Generates an Augmented Randomized Complete Block Design (ARCBD)

It randomly generates an augmented randomized complete block design across locations (ARCBD).

RCBD_augmented(
  lines = NULL,
  checks = NULL,
  b = NULL,
  l = 1,
  planter = "serpentine",
  plotNumber = 101,
  exptName = NULL,
  seed = NULL,
  locationNames = NULL,
  repsExpt = 1,
  data = NULL
)

Arguments

lines
Treatments, number of new lines for test.

checks
Number of checks per augmented block.

b
Number of augmented blocks.

l
Number of locations. By default l = 1.

planter
Option for serpentine or cartesian arrangement. By default planter = 'serpentine'.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

exptName
(optional) Name of experiment.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

locationNames
(optional) Name for each location.

repsExpt
(optional) Number of reps of experiment. By default repsExpt = 1.

data
(optional) Data frame with the labels of treatments.

Value

A list with information on the design parameters.

ARCBD layout for the first location.

Plot number layout for the first location.

Data frame with the ARCBD field book.

References

Federer, W. T. (1955). Experimental Design. Theory and Application. New York, USA. The Macmillan Company.

Examples

#Example 1: Generates an ARCBD with 6 blocks, 3 checks for each, and 50 treatments in two locations.
ARCBD1 <- RCBD_augmented(lines = 50, checks = 3, b = 6, l = 2, 
                         planter = "cartesian", 
                         plotNumber = c(1,1001),
                         seed = 23, 
                         locationNames = c("FARGO", "MINOT"))
ARCBD1$infoDesign

#> $Blocks
#> [1] 6
#> 
#> $plotsPerBlock
#> [1] 9 9 9 9 9 5
#> 
#> $Checks
#>   ENTRY    NAME
#> 1     1 Check 1
#> 2     2 Check 2
#> 3     3 Check 3
#> 
#> $repsExpt
#> [1] 1
#> 
#> $numberLocations
#> [1] 2
#> 
#> $Fillers
#> [1] 4
#> 
#> $seed
#> [1] 23
#> 

ARCBD1$layoutRandom

#>      Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12
#> Row6   15    3   21   26   36   38    1    2    0     0     0     0
#> Row5    2   48   46   11   42   32    3   24   37    20     1    31
#> Row4   33    2   47    7   19    4   12    1   23     3    22    29
#> Row3    3    6   52   27   44   45    2   49   35     1    14    51
#> Row2    2   50   13   16    1   25   41   34    9    43    39     3
#> Row1   28    8   18   53    3   10   40    2   30     5     1    17

ARCBD1$exptNames

#>      V1    V2    V3    V4    V5    V6    V7    V8    V9   V10   V11   V12
#> 1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 2 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 3 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 4 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 5 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 6 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1

ARCBD1$plotNumber

#>      V1 V2 V3 V4 V5 V6 V7 V8 V9 V10 V11 V12
#> [1,] 61 62 63 64 65 66 67 68  0   0   0   0
#> [2,] 49 50 51 52 53 54 55 56 57  58  59  60
#> [3,] 37 38 39 40 41 42 43 44 45  46  47  48
#> [4,] 25 26 27 28 29 30 31 32 33  34  35  36
#> [5,] 13 14 15 16 17 18 19 20 21  22  23  24
#> [6,]  1  2  3  4  5  6  7  8  9  10  11  12

head(ARCBD1$fieldbook, 12)

#>    ID  EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS BLOCK ENTRY TREATMENT
#> 1   1 Expt1    FARGO 2021    1   1      1      0     1    28    gen 28
#> 2   2 Expt1    FARGO 2021    2   1      2      0     1     8     gen 8
#> 3   3 Expt1    FARGO 2021    3   1      3      0     1    18    gen 18
#> 4   4 Expt1    FARGO 2021    4   1      4      0     1    53    gen 53
#> 5   5 Expt1    FARGO 2021    5   1      5      1     1     3   Check 3
#> 6   6 Expt1    FARGO 2021    6   1      6      0     1    10    gen 10
#> 7   7 Expt1    FARGO 2021    7   1      7      0     1    40    gen 40
#> 8   8 Expt1    FARGO 2021    8   1      8      1     1     2   Check 2
#> 9   9 Expt1    FARGO 2021    9   1      9      0     1    30    gen 30
#> 10 10 Expt1    FARGO 2021   10   1     10      0     1     5     gen 5
#> 11 11 Expt1    FARGO 2021   11   1     11      1     1     1   Check 1
#> 12 12 Expt1    FARGO 2021   12   1     12      0     1    17    gen 17

                   
#Example 2: Generates an ARCBD with 17 blocks, 4 checks for each, and 350 treatments in 3 locations.
# In this case, we show how to use the option data.
checks <- 4;
list_checks <- paste("CH", 1:checks, sep = "")
treatments <- paste("G", 5:354, sep = "")
treatment_list <- data.frame(list(ENTRY = 1:354, NAME = c(list_checks, treatments)))
head(treatment_list, 12)

#>    ENTRY NAME
#> 1      1  CH1
#> 2      2  CH2
#> 3      3  CH3
#> 4      4  CH4
#> 5      5   G5
#> 6      6   G6
#> 7      7   G7
#> 8      8   G8
#> 9      9   G9
#> 10    10  G10
#> 11    11  G11
#> 12    12  G12

ARCBD2 <- RCBD_augmented(lines = 350, checks = 4, b = 17, l = 3, 
                         planter = "serpentine", 
                         plotNumber = c(101,1001,2001), 
                         seed = 24, 
                         locationNames = LETTERS[1:3],
                         data = treatment_list)
ARCBD2$infoDesign

#> $Blocks
#> [1] 17
#> 
#> $plotsPerBlock
#>  [1] 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21 21 14
#> 
#> $Checks
#>   ENTRY NAME
#> 1     1  CH1
#> 2     2  CH2
#> 3     3  CH3
#> 4     4  CH4
#> 
#> $repsExpt
#> [1] 1
#> 
#> $numberLocations
#> [1] 3
#> 
#> $Fillers
#> [1] 7
#> 
#> $seed
#> [1] 24
#> 

ARCBD2$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row17  225    2  176  331    3  126  198  170  187     4   259   257   122
#> Row16  222  340  316    2  237   65  281  291   16    17   314    12   235
#> Row15  348  227   96    3  323   10  272  338  173   191     1    88   104
#> Row14   15  226    4  214  100  194  121  135  107    71   333   109   309
#> Row13    1   72   11  159   99    3  148   28  119   167    47   117    68
#> Row12   79  288   64  234  180  300  157  168    4   212    34     2     1
#> Row11  130   90    4  123  344    2  162  210    9   297   256   184   352
#> Row10  271   83  354  289  301    4    3  160    2   320   166    19   353
#> Row9   298    2    8  287   70  204  264    3  224   295    48   326   328
#> Row8   185   80    4  269  115  258  268  120   29    81   209     3   290
#> Row7   108  318  325  337  229   37  329  350  147   128    24     2    50
#> Row6   339  250   35  230  125  110   52   85    4   255    98    42     1
#> Row5     4    7   45  223   40   62   82  158  136     1   322     5   228
#> Row4    36  103  260  197    2    3   44  267  217     1    25    49   189
#> Row3     1   60    3  277  134  307  232  266  150   293   252    59   313
#> Row2   317    4  292   75  153  334   51  270   84   165   274     3    41
#> Row1   303  315   31  106  254  127  205    1    2   282   105   324    27
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row17     1    66   207   238   208     0     0     0     0     0     0     0
#> Row16     3    22    76   188   101    77   111   341     1   219   169     4
#> Row15    78   181   203     2   175     4   327   118   154   244   193   177
#> Row14   221   321   163   186   174     1    18     3   183     2   351    33
#> Row13    73    57    23   253   113   190   132   279     2     4   201    13
#> Row12    87   349   262   145    56   129    91   202   220   278   131     3
#> Row11   196   218   182   138    39   330   112   299     3   302   345     1
#> Row10   195   233     1    20   161   144   143   206    94   304   141     6
#> Row9    137   347   215   116   296   139    86   151     4   346     1   172
#> Row8    308     2   149   294     1    89   312   342    26   249   336   286
#> Row7    263    54     4     3   306    74    43    30     1   231    55   305
#> Row6    332   216    14     2    61   236   265    69   245     3   251   248
#> Row5    142    46   140   211   164   283   199    53     2     3   310    63
#> Row4    335   319   284   246   155   242   200     4    92    97   152   275
#> Row3     93     4   239   178   311   261    38   213   114     2    67    95
#> Row2    280   102     1   343   276   171   156     2   285   133   146   240
#> Row1    192    58     4     3    32   247   179   243    21   241   273   124

ARCBD2$exptNames

#>       V1    V2    V3    V4    V5    V6    V7    V8    V9   V10   V11   V12
#> 1  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 2  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 3  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 4  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 5  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 6  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 7  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 8  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 9  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 10 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 11 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 12 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 13 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 14 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 15 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 16 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 17 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#>      V13   V14   V15   V16   V17   V18   V19   V20   V21   V22   V23   V24
#> 1  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 2  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 3  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 4  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 5  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 6  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 7  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 8  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 9  Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 10 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 11 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 12 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 13 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 14 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 15 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 16 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#> 17 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1 Expt1
#>      V25
#> 1  Expt1
#> 2  Expt1
#> 3  Expt1
#> 4  Expt1
#> 5  Expt1
#> 6  Expt1
#> 7  Expt1
#> 8  Expt1
#> 9  Expt1
#> 10 Expt1
#> 11 Expt1
#> 12 Expt1
#> 13 Expt1
#> 14 Expt1
#> 15 Expt1
#> 16 Expt1
#> 17 Expt1

ARCBD2$plotNumber

#>        V1  V2  V3  V4  V5  V6  V7  V8  V9 V10 V11 V12 V13 V14 V15 V16 V17 V18
#>  [1,] 501 502 503 504 505 506 507 508 509 510 511 512 513 514 515 516 517 518
#>  [2,] 500 499 498 497 496 495 494 493 492 491 490 489 488 487 486 485 484 483
#>  [3,] 451 452 453 454 455 456 457 458 459 460 461 462 463 464 465 466 467 468
#>  [4,] 450 449 448 447 446 445 444 443 442 441 440 439 438 437 436 435 434 433
#>  [5,] 401 402 403 404 405 406 407 408 409 410 411 412 413 414 415 416 417 418
#>  [6,] 400 399 398 397 396 395 394 393 392 391 390 389 388 387 386 385 384 383
#>  [7,] 351 352 353 354 355 356 357 358 359 360 361 362 363 364 365 366 367 368
#>  [8,] 350 349 348 347 346 345 344 343 342 341 340 339 338 337 336 335 334 333
#>  [9,] 301 302 303 304 305 306 307 308 309 310 311 312 313 314 315 316 317 318
#> [10,] 300 299 298 297 296 295 294 293 292 291 290 289 288 287 286 285 284 283
#> [11,] 251 252 253 254 255 256 257 258 259 260 261 262 263 264 265 266 267 268
#> [12,] 250 249 248 247 246 245 244 243 242 241 240 239 238 237 236 235 234 233
#> [13,] 201 202 203 204 205 206 207 208 209 210 211 212 213 214 215 216 217 218
#> [14,] 200 199 198 197 196 195 194 193 192 191 190 189 188 187 186 185 184 183
#> [15,] 151 152 153 154 155 156 157 158 159 160 161 162 163 164 165 166 167 168
#> [16,] 150 149 148 147 146 145 144 143 142 141 140 139 138 137 136 135 134 133
#> [17,] 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117 118
#>       V19 V20 V21 V22 V23 V24 V25
#>  [1,]   0   0   0   0   0   0   0
#>  [2,] 482 481 480 479 478 477 476
#>  [3,] 469 470 471 472 473 474 475
#>  [4,] 432 431 430 429 428 427 426
#>  [5,] 419 420 421 422 423 424 425
#>  [6,] 382 381 380 379 378 377 376
#>  [7,] 369 370 371 372 373 374 375
#>  [8,] 332 331 330 329 328 327 326
#>  [9,] 319 320 321 322 323 324 325
#> [10,] 282 281 280 279 278 277 276
#> [11,] 269 270 271 272 273 274 275
#> [12,] 232 231 230 229 228 227 226
#> [13,] 219 220 221 222 223 224 225
#> [14,] 182 181 180 179 178 177 176
#> [15,] 169 170 171 172 173 174 175
#> [16,] 132 131 130 129 128 127 126
#> [17,] 119 120 121 122 123 124 125

head(ARCBD2$fieldbook, 12)

#>    ID  EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS BLOCK ENTRY TREATMENT
#> 1   1 Expt1        A 2021  101   1      1      0     1   303      G303
#> 2   2 Expt1        A 2021  102   1      2      0     1   315      G315
#> 3   3 Expt1        A 2021  103   1      3      0     1    31       G31
#> 4   4 Expt1        A 2021  104   1      4      0     1   106      G106
#> 5   5 Expt1        A 2021  105   1      5      0     1   254      G254
#> 6   6 Expt1        A 2021  106   1      6      0     1   127      G127
#> 7   7 Expt1        A 2021  107   1      7      0     1   205      G205
#> 8   8 Expt1        A 2021  108   1      8      1     1     1       CH1
#> 9   9 Expt1        A 2021  109   1      9      1     1     2       CH2
#> 10 10 Expt1        A 2021  110   1     10      0     1   282      G282
#> 11 11 Expt1        A 2021  111   1     11      0     1   105      G105
#> 12 12 Expt1        A 2021  112   1     12      0     1   324      G324