Spatial Un-replicated Diagonal Arrangement Design

Randomly generates an spatial un-replicated diagonal arrangement design.

diagonal_arrangement(
  nrows = NULL,
  ncols = NULL,
  lines = NULL,
  checks = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  kindExpt = "SUDC",
  splitBy = "row",
  seed = NULL,
  blocks = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows
Number of rows in the field.

ncols
Number of columns in the field.

lines
Number of genotypes, experimental lines or treatments.

checks
Integer number of genotypes checks or a numeric vector.

planter
Option for serpentine or cartesian plot arrangement. By default planter = 'serpentine'.

l
Number of locations or sites. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

kindExpt
Type of diagonal design, with options 'SUDC' and 'DBUDC'. By default kindExpt = 'SUDC'.

splitBy
Option to split the field when kindExpt = 'DBUDC' is selected. By default splitBy = 'row'.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

blocks
Number of experiments or blocks to generate an DBUDC design. If kindExpt = 'DBUDC' and data is null, blocks are mandatory.

exptName
(optional) Name of the experiment.

locationNames
(optional) Names each location.

data
(optional) Data frame with 3 columns: ENTRY | NAME | BLOCK or only 2 columns ENTRY | NAME if kindExpt = 'SUDC'.

Value

A list with information on the design parameters.

A matrix with the randomization layout.

A matrix with the layout plot number.

A data frame with fieldBook design. This includes the index (Row, Column).

Details

Constructs for a field of an arbitrary number of rows and columns randomized single or multi-level experiments for an arbitrary number of treatments lines and an arbitrary feasible number of checks. An arbitrary number of checks are allocated in a systematic diagonal movement.n un-replicated diagonal arrangement design where the checks be allocated in a systematic diagonal movement. In most of the cases is guarantee that each row and column has checks. Note that design generation needs the dimension of the field (number of rows and columns).

References

Hinkelmann, Klaus, Kempthorne, Oscar. Design and Analysis of Experiments, Volume 3: Special Designs and Applications. Hoboken, N.J.: Wiley-Interscience.

Examples

# Example 1: Generates a spatial single diagonal arrangement design in one location
# with 270 treatments and 30 check plots for a field with dimensions 15 rows x 20 cols
# in a serpentine arrangement.
spatd <- diagonal_arrangement(nrows = 15, ncols = 20, lines = 270, 
                              checks = 4, 
                              plotNumber = 101, 
                              kindExpt = "SUDC", 
                              planter = "serpentine", 
                              seed = 1987,
                              exptName = "20WRY1", 
                              locationNames = "MINOT")
spatd$infoDesign

#> $Lines
#> [1] 270
#> 
#> $checks
#> [1] 4
#> 
#> $RepChecks
#> 1 2 3 4 
#> 8 8 7 7 
#> 
#> $percentChecks
#> [1] "10%"
#> 
#> $Fillers
#> [1] 0
#> 
#> $seed
#> [1] 1987
#> 

spatd$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row15   59    4    9  271  139   92   87   37  136    74   113     2   270
#> Row14  146  267  248   12    3   88   82   47  210    66   153   122   211
#> Row13   42   75  173  178  253  107  181    4  150   131   105   124   249
#> Row12    2   19   96  201  117   29  121  215  242    34     1   265    24
#> Row11   35   91  164    2  126   30  161  224   11   255   194   246    17
#> Row10  205   73  158   84  127  244    3   62  137   157   106   195    80
#> Row9   167  133  251  145  168  204  116  119  247     4    39   123   256
#> Row8   238   57    1  162  193  190  142  101   26   239   222   227     1
#> Row7   100    8  261  208  172    2  202  184  120   219   218   196   177
#> Row6   129  134  138  245  221   79  191   46    2    72   169   152    61
#> Row5    76    3  216  258  180   67   55  250   18   236    52     4   140
#> Row4    50  132  254  272    3  109   68   53   70   264   103   212    28
#> Row3   187  148  135   43   41   49   21    2  217   104   213   237   186
#> Row2     1   36  154   38  234  233  160  252  240    98     4   108   197
#> Row1   185   31   94    1  274  200  163  207  232   257   203   118    25
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row15   189   183    33    78    54    20   214
#> Row14    99     2    85   102    71   175     5
#> Row13    83   260   130    44     4    32    65
#> Row12    48   206   259   235    90   263     6
#> Row11     3    27   165   174    81   223    69
#> Row10   166   182   110     1   111   155    77
#> Row9     14   229   188   230   241    45     4
#> Row8    170   225   268    23   209   143   144
#> Row7    171    51     2   141   112   243   192
#> Row6     95    10    60   151   220     3   262
#> Row5     89    93   156   179   115   176   199
#> Row4    266     1   149    63   128   226    16
#> Row3     13    15   125   228     1   198   114
#> Row2     64    58    40     7    22    97   147
#> Row1      3   159   273   269    56    86   231

spatd$plotsNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row15  381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row14  380  379  378  377  376  375  374  373  372   371   370   369   368
#> Row13  341  342  343  344  345  346  347  348  349   350   351   352   353
#> Row12  340  339  338  337  336  335  334  333  332   331   330   329   328
#> Row11  301  302  303  304  305  306  307  308  309   310   311   312   313
#> Row10  300  299  298  297  296  295  294  293  292   291   290   289   288
#> Row9   261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row8   260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row7   221  222  223  224  225  226  227  228  229   230   231   232   233
#> Row6   220  219  218  217  216  215  214  213  212   211   210   209   208
#> Row5   181  182  183  184  185  186  187  188  189   190   191   192   193
#> Row4   180  179  178  177  176  175  174  173  172   171   170   169   168
#> Row3   141  142  143  144  145  146  147  148  149   150   151   152   153
#> Row2   140  139  138  137  136  135  134  133  132   131   130   129   128
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row15   394   395   396   397   398   399   400
#> Row14   367   366   365   364   363   362   361
#> Row13   354   355   356   357   358   359   360
#> Row12   327   326   325   324   323   322   321
#> Row11   314   315   316   317   318   319   320
#> Row10   287   286   285   284   283   282   281
#> Row9    274   275   276   277   278   279   280
#> Row8    247   246   245   244   243   242   241
#> Row7    234   235   236   237   238   239   240
#> Row6    207   206   205   204   203   202   201
#> Row5    194   195   196   197   198   199   200
#> Row4    167   166   165   164   163   162   161
#> Row3    154   155   156   157   158   159   160
#> Row2    127   126   125   124   123   122   121
#> Row1    114   115   116   117   118   119   120

head(spatd$fieldBook, 12)

#>    ROW COLUMN ENTRY PLOT CHECKS   EXPT LOCATION LOC YEAR    NAME
#> 1    1      1   185  101      0 20WRY1    MINOT MNT 2021 gen 185
#> 2    1      2    31  102      0 20WRY1    MINOT MNT 2021  gen 31
#> 3    1      3    94  103      0 20WRY1    MINOT MNT 2021  gen 94
#> 4    1      4     1  104      1 20WRY1    MINOT MNT 2021 Check 1
#> 5    1      5   274  105      0 20WRY1    MINOT MNT 2021 gen 274
#> 6    1      6   200  106      0 20WRY1    MINOT MNT 2021 gen 200
#> 7    1      7   163  107      0 20WRY1    MINOT MNT 2021 gen 163
#> 8    1      8   207  108      0 20WRY1    MINOT MNT 2021 gen 207
#> 9    1      9   232  109      0 20WRY1    MINOT MNT 2021 gen 232
#> 10   1     10   257  110      0 20WRY1    MINOT MNT 2021 gen 257
#> 11   1     11   203  111      0 20WRY1    MINOT MNT 2021 gen 203
#> 12   1     12   118  112      0 20WRY1    MINOT MNT 2021 gen 118

# Example 2: Generates a spatial decision block diagonal arrangement design in one location
# with 720 treatments allocated in 5 experiments or blocks for a field with dimensions
# 30 rows x 26 cols in a serpentine arrangement.
spatDB <- diagonal_arrangement(nrows = 30, ncols = 26, lines = 720, 
                               checks = 5, 
                               plotNumber = 1, 
                               kindExpt = "DBUDC", 
                               planter = "serpentine", 
                               splitBy = "row", 
                               blocks = c(150,155,95,200,120))
spatDB$infoDesign

#> $Lines
#> [1] 150 155  95 200 120
#> 
#> $checks
#> [1] 5
#> 
#> $RepChecks
#>  1  2  3  4  5 
#> 14 13 12 11 10 
#> 
#> $percentChecks
#> [1] "7.7%"
#> 
#> $Fillers
#> [1] 0
#> 
#> $seed
#> [1] 3859.465
#> 

spatDB$layoutRandom

#>       Col1  Col2  Col3  Col4  Col5  Col6  Col7  Col8  Col9  Col10 Col11 Col12
#> Row30 "708" "5"   "647" "622" "625" "725" "628" "673" "704" "616" "631" "626"
#> Row29 "632" "683" "611" "661" "689" "3"   "646" "686" "614" "711" "663" "720"
#> Row28 "607" "680" "641" "615" "693" "697" "624" "717" "667" "1"   "656" "670"
#> Row27 "4"   "675" "716" "639" "635" "724" "665" "650" "652" "619" "715" "662"
#> Row26 "658" "610" "638" "718" "3"   "664" "682" "713" "703" "692" "627" "709"
#> Row25 "561" "580" "420" "508" "550" "565" "601" "411" "1"   "554" "425" "578"
#> Row24 "491" "476" "586" "559" "416" "581" "590" "469" "515" "547" "415" "460"
#> Row23 "557" "461" "526" "1"   "475" "465" "510" "436" "440" "585" "579" "468"
#> Row22 "595" "427" "457" "530" "501" "540" "495" "4"   "451" "442" "558" "592"
#> Row21 "542" "529" "569" "444" "478" "549" "484" "494" "413" "470" "588" "4"  
#> Row20 "505" "453" "2"   "454" "446" "479" "512" "502" "426" "500" "474" "604"
#> Row19 "524" "517" "528" "513" "509" "538" "2"   "488" "543" "576" "563" "439"
#> Row18 "498" "574" "560" "503" "466" "482" "449" "452" "527" "599" "3"   "489"
#> Row17 "390" "5"   "365" "337" "363" "349" "328" "372" "335" "317" "333" "326"
#> Row16 "371" "360" "391" "379" "366" "1"   "361" "358" "312" "324" "362" "343"
#> Row15 "401" "359" "393" "375" "330" "338" "347" "374" "311" "1"   "323" "385"
#> Row14 "2"   "380" "342" "394" "352" "402" "384" "377" "355" "405" "357" "369"
#> Row13 "230" "169" "157" "259" "4"   "222" "161" "298" "228" "300" "164" "307"
#> Row12 "288" "199" "223" "202" "236" "165" "180" "278" "5"   "247" "286" "254"
#> Row11 "296" "239" "291" "158" "283" "215" "237" "241" "177" "233" "308" "183"
#> Row10 "238" "255" "188" "3"   "262" "243" "275" "203" "193" "195" "174" "292"
#> Row9  "159" "196" "187" "261" "269" "175" "208" "5"   "162" "178" "172" "210"
#> Row8  "166" "251" "171" "299" "252" "273" "294" "181" "200" "198" "249" "2"  
#> Row7  "82"  "117" "5"   "90"  "18"  "125" "34"  "274" "289" "310" "189" "264"
#> Row6  "109" "65"  "132" "53"  "44"  "22"  "3"   "149" "17"  "12"  "52"  "120"
#> Row5  "6"   "40"  "41"  "61"  "139" "23"  "80"  "123" "15"  "63"  "4"   "140"
#> Row4  "60"  "3"   "129" "102" "33"  "134" "127" "130" "81"  "86"  "73"  "29" 
#> Row3  "93"  "150" "62"  "107" "99"  "5"   "48"  "75"  "114" "104" "143" "116"
#> Row2  "98"  "56"  "10"  "54"  "51"  "91"  "74"  "94"  "136" "1"   "13"  "100"
#> Row1  "1"   "87"  "95"  "72"  "142" "57"  "105" "83"  "25"  "113" "148" "66" 
#>       Col13 Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24
#> Row30 "637" "721" "4"   "642" "634" "645" "702" "707" "694" "685" "612" "696"
#> Row29 "655" "621" "681" "684" "629" "660" "2"   "695" "705" "653" "636" "674"
#> Row28 "633" "677" "668" "617" "657" "691" "654" "722" "666" "643" "4"   "712"
#> Row27 "723" "2"   "719" "608" "613" "714" "640" "649" "690" "606" "687" "698"
#> Row26 "678" "706" "710" "699" "676" "1"   "672" "651" "630" "669" "688" "618"
#> Row25 "481" "518" "521" "477" "514" "462" "414" "428" "572" "5"   "438" "507"
#> Row24 "3"   "473" "487" "421" "562" "568" "533" "570" "520" "583" "485" "418"
#> Row23 "496" "437" "430" "447" "1"   "600" "573" "408" "464" "511" "582" "486"
#> Row22 "412" "493" "463" "575" "435" "472" "422" "589" "3"   "594" "566" "406"
#> Row21 "584" "539" "458" "459" "564" "591" "535" "598" "504" "541" "409" "450"
#> Row20 "441" "536" "577" "5"   "434" "429" "497" "492" "432" "480" "555" "567"
#> Row19 "410" "417" "593" "424" "537" "467" "523" "2"   "532" "525" "456" "522"
#> Row18 "431" "587" "602" "548" "471" "546" "443" "556" "423" "448" "407" "1"  
#> Row17 "322" "400" "4"   "364" "388" "334" "597" "552" "544" "499" "433" "490"
#> Row16 "329" "354" "383" "351" "344" "368" "3"   "398" "373" "340" "387" "392"
#> Row15 "321" "367" "353" "316" "332" "399" "381" "313" "396" "376" "2"   "395"
#> Row14 "331" "3"   "404" "382" "318" "403" "325" "386" "356" "345" "336" "339"
#> Row13 "185" "256" "176" "156" "197" "1"   "235" "350" "378" "341" "346" "370"
#> Row12 "219" "224" "258" "232" "220" "272" "302" "184" "282" "2"   "213" "250"
#> Row11 "2"   "306" "207" "218" "257" "170" "303" "266" "227" "168" "194" "277"
#> Row10 "276" "271" "285" "190" "4"   "309" "173" "206" "211" "191" "280" "242"
#> Row9  "244" "270" "160" "284" "179" "268" "216" "229" "1"   "304" "163" "263"
#> Row8  "201" "212" "240" "265" "221" "225" "305" "226" "248" "267" "246" "295"
#> Row7  "301" "214" "293" "1"   "205" "182" "245" "279" "231" "186" "287" "290"
#> Row6  "59"  "77"  "88"  "154" "151" "50"  "67"  "4"   "138" "68"  "147" "37" 
#> Row5  "38"  "85"  "32"  "31"  "45"  "135" "146" "124" "39"  "9"   "43"  "4"  
#> Row4  "24"  "119" "2"   "28"  "16"  "144" "152" "11"  "145" "8"   "58"  "14" 
#> Row3  "111" "128" "112" "49"  "126" "19"  "5"   "115" "21"  "64"  "153" "92" 
#> Row2  "47"  "78"  "106" "71"  "20"  "122" "118" "42"  "131" "84"  "3"   "108"
#> Row1  "96"  "2"   "36"  "89"  "103" "30"  "27"  "133" "97"  "101" "69"  "110"
#>       Col25 Col26
#> Row30 "700" "679"
#> Row29 "644" "609"
#> Row28 "659" "648"
#> Row27 "701" "620"
#> Row26 "671" "623"
#> Row25 "545" "506"
#> Row24 "483" "2"  
#> Row23 "553" "551"
#> Row22 "534" "531"
#> Row21 "5"   "603"
#> Row20 "419" "571"
#> Row19 "519" "596"
#> Row18 "455" "516"
#> Row17 "605" "445"
#> Row16 "319" "320"
#> Row15 "315" "389"
#> Row14 "348" "327"
#> Row13 "314" "397"
#> Row12 "204" "167"
#> Row11 "297" "3"  
#> Row10 "234" "209"
#> Row9  "217" "192"
#> Row8  "1"   "253"
#> Row7  "281" "260"
#> Row6  "46"  "35" 
#> Row5  "121" "7"  
#> Row4  "79"  "55" 
#> Row3  "76"  "155"
#> Row2  "26"  "70" 
#> Row1  "141" "137"

spatDB$plotsNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  780  779  778  777  776  775  774  773  772   771   770   769   768
#> Row29  729  730  731  732  733  734  735  736  737   738   739   740   741
#> Row28  728  727  726  725  724  723  722  721  720   719   718   717   716
#> Row27  677  678  679  680  681  682  683  684  685   686   687   688   689
#> Row26  676  675  674  673  672  671  670  669  668   667   666   665   664
#> Row25  625  626  627  628  629  630  631  632  633   634   635   636   637
#> Row24  624  623  622  621  620  619  618  617  616   615   614   613   612
#> Row23  573  574  575  576  577  578  579  580  581   582   583   584   585
#> Row22  572  571  570  569  568  567  566  565  564   563   562   561   560
#> Row21  521  522  523  524  525  526  527  528  529   530   531   532   533
#> Row20  520  519  518  517  516  515  514  513  512   511   510   509   508
#> Row19  469  470  471  472  473  474  475  476  477   478   479   480   481
#> Row18  468  467  466  465  464  463  462  461  460   459   458   457   456
#> Row17  417  418  419  420  421  422  423  424  425   426   427   428   429
#> Row16  416  415  414  413  412  411  410  409  408   407   406   405   404
#> Row15  365  366  367  368  369  370  371  372  373   374   375   376   377
#> Row14  364  363  362  361  360  359  358  357  356   355   354   353   352
#> Row13  313  314  315  316  317  318  319  320  321   322   323   324   325
#> Row12  312  311  310  309  308  307  306  305  304   303   302   301   300
#> Row11  261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row10  260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row9   209  210  211  212  213  214  215  216  217   218   219   220   221
#> Row8   208  207  206  205  204  203  202  201  200   199   198   197   196
#> Row7   157  158  159  160  161  162  163  164  165   166   167   168   169
#> Row6   156  155  154  153  152  151  150  149  148   147   146   145   144
#> Row5   105  106  107  108  109  110  111  112  113   114   115   116   117
#> Row4   104  103  102  101  100   99   98   97   96    95    94    93    92
#> Row3    53   54   55   56   57   58   59   60   61    62    63    64    65
#> Row2    52   51   50   49   48   47   46   45   44    43    42    41    40
#> Row1     1    2    3    4    5    6    7    8    9    10    11    12    13
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row30   767   766   765   764   763   762   761   760   759   758   757   756
#> Row29   742   743   744   745   746   747   748   749   750   751   752   753
#> Row28   715   714   713   712   711   710   709   708   707   706   705   704
#> Row27   690   691   692   693   694   695   696   697   698   699   700   701
#> Row26   663   662   661   660   659   658   657   656   655   654   653   652
#> Row25   638   639   640   641   642   643   644   645   646   647   648   649
#> Row24   611   610   609   608   607   606   605   604   603   602   601   600
#> Row23   586   587   588   589   590   591   592   593   594   595   596   597
#> Row22   559   558   557   556   555   554   553   552   551   550   549   548
#> Row21   534   535   536   537   538   539   540   541   542   543   544   545
#> Row20   507   506   505   504   503   502   501   500   499   498   497   496
#> Row19   482   483   484   485   486   487   488   489   490   491   492   493
#> Row18   455   454   453   452   451   450   449   448   447   446   445   444
#> Row17   430   431   432   433   434   435   436   437   438   439   440   441
#> Row16   403   402   401   400   399   398   397   396   395   394   393   392
#> Row15   378   379   380   381   382   383   384   385   386   387   388   389
#> Row14   351   350   349   348   347   346   345   344   343   342   341   340
#> Row13   326   327   328   329   330   331   332   333   334   335   336   337
#> Row12   299   298   297   296   295   294   293   292   291   290   289   288
#> Row11   274   275   276   277   278   279   280   281   282   283   284   285
#> Row10   247   246   245   244   243   242   241   240   239   238   237   236
#> Row9    222   223   224   225   226   227   228   229   230   231   232   233
#> Row8    195   194   193   192   191   190   189   188   187   186   185   184
#> Row7    170   171   172   173   174   175   176   177   178   179   180   181
#> Row6    143   142   141   140   139   138   137   136   135   134   133   132
#> Row5    118   119   120   121   122   123   124   125   126   127   128   129
#> Row4     91    90    89    88    87    86    85    84    83    82    81    80
#> Row3     66    67    68    69    70    71    72    73    74    75    76    77
#> Row2     39    38    37    36    35    34    33    32    31    30    29    28
#> Row1     14    15    16    17    18    19    20    21    22    23    24    25
#>       Col26
#> Row30   755
#> Row29   754
#> Row28   703
#> Row27   702
#> Row26   651
#> Row25   650
#> Row24   599
#> Row23   598
#> Row22   547
#> Row21   546
#> Row20   495
#> Row19   494
#> Row18   443
#> Row17   442
#> Row16   391
#> Row15   390
#> Row14   339
#> Row13   338
#> Row12   287
#> Row11   286
#> Row10   235
#> Row9    234
#> Row8    183
#> Row7    182
#> Row6    131
#> Row5    130
#> Row4     79
#> Row3     78
#> Row2     27
#> Row1     26

head(spatDB$fieldBook,12)

#>    ROW COLUMN ENTRY PLOT CHECKS  EXPT LOCATION LOC YEAR    NAME
#> 1    1      1     1    1      1 Expt1        1   1 2021 Check 1
#> 2    1      2    87    2      0 Expt1        1   1 2021  gen 87
#> 3    1      3    95    3      0 Expt1        1   1 2021  gen 95
#> 4    1      4    72    4      0 Expt1        1   1 2021  gen 72
#> 5    1      5   142    5      0 Expt1        1   1 2021 gen 142
#> 6    1      6    57    6      0 Expt1        1   1 2021  gen 57
#> 7    1      7   105    7      0 Expt1        1   1 2021 gen 105
#> 8    1      8    83    8      0 Expt1        1   1 2021  gen 83
#> 9    1      9    25    9      0 Expt1        1   1 2021  gen 25
#> 10   1     10   113   10      0 Expt1        1   1 2021 gen 113
#> 11   1     11   148   11      0 Expt1        1   1 2021 gen 148
#> 12   1     12    66   12      0 Expt1        1   1 2021  gen 66

# Example 3: Generates a spatial decision block diagonal arrangement design in one location
# with 270 treatments allocated in 3 experiments or blocks for a field with dimensions
# 20 rows x 15 cols in a serpentine arrangement. Which in turn is an augmented block (3 blocks).
spatAB <- diagonal_arrangement(nrows = 20, ncols = 15, lines = 270, 
                               checks = 4, 
                               plotNumber = c(1,1001,2001), 
                               kindExpt = "DBUDC", 
                               planter = "serpentine",
                               exptName = c("20WRA", "20WRB", "20WRC"), 
                               blocks = c(90, 90, 90),
                               splitBy = "column")
spatAB$infoDesign

#> $Lines
#> [1] 90 90 90
#> 
#> $checks
#> [1] 4
#> 
#> $RepChecks
#> 1 2 3 4 
#> 9 9 6 6 
#> 
#> $percentChecks
#> [1] "10%"
#> 
#> $Fillers
#> [1] 0
#> 
#> $seed
#> [1] -6768.893
#> 

spatAB$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20   51    3   15   76   54  142  165   96  136   109   272     4   230
#> Row19   22   69   87   56    3  140  163  153  121   125   247   246   269
#> Row18   38   90   79    6    5  175  144    4  139   152   211   232   205
#> Row17    4   23   35   50   55  130  156  177  179   176     3   208   229
#> Row16   28   70   33    4   43  151  117  115  114    99   227   262   249
#> Row15   89   88   46   67   72  128    4  123  154   172   266   234   233
#> Row14    7   36   14   24   71  145  122  101  169     2   267   199   226
#> Row13   37   68    2   80   47  132  161   98  157   182   251   222     1
#> Row12   44   78   63   41   34    1  155  173  131   104   271   273   252
#> Row11   64   12   75   66   48  138  105  162    3   106   259   206   274
#> Row10   62    1   92   49   16   97  127  102  137   183   239     4   245
#> Row9    57   40   65   29    1  146   95  116  110   112   215   219   241
#> Row8    73   32    9   27   18  124  180    2  167   143   265   200   187
#> Row7     2   10   45   42   74  119  141  111  170   174     2   212   189
#> Row6    59   11   94    2   58  100  168  108  164   133   186   257   190
#> Row5    17   31   91   19   83  178    3  107  120   135   185   240   258
#> Row4    13    8   82   21   60  149  129  184  181     2   192   264   216
#> Row3    81   52    1   77   85  147  166  148  118   158   270   218     2
#> Row2    25   61   86   53   39    1  160  113  126   159   228   193   242
#> Row1    20   30   26   93   84  171  103  150    1   134   204   220   263
#>       Col14 Col15
#> Row20   195   235
#> Row19   213     3
#> Row18   268   221
#> Row17   194   237
#> Row16     1   238
#> Row15   261   225
#> Row14   197   210
#> Row13   196   231
#> Row12   217   244
#> Row11   202   254
#> Row10   214   191
#> Row9    248     2
#> Row8    253   243
#> Row7    201   260
#> Row6      1   223
#> Row5    256   255
#> Row4    224   203
#> Row3    207   236
#> Row2    250   198
#> Row1    209   188

spatAB$plotsNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  100   99   98   97   96 1100 1099 1098 1097  1096  2100  2099  2098
#> Row19   91   92   93   94   95 1091 1092 1093 1094  1095  2091  2092  2093
#> Row18   90   89   88   87   86 1090 1089 1088 1087  1086  2090  2089  2088
#> Row17   81   82   83   84   85 1081 1082 1083 1084  1085  2081  2082  2083
#> Row16   80   79   78   77   76 1080 1079 1078 1077  1076  2080  2079  2078
#> Row15   71   72   73   74   75 1071 1072 1073 1074  1075  2071  2072  2073
#> Row14   70   69   68   67   66 1070 1069 1068 1067  1066  2070  2069  2068
#> Row13   61   62   63   64   65 1061 1062 1063 1064  1065  2061  2062  2063
#> Row12   60   59   58   57   56 1060 1059 1058 1057  1056  2060  2059  2058
#> Row11   51   52   53   54   55 1051 1052 1053 1054  1055  2051  2052  2053
#> Row10   50   49   48   47   46 1050 1049 1048 1047  1046  2050  2049  2048
#> Row9    41   42   43   44   45 1041 1042 1043 1044  1045  2041  2042  2043
#> Row8    40   39   38   37   36 1040 1039 1038 1037  1036  2040  2039  2038
#> Row7    31   32   33   34   35 1031 1032 1033 1034  1035  2031  2032  2033
#> Row6    30   29   28   27   26 1030 1029 1028 1027  1026  2030  2029  2028
#> Row5    21   22   23   24   25 1021 1022 1023 1024  1025  2021  2022  2023
#> Row4    20   19   18   17   16 1020 1019 1018 1017  1016  2020  2019  2018
#> Row3    11   12   13   14   15 1011 1012 1013 1014  1015  2011  2012  2013
#> Row2    10    9    8    7    6 1010 1009 1008 1007  1006  2010  2009  2008
#> Row1     1    2    3    4    5 1001 1002 1003 1004  1005  2001  2002  2003
#>       Col14 Col15
#> Row20  2097  2096
#> Row19  2094  2095
#> Row18  2087  2086
#> Row17  2084  2085
#> Row16  2077  2076
#> Row15  2074  2075
#> Row14  2067  2066
#> Row13  2064  2065
#> Row12  2057  2056
#> Row11  2054  2055
#> Row10  2047  2046
#> Row9   2044  2045
#> Row8   2037  2036
#> Row7   2034  2035
#> Row6   2027  2026
#> Row5   2024  2025
#> Row4   2017  2016
#> Row3   2014  2015
#> Row2   2007  2006
#> Row1   2004  2005

head(spatAB$fieldBook,12)

#>    ROW COLUMN ENTRY PLOT CHECKS  EXPT LOCATION LOC YEAR    NAME
#> 1    1      1    20    1      0 20WRA        1   1 2021  gen 20
#> 2    1      2    30    2      0 20WRA        1   1 2021  gen 30
#> 3    1      3    26    3      0 20WRA        1   1 2021  gen 26
#> 4    1      4    93    4      0 20WRA        1   1 2021  gen 93
#> 5    1      5    84    5      0 20WRA        1   1 2021  gen 84
#> 6    1      6   171 1001      0 20WRB        1   1 2021 gen 171
#> 7    1      7   103 1002      0 20WRB        1   1 2021 gen 103
#> 8    1      8   150 1003      0 20WRB        1   1 2021 gen 150
#> 9    1      9     1 1004      1 20WRB        1   1 2021 Check 1
#> 10   1     10   134 1005      0 20WRB        1   1 2021 gen 134
#> 11   1     11   204 2001      0 20WRC        1   1 2021 gen 204
#> 12   1     12   220 2002      0 20WRC        1   1 2021 gen 220