Generates a Latin Square Design

Randomly generates a latin square design of up 10 treatments.

latin_square(
  t = NULL,
  reps = 1,
  plotNumber = 101,
  planter = "serpentine",
  seed = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

t
Number of treatments.

reps
Number of full resolvable squares. By default reps = 1.

plotNumber
Starting plot number. By default plotNumber = 101.

planter
Option for serpentine or cartesian arrangement. By default planter = 'serpentine'.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

locationNames
Name for each location.

data
optional) Data frame with label list of treatments.

Value

A list with information on the design parameters.

Data frame with the latin square field book.

References

Design and Analysis of Experiments, Volume 1, Introduction to Experimental Design. Second Edition. Klaus Hinkelmann & Oscar Kempthorne.John Wiley & Sons, Inc., Hoboken, New Jersey.

Examples

# Example 1: Generates a latin square design with 5 treatments and 3 reps.
latinSq1 <- latin_square(t = 5, reps = 3, plotNumber = 101, planter = "serpentine",
                         seed = 1981, locationNames = "FARGO") 
latinSq1$squares

#> [[1]]
#>       Column 1 Column 2 Column 3 Column 4 Column 5
#> Row 1 "T4"     "T3"     "T2"     "T1"     "T5"    
#> Row 2 "T1"     "T2"     "T4"     "T5"     "T3"    
#> Row 3 "T3"     "T5"     "T1"     "T2"     "T4"    
#> Row 4 "T2"     "T4"     "T5"     "T3"     "T1"    
#> Row 5 "T5"     "T1"     "T3"     "T4"     "T2"    
#> 
#> [[2]]
#>       Column 1 Column 2 Column 3 Column 4 Column 5
#> Row 1 "T2"     "T4"     "T5"     "T3"     "T1"    
#> Row 2 "T1"     "T2"     "T3"     "T4"     "T5"    
#> Row 3 "T3"     "T5"     "T1"     "T2"     "T4"    
#> Row 4 "T4"     "T1"     "T2"     "T5"     "T3"    
#> Row 5 "T5"     "T3"     "T4"     "T1"     "T2"    
#> 
#> [[3]]
#>       Column 1 Column 2 Column 3 Column 4 Column 5
#> Row 1 "T4"     "T2"     "T1"     "T3"     "T5"    
#> Row 2 "T1"     "T3"     "T2"     "T5"     "T4"    
#> Row 3 "T3"     "T5"     "T4"     "T2"     "T1"    
#> Row 4 "T5"     "T1"     "T3"     "T4"     "T2"    
#> Row 5 "T2"     "T4"     "T5"     "T1"     "T3"    
#> 

latinSq1$plotSquares

#> [[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  101  102  103  104  105
#> [2,]  110  109  108  107  106
#> [3,]  111  112  113  114  115
#> [4,]  120  119  118  117  116
#> [5,]  121  122  123  124  125
#> 
#> [[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  201  202  203  204  205
#> [2,]  210  209  208  207  206
#> [3,]  211  212  213  214  215
#> [4,]  220  219  218  217  216
#> [5,]  221  222  223  224  225
#> 
#> [[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  301  302  303  304  305
#> [2,]  310  309  308  307  306
#> [3,]  311  312  313  314  315
#> [4,]  320  319  318  317  316
#> [5,]  321  322  323  324  325
#> 

head(latinSq1$fieldBook)

#>   LOCATION PLOT SQUARE   ROW   COLUMN TREATMENT
#> 1    FARGO  101      1 Row 1 Column 1        T4
#> 2    FARGO  102      1 Row 1 Column 2        T3
#> 3    FARGO  103      1 Row 1 Column 3        T2
#> 4    FARGO  104      1 Row 1 Column 4        T1
#> 5    FARGO  105      1 Row 1 Column 5        T5
#> 6    FARGO  106      1 Row 2 Column 5        T3