Generates an Spatial Un-replicated Optimized Arrangement Design

Randomly generates a spatial un-replicated optimized arrangement design, where the distance between checks is maximized in such a way that each row and column have control plots. Note that design generation needs the dimension of the field (number of rows and columns).

optimized_arrangement(
  nrows = NULL,
  ncols = NULL,
  lines = NULL,
  amountChecks = NULL,
  checks = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  seed = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows
Number of rows in the field.

ncols
Number of columns in the field.

lines
Number of genotypes, experimental lines or treatments.

amountChecks
Integer with the amount total of checks or a numeric vector with the replicates of each check label.

checks
Number of genotypes as checks.

planter
Option for serpentine or cartesian arrangement. By default planter = 'serpentine'.

l
Number of locations. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

exptName
(optional) Name of the experiment.

locationNames
(optional) Name for each location.

data
(optional) Data frame with 3 columns: ENTRY | NAME | REPS.

Value

A list with five elements.

infoDesign is a list with information on the design parameters.

layoutRandom is a matrix with the randomization layout.

plotNumber is a matrix with the layout plot number.

data_entry is a data frame with the data input.

fieldBook is a data frame with field book design. This includes the index (Row, Column).

References

Clarke, G. P. Y., & Stefanova, K. T. (2011). Optimal design for early-generation plant breeding trials with unreplicated or partially replicated test lines. Australian & New Zealand Journal of Statistics, 53(4), 461–480.

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a spatial unreplicated optimized arrangement design in one location
# with 362 genotypes + 38 check plots (5 checks) for a field with dimension 20 rows x 20 cols.
OptimAd1 <- optimized_arrangement(nrows = 20, ncols = 20, lines = 362, 
                                  amountChecks = 38, 
                                  checks = 1:5,
                                  planter = "cartesian", 
                                  plotNumber = 101,
                                  seed = 14,
                                  exptName = "20RW1",
                                  locationNames = "CASSELTON")
OptimAd1$infoDesign

#> $Lines
#> [1] 362
#> 
#> $checks
#> [1] 1 2 3 4 5
#> 
#> $RepChecks
#> [1]  7  7 10  7  7
#> 
#> $seed
#> [1] 14
#> 
#> $idDesign
#> [1] 16
#> 

OptimAd1$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  296    4   16  123  276  190  251  117  229   166    62   160   165
#> Row19  131   97   82  307  344  293  218  258    2    48   239    66     2
#> Row18    5  230   69   71  164  286  114  248  142   280    88   156   223
#> Row17   30   68    4   23   22  254   44  314  243    17     3   225   309
#> Row16  299  167  285  125    1  270  301    1  234   320   326   274   332
#> Row15    1  304   18  198  232  181   43  358  246   173    56   306   233
#> Row14  237  195  138  150  357    2  183  140  308     5   215   271   353
#> Row13  241  202  279  336   74  328  126   39  106   158   345     3   323
#> Row12  316  222  103  264  249  120  275    2   40   289   236     1   220
#> Row11  199   94   45  322  265  231  331  343    3   163   101   252    76
#> Row10  119  294  216  335  356  364  155  319  342   146    42   327     5
#> Row9    32   12  272   98    4  192  291   46   83   350   261   200   346
#> Row8   205  196   33   59  300   31  211  311  201   351   212    21   365
#> Row7   305  180  367  313  177   10    4  310   11   136    14   159   161
#> Row6   100   50  139  151  208  116    5  226  110     1   113   197   361
#> Row5   149  105   79    5   91  339  227  214  262   141    81    93   134
#> Row4   360  179  315   51   20   36  312  157  278   337   213    89   187
#> Row3   266  281    4  366  325    2  269   37  273   189   244    96   341
#> Row2    87    7  352  124  185  107   13  122  221   253     4   175   219
#> Row1    24   72  349  137   41   65  128  191  298   130    86   147   256
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row20   260   170   282   359   121    34   324
#> Row19   186   283   206   250    99   210   203
#> Row18   267   127   303   104   109   235     2
#> Row17    38     6    90    19   174   295   217
#> Row16   268   207   297   292    75   362    58
#> Row15    25    60   111   363    28     5   321
#> Row14   115   288    84    53   168   108    29
#> Row13   176   135    67   112     5    35   259
#> Row12   148   247   188    15   102    92    52
#> Row11   354    64    95     3   263   245     8
#> Row10   145     2   287   242    70    57   129
#> Row9      3    26   152    80   154     9    49
#> Row8    240     2   333     2    27    73   330
#> Row7     63   318   209   355     3   255   133
#> Row6    182   144   329    47   224    77   290
#> Row5    172   118   347   143   302    85   228
#> Row4    338   284     1   348   162   194     3
#> Row3    169   317    78   340   153   171   238
#> Row2    277   178     1   193    54   334   132
#> Row1      2   204   184    61    55     4   257

OptimAd1$plotNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  481  482  483  484  485  486  487  488  489   490   491   492   493
#> Row19  461  462  463  464  465  466  467  468  469   470   471   472   473
#> Row18  441  442  443  444  445  446  447  448  449   450   451   452   453
#> Row17  421  422  423  424  425  426  427  428  429   430   431   432   433
#> Row16  401  402  403  404  405  406  407  408  409   410   411   412   413
#> Row15  381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row14  361  362  363  364  365  366  367  368  369   370   371   372   373
#> Row13  341  342  343  344  345  346  347  348  349   350   351   352   353
#> Row12  321  322  323  324  325  326  327  328  329   330   331   332   333
#> Row11  301  302  303  304  305  306  307  308  309   310   311   312   313
#> Row10  281  282  283  284  285  286  287  288  289   290   291   292   293
#> Row9   261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row8   241  242  243  244  245  246  247  248  249   250   251   252   253
#> Row7   221  222  223  224  225  226  227  228  229   230   231   232   233
#> Row6   201  202  203  204  205  206  207  208  209   210   211   212   213
#> Row5   181  182  183  184  185  186  187  188  189   190   191   192   193
#> Row4   161  162  163  164  165  166  167  168  169   170   171   172   173
#> Row3   141  142  143  144  145  146  147  148  149   150   151   152   153
#> Row2   121  122  123  124  125  126  127  128  129   130   131   132   133
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row20   494   495   496   497   498   499   500
#> Row19   474   475   476   477   478   479   480
#> Row18   454   455   456   457   458   459   460
#> Row17   434   435   436   437   438   439   440
#> Row16   414   415   416   417   418   419   420
#> Row15   394   395   396   397   398   399   400
#> Row14   374   375   376   377   378   379   380
#> Row13   354   355   356   357   358   359   360
#> Row12   334   335   336   337   338   339   340
#> Row11   314   315   316   317   318   319   320
#> Row10   294   295   296   297   298   299   300
#> Row9    274   275   276   277   278   279   280
#> Row8    254   255   256   257   258   259   260
#> Row7    234   235   236   237   238   239   240
#> Row6    214   215   216   217   218   219   220
#> Row5    194   195   196   197   198   199   200
#> Row4    174   175   176   177   178   179   180
#> Row3    154   155   156   157   158   159   160
#> Row2    134   135   136   137   138   139   140
#> Row1    114   115   116   117   118   119   120

head(OptimAd1$fieldBook,12)

#>    ID  EXPT  LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20RW1 CASSELTON 2021  101   1      1      0    24       G24
#> 2   2 20RW1 CASSELTON 2021  102   1      2      0    72       G72
#> 3   3 20RW1 CASSELTON 2021  103   1      3      0   349      G349
#> 4   4 20RW1 CASSELTON 2021  104   1      4      0   137      G137
#> 5   5 20RW1 CASSELTON 2021  105   1      5      0    41       G41
#> 6   6 20RW1 CASSELTON 2021  106   1      6      0    65       G65
#> 7   7 20RW1 CASSELTON 2021  107   1      7      0   128      G128
#> 8   8 20RW1 CASSELTON 2021  108   1      8      0   191      G191
#> 9   9 20RW1 CASSELTON 2021  109   1      9      0   298      G298
#> 10 10 20RW1 CASSELTON 2021  110   1     10      0   130      G130
#> 11 11 20RW1 CASSELTON 2021  111   1     11      0    86       G86
#> 12 12 20RW1 CASSELTON 2021  112   1     12      0   147      G147

                  
# Example 2: Generates a spatial unreplicated optimized arrangement design in one location
# with 635 genotypes + 65 check plots (4 checks) for a field with dimension 20 rows x 35 cols.
# As example, we set up the data option with the entries list.
checks <- 4
list_checks <- paste("CH", 1:checks, sep = "")
treatments <- paste("G", 5:639, sep = "")
REPS <- c(17, 16, 16, 16, rep(1, 635))
treatment_list <- data.frame(list(ENTRY = 1:639, NAME = c(list_checks, treatments), REPS = REPS))
head(treatment_list, 12) 

#>    ENTRY NAME REPS
#> 1      1  CH1   17
#> 2      2  CH2   16
#> 3      3  CH3   16
#> 4      4  CH4   16
#> 5      5   G5    1
#> 6      6   G6    1
#> 7      7   G7    1
#> 8      8   G8    1
#> 9      9   G9    1
#> 10    10  G10    1
#> 11    11  G11    1
#> 12    12  G12    1

tail(treatment_list, 12)

#>     ENTRY NAME REPS
#> 628   628 G628    1
#> 629   629 G629    1
#> 630   630 G630    1
#> 631   631 G631    1
#> 632   632 G632    1
#> 633   633 G633    1
#> 634   634 G634    1
#> 635   635 G635    1
#> 636   636 G636    1
#> 637   637 G637    1
#> 638   638 G638    1
#> 639   639 G639    1

OptimAd2 <- optimized_arrangement(nrows = 20, ncols = 35, 
                                  planter = "serpentine", 
                                  plotNumber = 101,
                                  seed = 12,
                                  exptName = "20YWA2",
                                  locationNames = "MINOT",
                                  data = treatment_list)
OptimAd2$infoDesign

#> $Lines
#> [1] 635
#> 
#> $checks
#> [1] 1 2 3 4
#> 
#> $RepChecks
#> [1] 17 16 16 16
#> 
#> $seed
#> [1] 12
#> 
#> $idDesign
#> [1] 16
#> 

OptimAd2$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  397  230  525  527    3  561  607    2  562   292    22   186   277
#> Row19  232   20    1  197  469  221   66  243  191    28     9   117   254
#> Row18  177   23  548  267  564  432    2  206  239    51   234   165     2
#> Row17   29    3   96  122  529  560   93  228    3   472   500   392    71
#> Row16  202  351  294   60   41  539  462  216  618   464   161   201   381
#> Row15  395  323  402  456  110  114  408  163    1   180   350   189   125
#> Row14  568  409  127  129    2  327  483  332  345   498   273   422   128
#> Row13  569   63  445  174  296  465  321  634  581    86   383   627   631
#> Row12  580  156  272  417  162  151  269  226  526   330   449   535   597
#> Row11    1   46  556  486  299   77  401   90  358   457    11   362     4
#> Row10  261  619  584  424  385  258  260  355  121   252   257   188   217
#> Row9   604  278  153  235  578    1  520  461  106   479   499    78   367
#> Row8   482  557  399  503   73  337  196    2   47   546     2   360   563
#> Row7   594  113   88   81  220  173  530  407  168   306   144     1   333
#> Row6   593  433  147  523  411  379    2  591  325     4   478    80   365
#> Row5     4  467  250    4   76    3  590  316   12   496     2   286    75
#> Row4   280   97  204  601  193  545  315  130  396   211   192    24   139
#> Row3   439  551  388  493  133  543  429  276  247   534   352     4     6
#> Row2    69  109  514    1  589   89  369   44  507     4   450   207   270
#> Row1   210    4   98  298  347  552  300  194  143   354   312    43   451
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row20   205   152   313   549    79   138   140   183    52   633   335   222
#> Row19    70   466     4   624     4   394   386   610     8   531   419   213
#> Row18   154   307   291   440   405   517   187   319   504   364   178   287
#> Row17   311   495   444   623   324     3   477    14   107   617     1    67
#> Row16     1   412   274    74   259   387    62   588   544   540   181   342
#> Row15   373   219   366   134   344   421   263   275    48     2   281   603
#> Row14   233   629   406     3   537    18   622     3   238   116   305   268
#> Row13   283   329    91   532   199     4   468   215   446   343   164   375
#> Row12   359    36     3   511   550   639   458   474   572     1     2   167
#> Row11   418     1   137   494   505   184   510    54   195   519   308    35
#> Row10   141   119    85    10   145   246     3   611    39   625   454     4
#> Row9    249   346   463   185   256    83   533    13     2   374   349   255
#> Row8    398    19   169   515   331   587    57    61   576   171   103   288
#> Row7    410     4   632    95   227   626   602   371   289   159    59     4
#> Row6    108   157   637   487   370   470   303     1   488   179   242   175
#> Row5    126   435   427   484   170   608   547   295   437   538   100   582
#> Row4      3   262   155     1   368   336    53   425   413   485   416   612
#> Row3     32   502   279   241   583   297     2   115   309   245   356   372
#> Row2    104   380   290   403   599    50    55   148   404   554   229   237
#> Row1    363   513   448    26    38   314   497   613     3   447   553   301
#>       Col26 Col27 Col28 Col29 Col30 Col31 Col32 Col33 Col34 Col35
#> Row20   190   573   225   524     2   150   231    25    45   248
#> Row19   251    68     5   326   431   160   570   453   101   455
#> Row18   480   443    16     1   266   512     1   253   320   284
#> Row17   528   198    33   620   491   541   120    82    17   596
#> Row16   536   378   415   565    21   223     7     2     4   341
#> Row15   285     1    42   208   400   521   630   317    27   393
#> Row14   264   318   310   304   328     3   567   158    30   436
#> Row13    40   598     4   595     1   501   203   452   124   200
#> Row12    31   123   265   111   430   571    87   339    58   146
#> Row11   389   338   149   522   614   566   509    49   600    65
#> Row10   628   459   518   322   616     2   224   575   609    84
#> Row9    423   218   244   438   414   112     1   574   212   118
#> Row8    176   293   142     3   340   442   476   555   132   353
#> Row7    236   102   441   348   434   166   172   420   508   635
#> Row6    105   473   605   579   471   585   492   182   302   475
#> Row5    214    15   384   426   506   376   615   382   460    34
#> Row4     99    72   209   558   586   490   559   489   361     3
#> Row3      3   136   621    94   577   606   636    92   592     3
#> Row2      2   481    56   516    64   357   377   271   240   428
#> Row1    391   542   638   282    37   135   131     4   334   390

OptimAd2$plotNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  800  799  798  797  796  795  794  793  792   791   790   789   788
#> Row19  731  732  733  734  735  736  737  738  739   740   741   742   743
#> Row18  730  729  728  727  726  725  724  723  722   721   720   719   718
#> Row17  661  662  663  664  665  666  667  668  669   670   671   672   673
#> Row16  660  659  658  657  656  655  654  653  652   651   650   649   648
#> Row15  591  592  593  594  595  596  597  598  599   600   601   602   603
#> Row14  590  589  588  587  586  585  584  583  582   581   580   579   578
#> Row13  521  522  523  524  525  526  527  528  529   530   531   532   533
#> Row12  520  519  518  517  516  515  514  513  512   511   510   509   508
#> Row11  451  452  453  454  455  456  457  458  459   460   461   462   463
#> Row10  450  449  448  447  446  445  444  443  442   441   440   439   438
#> Row9   381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row8   380  379  378  377  376  375  374  373  372   371   370   369   368
#> Row7   311  312  313  314  315  316  317  318  319   320   321   322   323
#> Row6   310  309  308  307  306  305  304  303  302   301   300   299   298
#> Row5   241  242  243  244  245  246  247  248  249   250   251   252   253
#> Row4   240  239  238  237  236  235  234  233  232   231   230   229   228
#> Row3   171  172  173  174  175  176  177  178  179   180   181   182   183
#> Row2   170  169  168  167  166  165  164  163  162   161   160   159   158
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row20   787   786   785   784   783   782   781   780   779   778   777   776
#> Row19   744   745   746   747   748   749   750   751   752   753   754   755
#> Row18   717   716   715   714   713   712   711   710   709   708   707   706
#> Row17   674   675   676   677   678   679   680   681   682   683   684   685
#> Row16   647   646   645   644   643   642   641   640   639   638   637   636
#> Row15   604   605   606   607   608   609   610   611   612   613   614   615
#> Row14   577   576   575   574   573   572   571   570   569   568   567   566
#> Row13   534   535   536   537   538   539   540   541   542   543   544   545
#> Row12   507   506   505   504   503   502   501   500   499   498   497   496
#> Row11   464   465   466   467   468   469   470   471   472   473   474   475
#> Row10   437   436   435   434   433   432   431   430   429   428   427   426
#> Row9    394   395   396   397   398   399   400   401   402   403   404   405
#> Row8    367   366   365   364   363   362   361   360   359   358   357   356
#> Row7    324   325   326   327   328   329   330   331   332   333   334   335
#> Row6    297   296   295   294   293   292   291   290   289   288   287   286
#> Row5    254   255   256   257   258   259   260   261   262   263   264   265
#> Row4    227   226   225   224   223   222   221   220   219   218   217   216
#> Row3    184   185   186   187   188   189   190   191   192   193   194   195
#> Row2    157   156   155   154   153   152   151   150   149   148   147   146
#> Row1    114   115   116   117   118   119   120   121   122   123   124   125
#>       Col26 Col27 Col28 Col29 Col30 Col31 Col32 Col33 Col34 Col35
#> Row20   775   774   773   772   771   770   769   768   767   766
#> Row19   756   757   758   759   760   761   762   763   764   765
#> Row18   705   704   703   702   701   700   699   698   697   696
#> Row17   686   687   688   689   690   691   692   693   694   695
#> Row16   635   634   633   632   631   630   629   628   627   626
#> Row15   616   617   618   619   620   621   622   623   624   625
#> Row14   565   564   563   562   561   560   559   558   557   556
#> Row13   546   547   548   549   550   551   552   553   554   555
#> Row12   495   494   493   492   491   490   489   488   487   486
#> Row11   476   477   478   479   480   481   482   483   484   485
#> Row10   425   424   423   422   421   420   419   418   417   416
#> Row9    406   407   408   409   410   411   412   413   414   415
#> Row8    355   354   353   352   351   350   349   348   347   346
#> Row7    336   337   338   339   340   341   342   343   344   345
#> Row6    285   284   283   282   281   280   279   278   277   276
#> Row5    266   267   268   269   270   271   272   273   274   275
#> Row4    215   214   213   212   211   210   209   208   207   206
#> Row3    196   197   198   199   200   201   202   203   204   205
#> Row2    145   144   143   142   141   140   139   138   137   136
#> Row1    126   127   128   129   130   131   132   133   134   135

head(OptimAd2$fieldBook,12)

#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20YWA2    MINOT 2021  101   1      1      0   210      G210
#> 2   2 20YWA2    MINOT 2021  102   1      2      1     4       CH4
#> 18  3 20YWA2    MINOT 2021  103   1      3      0    98       G98
#> 19  4 20YWA2    MINOT 2021  104   1      4      0   298      G298
#> 20  5 20YWA2    MINOT 2021  105   1      5      0   347      G347
#> 21  6 20YWA2    MINOT 2021  106   1      6      0   552      G552
#> 22  7 20YWA2    MINOT 2021  107   1      7      0   300      G300
#> 23  8 20YWA2    MINOT 2021  108   1      8      0   194      G194
#> 24  9 20YWA2    MINOT 2021  109   1      9      0   143      G143
#> 25 10 20YWA2    MINOT 2021  110   1     10      0   354      G354
#> 26 11 20YWA2    MINOT 2021  111   1     11      0   312      G312
#> 27 12 20YWA2    MINOT 2021  112   1     12      0    43       G43