Generates a Spatial Partially Replicated Arrangement Design

Randomly generates a spatial partially replicated design, where the distance between checks is maximized in such a way that each row and column have control plots. Note that design generation needs the dimension of the field (number of rows and columns).

partially_replicated(
  nrows = NULL,
  ncols = NULL,
  repGens = NULL,
  repUnits = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  seed = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows
Number of rows field.

ncols
Number of columns field.

repGens
Numeric vector with the amount genotypes to replicate.

repUnits
Numeric vector with the number of reps of each genotype.

planter
Option for serpentine or cartesian movement. By default planter = 'serpentine'.

l
Number of locations. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

exptName
(optional) Name of the experiment.

locationNames
(optional) Name for each location.

data
(optional) Dataframe with 3 columns: ENTRY | NAME | REPS.

Value

A list with information on the design parameters.

A matrix with the randomization layout.

A matrix with the layout plot number.

A data frame with fieldBook design. This includes the index (Row, Column).

References

Hinkelmann, Klaus, Kempthorne, Oscar. Design and Analysis of Experiments, Volume 3 : Special Designs and Applications. Hoboken, N.J.: Wiley-Interscience.

Examples

# Example 1: Generates a spatial optimized partially replicated arrangement design in one 
# location with 342 genotypes for a field with dimensions 25 rows x 18 cols. 
# Note that there are 280 genotypes unreplicated (only one time), 50 genotypes replicated 
# two times, and 10 genotypes replicated three times, and two checks 20 times each one.
SpatpREP1 <- partially_replicated(nrows = 25, 
                                  ncols = 18,  
                                  repGens = c(280,50,10,1,1),
                                  repUnits = c(1,2,3,20,20),
                                  planter = "cartesian", 
                                  plotNumber = 101,
                                  seed = 77)
SpatpREP1$infoDesign

#> $Greps
#> [1] 62
#> 
#> $RepGens
#>  [1] 20 20  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2
#> [26]  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2
#> [51]  2  2  2  2  2  2  2  2  2  2  2  2
#> 
#> $EntryReps
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> [51] 51 52 53 54 55 56 57 58 59 60 61 62
#> 

SpatpREP1$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row25  155   15   11  249  274   56  339  289   34   168     3     9   176
#> Row24    9  271   29   28  251  177  317    1  282   253    23   338   178
#> Row23   17  134   43  214  318  323  305  333  329     2   185     2     2
#> Row22  132   56  272   59  108  309   83  216   68    59   105   233   261
#> Row21    1  215   52  334   45  161  101   10  264     2   294   208    30
#> Row20    7   30  260  191  210    2  321  268   31   225   133    81    33
#> Row19  231  239    3    5    9   22   38  109    2    86   332    77   129
#> Row18  303   57   76  241  206  218    8  328   19    41    67    80   336
#> Row17  320  154    2  166  252  267  137  118  242   213   142   146   135
#> Row16  299   36    5  255    7  254   36  122  326   110    70   209    42
#> Row15   48  262  295  188  322   69  220  102  217   229    50     1   307
#> Row14    1   14   31   26  331  245    2  232   39   284    20    74     2
#> Row13   78  169  223  125   25  117   53    2    2    53     1     2   152
#> Row12    1  314  256   54  167    1  165  270  281   162   153     1    17
#> Row11   90    1   89  330  115    1  159   27    4    84   179     5    85
#> Row10  189  199   99   73   46   60  144   71  292    55     1    88    57
#> Row9   130  187  139   58   51   91    1  186  126    21    72    62    50
#> Row8   175  311  247   13  300  269   66   60  131   306    13   337     1
#> Row7   259  182   12   48  298   61   16  288  200    43   160    54   228
#> Row6   193  141   95  192   62  204  194   92   40   230   123   138   243
#> Row5   236   20  341  148  257  234   11   15  136   297    26   149   273
#> Row4    24  266  151  219   12   39   47  174  120   202    23    94   280
#> Row3    58   41  181    1  150  114  173   47   45    16   157   258   226
#> Row2   296   61  340    6   25   38  107    1  286   119   127    21    87
#> Row1    79  308   49  147    2  211  112    2   28    10   196   212   302
#>       Col14 Col15 Col16 Col17 Col18
#> Row25   315   325   244     2   100
#> Row24    35     1    46     6    65
#> Row23   235     8    82    18   316
#> Row22    42     3   277   111   207
#> Row21   312   128    35   285    44
#> Row20   227   335    12    98   246
#> Row19   342     1   197    97   324
#> Row18   221     2   319     2    32
#> Row17   279   250    75   237     8
#> Row16   170    96    11     2   203
#> Row15   327   222    14   171    63
#> Row14     7   301   183   310   143
#> Row13   116    93   304   291   124
#> Row12    51   290   205   195    22
#> Row11    33    19   265   163   158
#> Row10    55   184   240     4    37
#> Row9     27   104   248    24   106
#> Row8    313     2   263   224   190
#> Row7    121   164    37    64   287
#> Row6     18    49    29     2   145
#> Row5     32   238    10   103     1
#> Row4    198    40   172   276    52
#> Row3    278   140    34     4   113
#> Row2    156   201   283     1     1
#> Row1      6   293   180   275    44

SpatpREP1$plotNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row25  533  534  535  536  537  538  539  540  541   542   543   544   545
#> Row24  515  516  517  518  519  520  521  522  523   524   525   526   527
#> Row23  497  498  499  500  501  502  503  504  505   506   507   508   509
#> Row22  479  480  481  482  483  484  485  486  487   488   489   490   491
#> Row21  461  462  463  464  465  466  467  468  469   470   471   472   473
#> Row20  443  444  445  446  447  448  449  450  451   452   453   454   455
#> Row19  425  426  427  428  429  430  431  432  433   434   435   436   437
#> Row18  407  408  409  410  411  412  413  414  415   416   417   418   419
#> Row17  389  390  391  392  393  394  395  396  397   398   399   400   401
#> Row16  371  372  373  374  375  376  377  378  379   380   381   382   383
#> Row15  353  354  355  356  357  358  359  360  361   362   363   364   365
#> Row14  335  336  337  338  339  340  341  342  343   344   345   346   347
#> Row13  317  318  319  320  321  322  323  324  325   326   327   328   329
#> Row12  299  300  301  302  303  304  305  306  307   308   309   310   311
#> Row11  281  282  283  284  285  286  287  288  289   290   291   292   293
#> Row10  263  264  265  266  267  268  269  270  271   272   273   274   275
#> Row9   245  246  247  248  249  250  251  252  253   254   255   256   257
#> Row8   227  228  229  230  231  232  233  234  235   236   237   238   239
#> Row7   209  210  211  212  213  214  215  216  217   218   219   220   221
#> Row6   191  192  193  194  195  196  197  198  199   200   201   202   203
#> Row5   173  174  175  176  177  178  179  180  181   182   183   184   185
#> Row4   155  156  157  158  159  160  161  162  163   164   165   166   167
#> Row3   137  138  139  140  141  142  143  144  145   146   147   148   149
#> Row2   119  120  121  122  123  124  125  126  127   128   129   130   131
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18
#> Row25   546   547   548   549   550
#> Row24   528   529   530   531   532
#> Row23   510   511   512   513   514
#> Row22   492   493   494   495   496
#> Row21   474   475   476   477   478
#> Row20   456   457   458   459   460
#> Row19   438   439   440   441   442
#> Row18   420   421   422   423   424
#> Row17   402   403   404   405   406
#> Row16   384   385   386   387   388
#> Row15   366   367   368   369   370
#> Row14   348   349   350   351   352
#> Row13   330   331   332   333   334
#> Row12   312   313   314   315   316
#> Row11   294   295   296   297   298
#> Row10   276   277   278   279   280
#> Row9    258   259   260   261   262
#> Row8    240   241   242   243   244
#> Row7    222   223   224   225   226
#> Row6    204   205   206   207   208
#> Row5    186   187   188   189   190
#> Row4    168   169   170   171   172
#> Row3    150   151   152   153   154
#> Row2    132   133   134   135   136
#> Row1    114   115   116   117   118

head(SpatpREP1$fieldBook,12)

#>    ROW COLUMN ENTRY PLOT CHECKS       EXPT LOCATION LOC YEAR NAME
#> 1    1      1    79  101      0 20ExptpREP        1   1 2021  G79
#> 2    1      2   308  102      0 20ExptpREP        1   1 2021 G308
#> 3    1      3    49  103      1 20ExptpREP        1   1 2021  G49
#> 4    1      4   147  104      0 20ExptpREP        1   1 2021 G147
#> 5    1      5     2  105      1 20ExptpREP        1   1 2021   G2
#> 6    1      6   211  106      0 20ExptpREP        1   1 2021 G211
#> 7    1      7   112  107      0 20ExptpREP        1   1 2021 G112
#> 8    1      8     2  108      1 20ExptpREP        1   1 2021   G2
#> 9    1      9    28  109      1 20ExptpREP        1   1 2021  G28
#> 10   1     10    10  110      1 20ExptpREP        1   1 2021  G10
#> 11   1     11   196  111      0 20ExptpREP        1   1 2021 G196
#> 12   1     12   212  112      0 20ExptpREP        1   1 2021 G212

# Example 2: Generates a spatial optimized partially replicated arrangement design with 459 
# genotypes in a field with dimensions 30 rows x 20 cols. Note that there 351 genotypes 
# unreplicated (only one time), 75 genotypes replicated two times, and 33 genotypes
# replicated three times. In this case we don't have check plots.
SpatpREP2 <- partially_replicated(nrows = 30, 
                                  ncols = 20, 
                                  repGens = c(351,75,33),
                                  repUnits = c(1,2,3),
                                  planter = "serpentine", 
                                  plotNumber = 101,
                                  seed = 41)
SpatpREP2$infoDesign

#> $Greps
#> [1] 108
#> 
#> $RepGens
#>   [1] 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 3 2 2 2 2
#>  [38] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [75] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> 
#> $EntryReps
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> 

SpatpREP2$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30   10  206   64   83  247   11  148  108  369   348   149   315    30
#> Row29   97  439   18   48  298  153   19  289  418   399    62   332   105
#> Row28   65  367   71  449  404  345   94   61  419   146    30   220   129
#> Row27  390  230   76  362   74   60  127  436  136     8   328     3   106
#> Row26  139  355   53  286   13   50   49  407  278   104   329   212   172
#> Row25  450  120  443  116  210  239   75   44  159    82   379   204   183
#> Row24    9  282  253   12  283   91  455  426  205   266     3   459   222
#> Row23   64   26  274  324  319    6  306  275   17    58   143   142    93
#> Row22  140   33  281  147  380  295  190   43  100   119    28    56    94
#> Row21   20  246   39  238  360  307  227  251   51   100   145    88    82
#> Row20  344  382   69  103  300  262  113   40  157    70   171    47   406
#> Row19  197   20  339  291   17   89   54  330  293    45    32   325     4
#> Row18   23   85  333  270  107  162  337  327  214    48   458   431   411
#> Row17   90   13   23   17   28  108    8   42   37    78    11    31   174
#> Row16  191   27   24  389  351  372  249    7  413    24    32    42   232
#> Row15  313    4  277  393   55  447  173  451  388   109    69   358   428
#> Row14   85   65  405   81   92  354  181  347  117   334     5    24   199
#> Row13   67  179    5  287  400  200   84  364  101   356   252   170   392
#> Row12  391  417  178  243  150  137  366  188  203   317   235    31    26
#> Row11  126   35  201   86  285  163  102  198  310   154   318   299    22
#> Row10  414  186  423   46   27   41  155   63  433   342   397   259   130
#> Row9    33  394  237  422  273    2   33   10   83   208   216   421   386
#> Row8    39   12  264  215  441   97   28    6  312   336   427    80    50
#> Row7   320    9  248  412   25  118   32   93   14   343   234   135    57
#> Row6   256    4   96    3  144   34  444  437  101   374   370   268   448
#> Row5   308  440  194   26   38  395  416   87   35   138   224    98    74
#> Row4   271  438   25    1  297  265  131   63   95     6    22   207   260
#> Row3   357  112  202  442   19   25  403  350   86   226    55    98   290
#> Row2   125  165  340  103   66   36  187  217  304    90    54   255    99
#> Row1   169   18  314   76  377  122   41  279   52   376   244    21   429
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row30   168   111   385    58   425   182   164
#> Row29    79    37   124   368    91   156   424
#> Row28   435   415   294   309   218    77   311
#> Row27   128   229    45   301   189    29   383
#> Row26    89     7   110   410   242   373    21
#> Row25   303   231    15   335     7   398   322
#> Row24    52   240    78   180   363    38    60
#> Row23    16   305   211   196    72   141   133
#> Row22    15   241   408   177    80   359   195
#> Row21   353    92   225     8    73   228   209
#> Row20    29   371   387   192    67    59    84
#> Row19   341    22   245    87    72    81    16
#> Row18    15   257   409    36   236    95   261
#> Row17   402   166   430   321   263   176    59
#> Row16   250     1    75     2   457   420   123
#> Row15   106    34    62    70    61   323    57
#> Row14   338   316    79   280     5   105   184
#> Row13    77   288    30    13     1   284    99
#> Row12   132   361   302    14   272    73    29
#> Row11    19   396   115    10   151   456    66
#> Row10   401    51   375   432    21    71   296
#> Row9     18   107   331   223   276    49   158
#> Row8    213   175    68    44   378    31    53
#> Row7    381   102    11   114   326   267    12
#> Row6    121   167    16   161   221   352   134
#> Row5     88    14   104     9    47   384   349
#> Row4     46    27   193     2   446    40    23
#> Row3    269   445    68    56   185   454   254
#> Row2    434   346   219   365   452   233    43
#> Row1    453    96   258   292   152    20   160

SpatpREP2$plotNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  700  699  698  697  696  695  694  693  692   691   690   689   688
#> Row29  661  662  663  664  665  666  667  668  669   670   671   672   673
#> Row28  660  659  658  657  656  655  654  653  652   651   650   649   648
#> Row27  621  622  623  624  625  626  627  628  629   630   631   632   633
#> Row26  620  619  618  617  616  615  614  613  612   611   610   609   608
#> Row25  581  582  583  584  585  586  587  588  589   590   591   592   593
#> Row24  580  579  578  577  576  575  574  573  572   571   570   569   568
#> Row23  541  542  543  544  545  546  547  548  549   550   551   552   553
#> Row22  540  539  538  537  536  535  534  533  532   531   530   529   528
#> Row21  501  502  503  504  505  506  507  508  509   510   511   512   513
#> Row20  500  499  498  497  496  495  494  493  492   491   490   489   488
#> Row19  461  462  463  464  465  466  467  468  469   470   471   472   473
#> Row18  460  459  458  457  456  455  454  453  452   451   450   449   448
#> Row17  421  422  423  424  425  426  427  428  429   430   431   432   433
#> Row16  420  419  418  417  416  415  414  413  412   411   410   409   408
#> Row15  381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row14  380  379  378  377  376  375  374  373  372   371   370   369   368
#> Row13  341  342  343  344  345  346  347  348  349   350   351   352   353
#> Row12  340  339  338  337  336  335  334  333  332   331   330   329   328
#> Row11  301  302  303  304  305  306  307  308  309   310   311   312   313
#> Row10  300  299  298  297  296  295  294  293  292   291   290   289   288
#> Row9   261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row8   260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row7   221  222  223  224  225  226  227  228  229   230   231   232   233
#> Row6   220  219  218  217  216  215  214  213  212   211   210   209   208
#> Row5   181  182  183  184  185  186  187  188  189   190   191   192   193
#> Row4   180  179  178  177  176  175  174  173  172   171   170   169   168
#> Row3   141  142  143  144  145  146  147  148  149   150   151   152   153
#> Row2   140  139  138  137  136  135  134  133  132   131   130   129   128
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row30   687   686   685   684   683   682   681
#> Row29   674   675   676   677   678   679   680
#> Row28   647   646   645   644   643   642   641
#> Row27   634   635   636   637   638   639   640
#> Row26   607   606   605   604   603   602   601
#> Row25   594   595   596   597   598   599   600
#> Row24   567   566   565   564   563   562   561
#> Row23   554   555   556   557   558   559   560
#> Row22   527   526   525   524   523   522   521
#> Row21   514   515   516   517   518   519   520
#> Row20   487   486   485   484   483   482   481
#> Row19   474   475   476   477   478   479   480
#> Row18   447   446   445   444   443   442   441
#> Row17   434   435   436   437   438   439   440
#> Row16   407   406   405   404   403   402   401
#> Row15   394   395   396   397   398   399   400
#> Row14   367   366   365   364   363   362   361
#> Row13   354   355   356   357   358   359   360
#> Row12   327   326   325   324   323   322   321
#> Row11   314   315   316   317   318   319   320
#> Row10   287   286   285   284   283   282   281
#> Row9    274   275   276   277   278   279   280
#> Row8    247   246   245   244   243   242   241
#> Row7    234   235   236   237   238   239   240
#> Row6    207   206   205   204   203   202   201
#> Row5    194   195   196   197   198   199   200
#> Row4    167   166   165   164   163   162   161
#> Row3    154   155   156   157   158   159   160
#> Row2    127   126   125   124   123   122   121
#> Row1    114   115   116   117   118   119   120

head(SpatpREP2$fieldBook,10)

#>    ROW COLUMN ENTRY PLOT CHECKS       EXPT LOCATION LOC YEAR NAME
#> 1    1      1   169  101      0 20ExptpREP        1   1 2021 G169
#> 2    1      2    18  102      1 20ExptpREP        1   1 2021  G18
#> 3    1      3   314  103      0 20ExptpREP        1   1 2021 G314
#> 4    1      4    76  104      1 20ExptpREP        1   1 2021  G76
#> 5    1      5   377  105      0 20ExptpREP        1   1 2021 G377
#> 6    1      6   122  106      0 20ExptpREP        1   1 2021 G122
#> 7    1      7    41  107      1 20ExptpREP        1   1 2021  G41
#> 8    1      8   279  108      0 20ExptpREP        1   1 2021 G279
#> 9    1      9    52  109      1 20ExptpREP        1   1 2021  G52
#> 10   1     10   376  110      0 20ExptpREP        1   1 2021 G376