Print an FielDHub object

Prints information about randomization of Completely Randomized Design (CRD)

# S3 method for CRD
print(x, n, ...)

Arguments

x
an object inheriting from class

n
a single integer. If positive or zero, size for the resulting object: number of elements for a vector (including lists), rows for a matrix or data frame or lines for a function. If negative, all but the n last/first number of elements of x.

...
further arguments passed to head.

Value

an object inheriting from class FielDHub

Author

Thiago de Paula Oliveira, thiago.paula.oliveira@alumni.usp.br

Examples

# Example 1: Generates a CRD design with 5 treatments and 5 reps each.
crd1 <- CRD(t = 5, reps = 5, plotNumber = 101,
seed = 1985, locationName = "Fargo")
crd1$infoDesign

#> $numberofTreatments
#> [1] 5
#> 
#> $treatments
#> [1] "T1" "T2" "T3" "T4" "T5"
#> 
#> $Reps
#> [1] 5
#> 
#> $locationName
#> [1] "Fargo"
#> 
#> $seed
#> [1] 1985
#> 

print(crd1)

#> $infoDesign
#> $infoDesign$numberofTreatments
#> [1] 5
#> 
#> $infoDesign$treatments
#> [1] "T1" "T2" "T3" "T4" "T5"
#> 
#> $infoDesign$Reps
#> [1] 5
#> 
#> $infoDesign$locationName
#> [1] "Fargo"
#> 
#> $infoDesign$seed
#> [1] 1985
#> 
#> 
#> $fieldBook
#>    ID LOCATION PLOT REP TREATMENT
#> 1   1    Fargo  101   5        T3
#> 2   2    Fargo  102   4        T5
#> 3   3    Fargo  103   3        T2
#> 4   4    Fargo  104   5        T1
#> 5   5    Fargo  105   3        T1
#> 6   6    Fargo  106   3        T3
#> 7   7    Fargo  107   2        T1
#> 8   8    Fargo  108   2        T4
#> 9   9    Fargo  109   5        T2
#> 10 10    Fargo  110   4        T2
#> 11 11    Fargo  111   2        T5
#> 12 12    Fargo  112   2        T3
#> 13 13    Fargo  113   4        T3
#> 14 14    Fargo  114   5        T4
#> 15 15    Fargo  115   4        T4
#> 16 16    Fargo  116   2        T2
#> 17 17    Fargo  117   3        T4
#> 18 18    Fargo  118   1        T4
#> 19 19    Fargo  119   5        T5
#> 20 20    Fargo  120   1        T5
#> 21 21    Fargo  121   1        T3
#> 22 22    Fargo  122   4        T1
#> 23 23    Fargo  123   3        T5
#> 24 24    Fargo  124   1        T1
#> 25 25    Fargo  125   1        T2
#> 
#> attr(,"class")
#> [1] "FielDHub"