Print an FielDHub object

Prints information about latin square design

# S3 method for latin_square
print(x, n, ...)

Arguments

x
an object inheriting from class

n
a single integer. If positive or zero, size for the resulting object: number of elements for a vector (including lists), rows for a matrix or data frame or lines for a function. If negative, all but the n last/first number of elements of x.

...
further arguments passed to head.

Value

an object inheriting from class FielDHub

Author

Thiago de Paula Oliveira, thiago.paula.oliveira@alumni.usp.br

Examples

# Example 1: Generates a latin square design with 4 treatments and 2 reps.
latinSq1 <- latin_square(t = 4,
                         reps = 2,
                         plotNumber = 101,
                         planter = "cartesian",
                         seed = 1980)
print(latinSq1)

#> $infoDesign
#> $infoDesign$numberofTreatments
#> [1] 4
#> 
#> $infoDesign$numberofRowns
#> [1] 4
#> 
#> $infoDesign$numberofColumns
#> [1] 4
#> 
#> $infoDesign$rep
#> [1] 2
#> 
#> $infoDesign$locationName
#> NULL
#> 
#> $infoDesign$seed
#> [1] 1980
#> 
#> 
#> $squares
#> $squares$rep1
#>       Column 1 Column 2 Column 3 Column 4
#> Row 1 "T1"     "T4"     "T2"     "T3"    
#> Row 2 "T3"     "T1"     "T4"     "T2"    
#> Row 3 "T4"     "T2"     "T3"     "T1"    
#> Row 4 "T2"     "T3"     "T1"     "T4"    
#> 
#> $squares$rep2
#>       Column 1 Column 2 Column 3 Column 4
#> Row 1 "T1"     "T3"     "T4"     "T2"    
#> Row 2 "T2"     "T4"     "T3"     "T1"    
#> Row 3 "T4"     "T1"     "T2"     "T3"    
#> Row 4 "T3"     "T2"     "T1"     "T4"    
#> 
#> 
#> $plotSquares
#> $plotSquares$rep1
#>      [,1] [,2] [,3] [,4]
#> [1,]  101  102  103  104
#> [2,]  105  106  107  108
#> [3,]  109  110  111  112
#> [4,]  113  114  115  116
#> 
#> $plotSquares$rep2
#>      [,1] [,2] [,3] [,4]
#> [1,]  201  202  203  204
#> [2,]  205  206  207  208
#> [3,]  209  210  211  212
#> [4,]  213  214  215  216
#> 
#> 
#> $fieldBook
#>    ID LOCATION PLOT SQUARE   ROW   COLUMN TREATMENT
#> 1   1        1  101      1 Row 1 Column 1        T1
#> 2   2        1  102      1 Row 1 Column 2        T4
#> 3   3        1  103      1 Row 1 Column 3        T2
#> 4   4        1  104      1 Row 1 Column 4        T3
#> 5   5        1  105      1 Row 2 Column 1        T3
#> 6   6        1  106      1 Row 2 Column 2        T1
#> 7   7        1  107      1 Row 2 Column 3        T4
#> 8   8        1  108      1 Row 2 Column 4        T2
#> 9   9        1  109      1 Row 3 Column 1        T4
#> 10 10        1  110      1 Row 3 Column 2        T2
#> 11 11        1  111      1 Row 3 Column 3        T3
#> 12 12        1  112      1 Row 3 Column 4        T1
#> 13 13        1  113      1 Row 4 Column 1        T2
#> 14 14        1  114      1 Row 4 Column 2        T3
#> 15 15        1  115      1 Row 4 Column 3        T1
#> 16 16        1  116      1 Row 4 Column 4        T4
#> 17 17        1  201      2 Row 1 Column 1        T1
#> 18 18        1  202      2 Row 1 Column 2        T3
#> 19 19        1  203      2 Row 1 Column 3        T4
#> 20 20        1  204      2 Row 1 Column 4        T2
#> 21 21        1  205      2 Row 2 Column 1        T2
#> 22 22        1  206      2 Row 2 Column 2        T4
#> 23 23        1  207      2 Row 2 Column 3        T3
#> 24 24        1  208      2 Row 2 Column 4        T1
#> 25 25        1  209      2 Row 3 Column 1        T4
#> 26 26        1  210      2 Row 3 Column 2        T1
#> 27 27        1  211      2 Row 3 Column 3        T2
#> 28 28        1  212      2 Row 3 Column 4        T3
#> 29 29        1  213      2 Row 4 Column 1        T3
#> 30 30        1  214      2 Row 4 Column 2        T2
#> 31 31        1  215      2 Row 4 Column 3        T1
#> 32 32        1  216      2 Row 4 Column 4        T4
#> 
#> attr(,"class")
#> [1] "FielDHub"