Strip Plot Design

It randomly generates a strip plot design across locations.

strip_plot(
  Hplots = NULL,
  Vplots = NULL,
  b = 1,
  l = 1,
  plotNumber = NULL,
  planter = "serpentine",
  locationNames = NULL,
  seed = NULL,
  data = NULL
)

Arguments

Hplots
Number of horizontal factors, as an integer or a vector.

Vplots
Number of vertical factors, as an integer or a vector.

b
Number of blocks (full replicates).

l
Number of locations. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

planter
Option for serpentine or cartesian arrangement. By default planter = 'serpentine'.

locationNames
(optional) Names for each location.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

data
(optional) data frame with the labels of vertical and hirizontal plots.

Value

A list with information on the design parameters.

Strip blocks for each location.

Data frame with the strip plot field book.

References

Design and Analysis of Experiments, Volume 1, Introduction to Experimental Design. Second Edition. Klaus Hinkelmann & Oscar Kempthorne.John Wiley & Sons, Inc., Hoboken, New Jersey.

Examples

# Example 1: Generates a strip plot design with 5 vertical strips and 4 horizontal strips,
# with 3 reps in one location.
H <- paste("H", 1:4, sep = "")
V <- paste("V", 1:5, sep = "")
strip1 <- strip_plot(Hplots = H, 
                     Vplots = V, 
                     b = 3, 
                     l = 1, 
                     plotNumber = 101,
                     planter = "serpentine",
                     locationNames = "A", 
                     seed = 333)
strip1$infoDesign                  

#> $Hplots
#> [1] 4
#> 
#> $Vplots
#> [1] 5
#> 
#> $blocks
#> [1] 3
#> 
#> $numberLocations
#> [1] 1
#> 
#> $nameLocations
#> [1] "A"
#> 
#> $seed
#> [1] 333
#> 

strip1$stripsBlockLoc

#> [[1]]
#> [[1]][[1]]
#>    V4      V2      V5      V1      V3     
#> H2 "H2|V4" "H2|V2" "H2|V5" "H2|V1" "H2|V3"
#> H1 "H1|V4" "H1|V2" "H1|V5" "H1|V1" "H1|V3"
#> H3 "H3|V4" "H3|V2" "H3|V5" "H3|V1" "H3|V3"
#> H4 "H4|V4" "H4|V2" "H4|V5" "H4|V1" "H4|V3"
#> 
#> [[1]][[2]]
#>    V1      V3      V4      V2      V5     
#> H3 "H3|V1" "H3|V3" "H3|V4" "H3|V2" "H3|V5"
#> H4 "H4|V1" "H4|V3" "H4|V4" "H4|V2" "H4|V5"
#> H2 "H2|V1" "H2|V3" "H2|V4" "H2|V2" "H2|V5"
#> H1 "H1|V1" "H1|V3" "H1|V4" "H1|V2" "H1|V5"
#> 
#> [[1]][[3]]
#>    V3      V1      V2      V4      V5     
#> H2 "H2|V3" "H2|V1" "H2|V2" "H2|V4" "H2|V5"
#> H1 "H1|V3" "H1|V1" "H1|V2" "H1|V4" "H1|V5"
#> H4 "H4|V3" "H4|V1" "H4|V2" "H4|V4" "H4|V5"
#> H3 "H3|V3" "H3|V1" "H3|V2" "H3|V4" "H3|V5"
#> 
#> 

strip1$plotLayouts

#> [[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  101  102  103  104  105
#> [2,]  110  109  108  107  106
#> [3,]  111  112  113  114  115
#> [4,]  120  119  118  117  116
#> 
#> [[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  201  202  203  204  205
#> [2,]  210  209  208  207  206
#> [3,]  211  212  213  214  215
#> [4,]  220  219  218  217  216
#> 
#> [[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  301  302  303  304  305
#> [2,]  310  309  308  307  306
#> [3,]  311  312  313  314  315
#> [4,]  320  319  318  317  316
#> 

head(strip1$fieldBook,12)                     

#>    LOCATION PLOT REP HSTRIP VSTRIP TRT_COMB
#> 1         A  101   1     H2     V4    H2|V4
#> 2         A  102   1     H2     V2    H2|V2
#> 3         A  103   1     H2     V5    H2|V5
#> 4         A  104   1     H2     V1    H2|V1
#> 5         A  105   1     H2     V3    H2|V3
#> 6         A  110   1     H1     V4    H1|V4
#> 7         A  109   1     H1     V2    H1|V2
#> 8         A  108   1     H1     V5    H1|V5
#> 9         A  107   1     H1     V1    H1|V1
#> 10        A  106   1     H1     V3    H1|V3
#> 11        A  111   1     H3     V4    H3|V4
#> 12        A  112   1     H3     V2    H3|V2

# Example 2: Generates a strip plot design with 5 vertical strips and 5 horizontal strips,
# with 6 reps across to 3 locations.
strip2 <- strip_plot(Hplots = 5, 
                     Vplots = 5, 
                     b = 6, 
                     l = 3, 
                     plotNumber = c(101,1001,2001),
                     planter = "cartesian",
                     locationNames = c("A", "B", "C"), 
                     seed = 222)
strip2$infoDesign                  

#> $Hplots
#> [1] 5
#> 
#> $Vplots
#> [1] 5
#> 
#> $blocks
#> [1] 6
#> 
#> $numberLocations
#> [1] 3
#> 
#> $nameLocations
#> [1] "A" "B" "C"
#> 
#> $seed
#> [1] 222
#> 

strip2$stripsBlockLoc

#> [[1]]
#> [[1]][[1]]
#>    a1      a3      a0      a4      a2     
#> b4 "b4|a1" "b4|a3" "b4|a0" "b4|a4" "b4|a2"
#> b1 "b1|a1" "b1|a3" "b1|a0" "b1|a4" "b1|a2"
#> b2 "b2|a1" "b2|a3" "b2|a0" "b2|a4" "b2|a2"
#> b3 "b3|a1" "b3|a3" "b3|a0" "b3|a4" "b3|a2"
#> b0 "b0|a1" "b0|a3" "b0|a0" "b0|a4" "b0|a2"
#> 
#> [[1]][[2]]
#>    a0      a3      a4      a1      a2     
#> b1 "b1|a0" "b1|a3" "b1|a4" "b1|a1" "b1|a2"
#> b0 "b0|a0" "b0|a3" "b0|a4" "b0|a1" "b0|a2"
#> b4 "b4|a0" "b4|a3" "b4|a4" "b4|a1" "b4|a2"
#> b2 "b2|a0" "b2|a3" "b2|a4" "b2|a1" "b2|a2"
#> b3 "b3|a0" "b3|a3" "b3|a4" "b3|a1" "b3|a2"
#> 
#> [[1]][[3]]
#>    a0      a2      a4      a1      a3     
#> b4 "b4|a0" "b4|a2" "b4|a4" "b4|a1" "b4|a3"
#> b0 "b0|a0" "b0|a2" "b0|a4" "b0|a1" "b0|a3"
#> b3 "b3|a0" "b3|a2" "b3|a4" "b3|a1" "b3|a3"
#> b2 "b2|a0" "b2|a2" "b2|a4" "b2|a1" "b2|a3"
#> b1 "b1|a0" "b1|a2" "b1|a4" "b1|a1" "b1|a3"
#> 
#> [[1]][[4]]
#>    a4      a2      a0      a3      a1     
#> b4 "b4|a4" "b4|a2" "b4|a0" "b4|a3" "b4|a1"
#> b1 "b1|a4" "b1|a2" "b1|a0" "b1|a3" "b1|a1"
#> b0 "b0|a4" "b0|a2" "b0|a0" "b0|a3" "b0|a1"
#> b3 "b3|a4" "b3|a2" "b3|a0" "b3|a3" "b3|a1"
#> b2 "b2|a4" "b2|a2" "b2|a0" "b2|a3" "b2|a1"
#> 
#> [[1]][[5]]
#>    a3      a4      a1      a2      a0     
#> b4 "b4|a3" "b4|a4" "b4|a1" "b4|a2" "b4|a0"
#> b1 "b1|a3" "b1|a4" "b1|a1" "b1|a2" "b1|a0"
#> b2 "b2|a3" "b2|a4" "b2|a1" "b2|a2" "b2|a0"
#> b3 "b3|a3" "b3|a4" "b3|a1" "b3|a2" "b3|a0"
#> b0 "b0|a3" "b0|a4" "b0|a1" "b0|a2" "b0|a0"
#> 
#> [[1]][[6]]
#>    a1      a3      a4      a0      a2     
#> b1 "b1|a1" "b1|a3" "b1|a4" "b1|a0" "b1|a2"
#> b3 "b3|a1" "b3|a3" "b3|a4" "b3|a0" "b3|a2"
#> b2 "b2|a1" "b2|a3" "b2|a4" "b2|a0" "b2|a2"
#> b4 "b4|a1" "b4|a3" "b4|a4" "b4|a0" "b4|a2"
#> b0 "b0|a1" "b0|a3" "b0|a4" "b0|a0" "b0|a2"
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#>    a3      a4      a2      a0      a1     
#> b4 "b4|a3" "b4|a4" "b4|a2" "b4|a0" "b4|a1"
#> b3 "b3|a3" "b3|a4" "b3|a2" "b3|a0" "b3|a1"
#> b0 "b0|a3" "b0|a4" "b0|a2" "b0|a0" "b0|a1"
#> b2 "b2|a3" "b2|a4" "b2|a2" "b2|a0" "b2|a1"
#> b1 "b1|a3" "b1|a4" "b1|a2" "b1|a0" "b1|a1"
#> 
#> [[2]][[2]]
#>    a3      a0      a4      a1      a2     
#> b1 "b1|a3" "b1|a0" "b1|a4" "b1|a1" "b1|a2"
#> b4 "b4|a3" "b4|a0" "b4|a4" "b4|a1" "b4|a2"
#> b2 "b2|a3" "b2|a0" "b2|a4" "b2|a1" "b2|a2"
#> b3 "b3|a3" "b3|a0" "b3|a4" "b3|a1" "b3|a2"
#> b0 "b0|a3" "b0|a0" "b0|a4" "b0|a1" "b0|a2"
#> 
#> [[2]][[3]]
#>    a2      a0      a3      a1      a4     
#> b3 "b3|a2" "b3|a0" "b3|a3" "b3|a1" "b3|a4"
#> b2 "b2|a2" "b2|a0" "b2|a3" "b2|a1" "b2|a4"
#> b4 "b4|a2" "b4|a0" "b4|a3" "b4|a1" "b4|a4"
#> b0 "b0|a2" "b0|a0" "b0|a3" "b0|a1" "b0|a4"
#> b1 "b1|a2" "b1|a0" "b1|a3" "b1|a1" "b1|a4"
#> 
#> [[2]][[4]]
#>    a0      a3      a2      a1      a4     
#> b4 "b4|a0" "b4|a3" "b4|a2" "b4|a1" "b4|a4"
#> b3 "b3|a0" "b3|a3" "b3|a2" "b3|a1" "b3|a4"
#> b1 "b1|a0" "b1|a3" "b1|a2" "b1|a1" "b1|a4"
#> b0 "b0|a0" "b0|a3" "b0|a2" "b0|a1" "b0|a4"
#> b2 "b2|a0" "b2|a3" "b2|a2" "b2|a1" "b2|a4"
#> 
#> [[2]][[5]]
#>    a4      a3      a1      a2      a0     
#> b4 "b4|a4" "b4|a3" "b4|a1" "b4|a2" "b4|a0"
#> b0 "b0|a4" "b0|a3" "b0|a1" "b0|a2" "b0|a0"
#> b3 "b3|a4" "b3|a3" "b3|a1" "b3|a2" "b3|a0"
#> b2 "b2|a4" "b2|a3" "b2|a1" "b2|a2" "b2|a0"
#> b1 "b1|a4" "b1|a3" "b1|a1" "b1|a2" "b1|a0"
#> 
#> [[2]][[6]]
#>    a4      a3      a0      a2      a1     
#> b0 "b0|a4" "b0|a3" "b0|a0" "b0|a2" "b0|a1"
#> b4 "b4|a4" "b4|a3" "b4|a0" "b4|a2" "b4|a1"
#> b3 "b3|a4" "b3|a3" "b3|a0" "b3|a2" "b3|a1"
#> b1 "b1|a4" "b1|a3" "b1|a0" "b1|a2" "b1|a1"
#> b2 "b2|a4" "b2|a3" "b2|a0" "b2|a2" "b2|a1"
#> 
#> 
#> [[3]]
#> [[3]][[1]]
#>    a3      a2      a4      a0      a1     
#> b1 "b1|a3" "b1|a2" "b1|a4" "b1|a0" "b1|a1"
#> b2 "b2|a3" "b2|a2" "b2|a4" "b2|a0" "b2|a1"
#> b0 "b0|a3" "b0|a2" "b0|a4" "b0|a0" "b0|a1"
#> b3 "b3|a3" "b3|a2" "b3|a4" "b3|a0" "b3|a1"
#> b4 "b4|a3" "b4|a2" "b4|a4" "b4|a0" "b4|a1"
#> 
#> [[3]][[2]]
#>    a3      a1      a4      a2      a0     
#> b3 "b3|a3" "b3|a1" "b3|a4" "b3|a2" "b3|a0"
#> b0 "b0|a3" "b0|a1" "b0|a4" "b0|a2" "b0|a0"
#> b1 "b1|a3" "b1|a1" "b1|a4" "b1|a2" "b1|a0"
#> b4 "b4|a3" "b4|a1" "b4|a4" "b4|a2" "b4|a0"
#> b2 "b2|a3" "b2|a1" "b2|a4" "b2|a2" "b2|a0"
#> 
#> [[3]][[3]]
#>    a1      a4      a0      a3      a2     
#> b4 "b4|a1" "b4|a4" "b4|a0" "b4|a3" "b4|a2"
#> b0 "b0|a1" "b0|a4" "b0|a0" "b0|a3" "b0|a2"
#> b1 "b1|a1" "b1|a4" "b1|a0" "b1|a3" "b1|a2"
#> b3 "b3|a1" "b3|a4" "b3|a0" "b3|a3" "b3|a2"
#> b2 "b2|a1" "b2|a4" "b2|a0" "b2|a3" "b2|a2"
#> 
#> [[3]][[4]]
#>    a1      a3      a2      a0      a4     
#> b4 "b4|a1" "b4|a3" "b4|a2" "b4|a0" "b4|a4"
#> b3 "b3|a1" "b3|a3" "b3|a2" "b3|a0" "b3|a4"
#> b2 "b2|a1" "b2|a3" "b2|a2" "b2|a0" "b2|a4"
#> b0 "b0|a1" "b0|a3" "b0|a2" "b0|a0" "b0|a4"
#> b1 "b1|a1" "b1|a3" "b1|a2" "b1|a0" "b1|a4"
#> 
#> [[3]][[5]]
#>    a4      a3      a2      a1      a0     
#> b1 "b1|a4" "b1|a3" "b1|a2" "b1|a1" "b1|a0"
#> b4 "b4|a4" "b4|a3" "b4|a2" "b4|a1" "b4|a0"
#> b0 "b0|a4" "b0|a3" "b0|a2" "b0|a1" "b0|a0"
#> b3 "b3|a4" "b3|a3" "b3|a2" "b3|a1" "b3|a0"
#> b2 "b2|a4" "b2|a3" "b2|a2" "b2|a1" "b2|a0"
#> 
#> [[3]][[6]]
#>    a2      a3      a1      a4      a0     
#> b0 "b0|a2" "b0|a3" "b0|a1" "b0|a4" "b0|a0"
#> b1 "b1|a2" "b1|a3" "b1|a1" "b1|a4" "b1|a0"
#> b3 "b3|a2" "b3|a3" "b3|a1" "b3|a4" "b3|a0"
#> b4 "b4|a2" "b4|a3" "b4|a1" "b4|a4" "b4|a0"
#> b2 "b2|a2" "b2|a3" "b2|a1" "b2|a4" "b2|a0"
#> 
#> 

strip2$plotLayouts

#> [[1]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  101  102  103  104  105
#> [2,]  106  107  108  109  110
#> [3,]  111  112  113  114  115
#> [4,]  116  117  118  119  120
#> [5,]  121  122  123  124  125
#> 
#> [[2]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  201  202  203  204  205
#> [2,]  206  207  208  209  210
#> [3,]  211  212  213  214  215
#> [4,]  216  217  218  219  220
#> [5,]  221  222  223  224  225
#> 
#> [[3]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  301  302  303  304  305
#> [2,]  306  307  308  309  310
#> [3,]  311  312  313  314  315
#> [4,]  316  317  318  319  320
#> [5,]  321  322  323  324  325
#> 
#> [[4]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  401  402  403  404  405
#> [2,]  406  407  408  409  410
#> [3,]  411  412  413  414  415
#> [4,]  416  417  418  419  420
#> [5,]  421  422  423  424  425
#> 
#> [[5]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  501  502  503  504  505
#> [2,]  506  507  508  509  510
#> [3,]  511  512  513  514  515
#> [4,]  516  517  518  519  520
#> [5,]  521  522  523  524  525
#> 
#> [[6]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]  601  602  603  604  605
#> [2,]  606  607  608  609  610
#> [3,]  611  612  613  614  615
#> [4,]  616  617  618  619  620
#> [5,]  621  622  623  624  625
#> 
#> [[7]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1001 1002 1003 1004 1005
#> [2,] 1006 1007 1008 1009 1010
#> [3,] 1011 1012 1013 1014 1015
#> [4,] 1016 1017 1018 1019 1020
#> [5,] 1021 1022 1023 1024 1025
#> 
#> [[8]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1101 1102 1103 1104 1105
#> [2,] 1106 1107 1108 1109 1110
#> [3,] 1111 1112 1113 1114 1115
#> [4,] 1116 1117 1118 1119 1120
#> [5,] 1121 1122 1123 1124 1125
#> 
#> [[9]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1201 1202 1203 1204 1205
#> [2,] 1206 1207 1208 1209 1210
#> [3,] 1211 1212 1213 1214 1215
#> [4,] 1216 1217 1218 1219 1220
#> [5,] 1221 1222 1223 1224 1225
#> 
#> [[10]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1301 1302 1303 1304 1305
#> [2,] 1306 1307 1308 1309 1310
#> [3,] 1311 1312 1313 1314 1315
#> [4,] 1316 1317 1318 1319 1320
#> [5,] 1321 1322 1323 1324 1325
#> 
#> [[11]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1401 1402 1403 1404 1405
#> [2,] 1406 1407 1408 1409 1410
#> [3,] 1411 1412 1413 1414 1415
#> [4,] 1416 1417 1418 1419 1420
#> [5,] 1421 1422 1423 1424 1425
#> 
#> [[12]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 1501 1502 1503 1504 1505
#> [2,] 1506 1507 1508 1509 1510
#> [3,] 1511 1512 1513 1514 1515
#> [4,] 1516 1517 1518 1519 1520
#> [5,] 1521 1522 1523 1524 1525
#> 
#> [[13]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 2001 2002 2003 2004 2005
#> [2,] 2006 2007 2008 2009 2010
#> [3,] 2011 2012 2013 2014 2015
#> [4,] 2016 2017 2018 2019 2020
#> [5,] 2021 2022 2023 2024 2025
#> 
#> [[14]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 2101 2102 2103 2104 2105
#> [2,] 2106 2107 2108 2109 2110
#> [3,] 2111 2112 2113 2114 2115
#> [4,] 2116 2117 2118 2119 2120
#> [5,] 2121 2122 2123 2124 2125
#> 
#> [[15]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 2201 2202 2203 2204 2205
#> [2,] 2206 2207 2208 2209 2210
#> [3,] 2211 2212 2213 2214 2215
#> [4,] 2216 2217 2218 2219 2220
#> [5,] 2221 2222 2223 2224 2225
#> 
#> [[16]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 2301 2302 2303 2304 2305
#> [2,] 2306 2307 2308 2309 2310
#> [3,] 2311 2312 2313 2314 2315
#> [4,] 2316 2317 2318 2319 2320
#> [5,] 2321 2322 2323 2324 2325
#> 
#> [[17]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 2401 2402 2403 2404 2405
#> [2,] 2406 2407 2408 2409 2410
#> [3,] 2411 2412 2413 2414 2415
#> [4,] 2416 2417 2418 2419 2420
#> [5,] 2421 2422 2423 2424 2425
#> 
#> [[18]]
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,] 2501 2502 2503 2504 2505
#> [2,] 2506 2507 2508 2509 2510
#> [3,] 2511 2512 2513 2514 2515
#> [4,] 2516 2517 2518 2519 2520
#> [5,] 2521 2522 2523 2524 2525
#> 

head(strip2$fieldBook,12)

#>    LOCATION PLOT REP HSTRIP VSTRIP TRT_COMB
#> 1         A  101   1     b4     a1    b4|a1
#> 2         A  102   1     b4     a3    b4|a3
#> 3         A  103   1     b4     a0    b4|a0
#> 4         A  104   1     b4     a4    b4|a4
#> 5         A  105   1     b4     a2    b4|a2
#> 6         A  106   1     b1     a1    b1|a1
#> 7         A  107   1     b1     a3    b1|a3
#> 8         A  108   1     b1     a0    b1|a0
#> 9         A  109   1     b1     a4    b1|a4
#> 10        A  110   1     b1     a2    b1|a2
#> 11        A  111   1     b2     a1    b2|a1
#> 12        A  112   1     b2     a3    b2|a3