Generates a Completely Randomized Design (CRD)

It randomly generates a completely randomized design.

CRD(
  t = NULL,
  reps = NULL,
  plotNumber = 101,
  locationName = NULL,
  seed = NULL,
  data = NULL
)

Arguments

t
an integer number with total number of treatments or a vector of dimension t with labels.

reps
Number of replicates (full blocks) of each treatment.

plotNumber
Starting plot number. By default plotNumber = 101.

locationName
(optional) Name of the location.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

data
(optional) Data frame with the 2 columns with labels of each treatments and its number of replicates.

Value

A list with information on the design parameters.

Data frame with the CRD field book.

References

Design and Analysis of Experiments, Volume 1, Introduction to Experimental Design. Second Edition. Klaus Hinkelmann & Oscar Kempthorne.John Wiley & Sons, Inc., Hoboken, New Jersey.

Examples

# Example 1: Generates a CRD design with 10 treatments and 5 reps each.
crd1 <- CRD(t = 10, 
            reps = 5, 
            plotNumber = 101, 
            seed = 1987, 
            locationName = "Fargo")
crd1$infoDesign

#> $numberofTreatments
#> [1] 10
#> 
#> $treatments
#>  [1] "T1"  "T2"  "T3"  "T4"  "T5"  "T6"  "T7"  "T8"  "T9"  "T10"
#> 
#> $Reps
#> [1] 5
#> 
#> $locationName
#> [1] "Fargo"
#> 
#> $seed
#> [1] 1987
#> 

head(crd1$fieldBook,10)

#>    LOCATION PLOT REP TREATMENT
#> 1     Fargo  101   3        T5
#> 2     Fargo  102   3        T7
#> 3     Fargo  103   2        T1
#> 4     Fargo  104   2        T5
#> 5     Fargo  105   5        T5
#> 6     Fargo  106   3        T3
#> 7     Fargo  107   3       T10
#> 8     Fargo  108   2        T8
#> 9     Fargo  109   3        T9
#> 10    Fargo  110   3        T8

# Example 2: Generates a CRD design with 15 treatments and 6 reps each.
Gens <- paste("Wheat", 1:15, sep = "")
crd2 <- CRD(t = Gens, 
            reps = 6, 
            plotNumber = 1001, 
            seed = 1654, 
            locationName = "Fargo")
crd2$infoDesign

#> $numberofTreatments
#> [1] 15
#> 
#> $treatments
#>  [1] "Wheat1"  "Wheat2"  "Wheat3"  "Wheat4"  "Wheat5"  "Wheat6"  "Wheat7" 
#>  [8] "Wheat8"  "Wheat9"  "Wheat10" "Wheat11" "Wheat12" "Wheat13" "Wheat14"
#> [15] "Wheat15"
#> 
#> $Reps
#> [1] 6
#> 
#> $locationName
#> [1] "Fargo"
#> 
#> $seed
#> [1] 1654
#> 

head(crd2$fieldBook,10)

#>    LOCATION PLOT REP TREATMENT
#> 1     Fargo 1001   6    Wheat3
#> 2     Fargo 1002   2    Wheat8
#> 3     Fargo 1003   4    Wheat4
#> 4     Fargo 1004   5    Wheat8
#> 5     Fargo 1005   3   Wheat12
#> 6     Fargo 1006   1   Wheat10
#> 7     Fargo 1007   6    Wheat9
#> 8     Fargo 1008   2   Wheat11
#> 9     Fargo 1009   1    Wheat1
#> 10    Fargo 1010   1    Wheat4