Generates a Randomized Complete Block Design (RCBD)

It randomly generates a randomized complete block design (RCBD) across locations.

RCBD(
  t = NULL,
  reps = NULL,
  l = 1,
  plotNumber = 101,
  continuous = FALSE,
  planter = "serpentine",
  seed = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

t
an integer number with total number of treatments or a vector of dimension t with labels.

reps
Number of replicates (full blocks) of each treatment.

l
Number of locations. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

continuous
Logical value for plot number continuous or not. By default continuous = FALSE.

planter
Option for serpentine or cartesian arrangement. By default planter = 'serpentine'.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

locationNames
(optional) Names for each location.

data
(optional) Data frame with the labels of treatments.

Value

A list with information on the design parameters.

RCBD layout for each location.

plot number layout for each location.

Data frame with the RCBD field book.

References

Design and Analysis of Experiments, Volume 1, Introduction to Experimental Design. Second Edition. Klaus Hinkelmann & Oscar Kempthorne.John Wiley & Sons, Inc., Hoboken, New Jersey.

Examples

# Example 1: Generates a RCBD design with 3 blocks and 20 treatments across 3 locations.
rcbd1 <- RCBD(t = LETTERS[1:20], reps = 5, l = 3, 
              plotNumber = c(101,1001, 2001), 
              continuous = TRUE,
              planter = "serpentine", 
              seed = 1020, 
              locationNames = c("FARGO", "MINOT", "CASSELTON"))
rcbd1$infoDesign                  

#> $blocks
#> [1] 5
#> 
#> $number.of.treatments
#> [1] 20
#> 
#> $treatments
#>  [1] "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "N" "O" "P" "Q" "R" "S"
#> [20] "T"
#> 
#> $locations
#> [1] 3
#> 
#> $plotNumber
#> $plotNumber[[1]]
#> [1] 101 201 301 401 501
#> 
#> $plotNumber[[2]]
#> [1] 1001 1101 1201 1301 1401
#> 
#> $plotNumber[[3]]
#> [1] 2001 2101 2201 2301 2401
#> 
#> 
#> $locationNames
#> [1] "FARGO"     "MINOT"     "CASSELTON"
#> 
#> $seed
#> [1] 1020
#> 

rcbd1$RCBD.layout

#> [[1]]
#>      Block --Treatments--                           
#> [1,] "1"   "N A K L C O T M R Q H F I D S G P E B J"
#> [2,] "2"   "E F K B M H S J Q A L O G T N I D C R P"
#> [3,] "3"   "D G P C H J I T L F K A Q R M O N E B S"
#> [4,] "4"   "R K D J I M N E P O L T G B S F C H A Q"
#> [5,] "5"   "O R P J G D E T A B I F C N S Q M L K H"
#> 
#> [[2]]
#>      Block --Treatments--                           
#> [1,] "1"   "K M E A G H I S T F C N Q L R D P J O B"
#> [2,] "2"   "P S O D Q B L E K R T F N M J G I A H C"
#> [3,] "3"   "T Q K G P M N O E B J D H A R S L C I F"
#> [4,] "4"   "R M I L Q J D F A K C E B T O P N H G S"
#> [5,] "5"   "I S K Q F T O L B H P C E N R J D M A G"
#> 
#> [[3]]
#>      Block --Treatments--                           
#> [1,] "1"   "N K O J R H D B A M I E S T F P G Q C L"
#> [2,] "2"   "F T A G N D P E J Q H L S R I O B K M C"
#> [3,] "3"   "R M E A F H D T P L Q I B C J O S G K N"
#> [4,] "4"   "S Q M L R C T E F D N A G J P B O K I H"
#> [5,] "5"   "J F K P T R I G L S D B E N O Q C H M A"
#> 

rcbd1$plotNumber.layout

#> [[1]]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]  101  102  103  104  105  106  107  108  109   110   111   112   113   114
#> [2,]  140  139  138  137  136  135  134  133  132   131   130   129   128   127
#> [3,]  141  142  143  144  145  146  147  148  149   150   151   152   153   154
#> [4,]  180  179  178  177  176  175  174  173  172   171   170   169   168   167
#> [5,]  181  182  183  184  185  186  187  188  189   190   191   192   193   194
#>      [,15] [,16] [,17] [,18] [,19] [,20]
#> [1,]   115   116   117   118   119   120
#> [2,]   126   125   124   123   122   121
#> [3,]   155   156   157   158   159   160
#> [4,]   166   165   164   163   162   161
#> [5,]   195   196   197   198   199   200
#> 
#> [[2]]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,] 1101 1102 1103 1104 1105 1106 1107 1108 1109  1110  1111  1112  1113  1114
#> [2,] 1140 1139 1138 1137 1136 1135 1134 1133 1132  1131  1130  1129  1128  1127
#> [3,] 1141 1142 1143 1144 1145 1146 1147 1148 1149  1150  1151  1152  1153  1154
#> [4,] 1180 1179 1178 1177 1176 1175 1174 1173 1172  1171  1170  1169  1168  1167
#> [5,] 1181 1182 1183 1184 1185 1186 1187 1188 1189  1190  1191  1192  1193  1194
#>      [,15] [,16] [,17] [,18] [,19] [,20]
#> [1,]  1115  1116  1117  1118  1119  1120
#> [2,]  1126  1125  1124  1123  1122  1121
#> [3,]  1155  1156  1157  1158  1159  1160
#> [4,]  1166  1165  1164  1163  1162  1161
#> [5,]  1195  1196  1197  1198  1199  1200
#> 
#> [[3]]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,] 2201 2202 2203 2204 2205 2206 2207 2208 2209  2210  2211  2212  2213  2214
#> [2,] 2240 2239 2238 2237 2236 2235 2234 2233 2232  2231  2230  2229  2228  2227
#> [3,] 2241 2242 2243 2244 2245 2246 2247 2248 2249  2250  2251  2252  2253  2254
#> [4,] 2280 2279 2278 2277 2276 2275 2274 2273 2272  2271  2270  2269  2268  2267
#> [5,] 2281 2282 2283 2284 2285 2286 2287 2288 2289  2290  2291  2292  2293  2294
#>      [,15] [,16] [,17] [,18] [,19] [,20]
#> [1,]  2215  2216  2217  2218  2219  2220
#> [2,]  2226  2225  2224  2223  2222  2221
#> [3,]  2255  2256  2257  2258  2259  2260
#> [4,]  2266  2265  2264  2263  2262  2261
#> [5,]  2295  2296  2297  2298  2299  2300
#> 

head(rcbd1$fieldBook)

#>   LOCATION PLOT REP TREATMENT
#> 1    FARGO  101   1         N
#> 2    FARGO  102   1         A
#> 3    FARGO  103   1         K
#> 4    FARGO  104   1         L
#> 5    FARGO  105   1         C
#> 6    FARGO  106   1         O

# Example 2: Generates a RCBD design with 6 blocks and 18 treatments in one location.
rcbd2 <- RCBD(t = 18, reps = 6, l = 1, 
              plotNumber = 101, 
              continuous = FALSE, 
              planter = "serpentine", 
              seed = 13, 
              locationNames = "IBAGUE")
rcbd2$infoDesign                  

#> $blocks
#> [1] 6
#> 
#> $number.of.treatments
#> [1] 18
#> 
#> $treatments
#>  [1] "T1"  "T2"  "T3"  "T4"  "T5"  "T6"  "T7"  "T8"  "T9"  "T10" "T11" "T12"
#> [13] "T13" "T14" "T15" "T16" "T17" "T18"
#> 
#> $locations
#> [1] 1
#> 
#> $plotNumber
#> $plotNumber[[1]]
#> [1] 101 201 301 401 501 601
#> 
#> 
#> $locationNames
#> [1] "IBAGUE"
#> 
#> $seed
#> [1] 13
#> 

rcbd2$RCBD.layout

#> [[1]]
#>      Block --Treatments--                                                  
#> [1,] "1"   "T17 T13 T9 T16 T18 T10 T12 T7 T2 T8 T1 T14 T4 T6 T11 T15 T5 T3"
#> [2,] "2"   "T17 T18 T11 T6 T7 T12 T10 T4 T15 T8 T1 T13 T16 T5 T14 T3 T2 T9"
#> [3,] "3"   "T3 T10 T14 T6 T17 T9 T11 T5 T16 T8 T4 T18 T1 T13 T7 T15 T12 T2"
#> [4,] "4"   "T5 T12 T2 T10 T9 T8 T4 T17 T15 T7 T14 T11 T13 T16 T18 T3 T6 T1"
#> [5,] "5"   "T11 T17 T4 T13 T3 T1 T5 T16 T7 T10 T6 T2 T18 T8 T12 T9 T15 T14"
#> [6,] "6"   "T14 T9 T17 T3 T8 T5 T18 T6 T16 T13 T10 T7 T4 T15 T12 T1 T2 T11"
#> 

rcbd2$plotNumber.layout

#> [[1]]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]  101  102  103  104  105  106  107  108  109   110   111   112   113   114
#> [2,]  218  217  216  215  214  213  212  211  210   209   208   207   206   205
#> [3,]  301  302  303  304  305  306  307  308  309   310   311   312   313   314
#> [4,]  418  417  416  415  414  413  412  411  410   409   408   407   406   405
#> [5,]  501  502  503  504  505  506  507  508  509   510   511   512   513   514
#> [6,]  618  617  616  615  614  613  612  611  610   609   608   607   606   605
#>      [,15] [,16] [,17] [,18]
#> [1,]   115   116   117   118
#> [2,]   204   203   202   201
#> [3,]   315   316   317   318
#> [4,]   404   403   402   401
#> [5,]   515   516   517   518
#> [6,]   604   603   602   601
#> 

head(rcbd2$fieldBook)

#>   LOCATION PLOT REP TREATMENT
#> 1   IBAGUE  101   1       T17
#> 2   IBAGUE  102   1       T13
#> 3   IBAGUE  103   1        T9
#> 4   IBAGUE  104   1       T16
#> 5   IBAGUE  105   1       T18
#> 6   IBAGUE  106   1       T10