Spatial Un-replicated Diagonal Arrangement Design

Randomly generates an spatial un-replicated diagonal arrangement design.

Usage

diagonal_arrangement(
  nrows = NULL,
  ncols = NULL,
  lines = NULL,
  checks = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  kindExpt = "SUDC",
  splitBy = "row",
  seed = NULL,
  blocks = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows

Number of rows in the field.

ncols

Number of columns in the field.

lines

Number of genotypes, experimental lines or treatments.

checks

Number of genotypes checks.

planter

Option for serpentine or cartesian plot arrangement. By default planter = 'serpentine'.

l

Number of locations or sites. By default l = 1.

plotNumber

Numeric vector with the starting plot number for each location. By default plotNumber = 101.

kindExpt

Type of diagonal design, with single options: Single Un-replicated Diagonal Checks 'SUDC' and Decision Blocks Un-replicated Design with Diagonal Checks 'DBUDC' for multiple experiments. By default kindExpt = 'SUDC'.

splitBy

Option to split the field when kindExpt = 'DBUDC' is selected. By default splitBy = 'row'.

seed

(optional) Real number that specifies the starting seed to obtain reproducible designs.

blocks

Number of experiments or blocks to generate an DBUDC design. If kindExpt = 'DBUDC' and data is null, blocks are mandatory.

exptName

(optional) Name of the experiment.

locationNames

(optional) Names each location.

data

(optional) Data frame with 2 columns: ENTRY | NAME .

Value

A list with five elements.

infoDesign is a list with information on the design parameters.

layoutRandom is a matrix with the randomization layout.

plotsNumber is a matrix with the layout plot number.

data_entry is a data frame with the data input.

fieldBook is a data frame with field book design. This includes the index (Row, Column).

References

Clarke, G. P. Y., & Stefanova, K. T. (2011). Optimal design for early-generation plant breeding trials with unreplicated or partially replicated test lines. Australian & New Zealand Journal of Statistics, 53(4), 461–480.

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a spatial single diagonal arrangement design in one location
# with 270 treatments and 30 check plots for a field with dimensions 15 rows x 20 cols
# in a serpentine arrangement.
spatd <- diagonal_arrangement(
  nrows = 15, 
  ncols = 20, 
  lines = 270, 
  checks = 4, 
  plotNumber = 101, 
  kindExpt = "SUDC", 
  planter = "serpentine", 
  seed = 1987,
  exptName = "20WRY1", 
  locationNames = "MINOT"
)
spatd$infoDesign
#> $rows
#> [1] 15
#> 
#> $columns
#> [1] 20
#> 
#> $treatments
#> [1] 270
#> 
#> $checks
#> [1] 4
#> 
#> $entry_checks
#> [1] 1 2 3 4
#> 
#> $rep_checks
#> [1] 8 7 7 8
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $percent_checks
#> [1] "10%"
#> 
#> $fillers
#> [1] 0
#> 
#> $seed
#> [1] 1987
#> 
#> $id_design
#> [1] 15
#> 
spatd$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row15   98    4  217  213  139   59  247  208  272    24   232     1   222
#> Row14  221   34  243   69    2  145   37  177  233    67    84    78   239
#> Row13  111  237  254  249   33  142  137    4   39   156   101   159    93
#> Row12    4   27  135   23  191  167   71  244  263   176     4   148   227
#> Row11   89  203  269    2   94  205   43  199  228    95   206   224    10
#> Row10  113  151   64  179   79  184    3  165  220     5    49   183    26
#> Row9   115  138  274   51  264  103  105   82  152     2   158   129   119
#> Row8   147  157    3   40  120  190   66   96   13   163   207   127     4
#> Row7   259  255   92   16    9    1   45  180  154    76   189   144    42
#> Row6   245  114  195   15  173  223  117   60    2   171    87   262    62
#> Row5   108    3  230  186   81  141  219  268  132    30   197     1    97
#> Row4   172  118   61   58    3   80    8  218   14    47   194    54   240
#> Row3    12  116  212   75  209   90  193    3  126   131   169   168   133
#> Row2     3  198   28  210  236   25   53  178  130    17     4   238   146
#> Row1   122  231   77    2  175   19   86  136   57   192   258   106   229
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row15   246   261   100   266   160   162    52
#> Row14    56     1     7   270   250    22   235
#> Row13   174   204    20   134     4   155   112
#> Row12   150    36   164    63    70   153   256
#> Row11     1    29   234    65   248    99     6
#> Row10   196    55   102     1   170    83   187
#> Row9     68   265    72   200    21   201     4
#> Row8    109    73   128   241   253   202    11
#> Row7    140   226     2   225   271   166   161
#> Row6    215   273    35   242   124     3    74
#> Row5    188    18   121   260    41   216    91
#> Row4    143     2    48   252    85    44   211
#> Row3    182   185    88    50     1    32   104
#> Row2    257    31   149   125   110   181    46
#> Row1      1   123    38   107   214   251   267
#> 
spatd$plotsNumber
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row15  381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row14  380  379  378  377  376  375  374  373  372   371   370   369   368
#> Row13  341  342  343  344  345  346  347  348  349   350   351   352   353
#> Row12  340  339  338  337  336  335  334  333  332   331   330   329   328
#> Row11  301  302  303  304  305  306  307  308  309   310   311   312   313
#> Row10  300  299  298  297  296  295  294  293  292   291   290   289   288
#> Row9   261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row8   260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row7   221  222  223  224  225  226  227  228  229   230   231   232   233
#> Row6   220  219  218  217  216  215  214  213  212   211   210   209   208
#> Row5   181  182  183  184  185  186  187  188  189   190   191   192   193
#> Row4   180  179  178  177  176  175  174  173  172   171   170   169   168
#> Row3   141  142  143  144  145  146  147  148  149   150   151   152   153
#> Row2   140  139  138  137  136  135  134  133  132   131   130   129   128
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row15   394   395   396   397   398   399   400
#> Row14   367   366   365   364   363   362   361
#> Row13   354   355   356   357   358   359   360
#> Row12   327   326   325   324   323   322   321
#> Row11   314   315   316   317   318   319   320
#> Row10   287   286   285   284   283   282   281
#> Row9    274   275   276   277   278   279   280
#> Row8    247   246   245   244   243   242   241
#> Row7    234   235   236   237   238   239   240
#> Row6    207   206   205   204   203   202   201
#> Row5    194   195   196   197   198   199   200
#> Row4    167   166   165   164   163   162   161
#> Row3    154   155   156   157   158   159   160
#> Row2    127   126   125   124   123   122   121
#> Row1    114   115   116   117   118   119   120
#> 
head(spatd$fieldBook, 12)
#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20WRY1    MINOT 2022  101   1      1      0   122   Gen-122
#> 2   2 20WRY1    MINOT 2022  102   1      2      0   231   Gen-231
#> 3   3 20WRY1    MINOT 2022  103   1      3      0    77    Gen-77
#> 4   4 20WRY1    MINOT 2022  104   1      4      2     2   Check-2
#> 5   5 20WRY1    MINOT 2022  105   1      5      0   175   Gen-175
#> 6   6 20WRY1    MINOT 2022  106   1      6      0    19    Gen-19
#> 7   7 20WRY1    MINOT 2022  107   1      7      0    86    Gen-86
#> 8   8 20WRY1    MINOT 2022  108   1      8      0   136   Gen-136
#> 9   9 20WRY1    MINOT 2022  109   1      9      0    57    Gen-57
#> 10 10 20WRY1    MINOT 2022  110   1     10      0   192   Gen-192
#> 11 11 20WRY1    MINOT 2022  111   1     11      0   258   Gen-258
#> 12 12 20WRY1    MINOT 2022  112   1     12      0   106   Gen-106

# Example 2: Generates a spatial decision block diagonal arrangement design in one location
# with 720 treatments allocated in 5 experiments or blocks for a field with dimensions
# 30 rows x 26 cols in a serpentine arrangement. In this case, we show how to set up the data 
# option with the entries list.
checks <- 5;expts <- 5
list_checks <- paste("CH", 1:checks, sep = "")
treatments <- paste("G", 6:725, sep = "")
treatment_list <- data.frame(list(ENTRY = 1:725, NAME = c(list_checks, treatments)))
head(treatment_list, 12) 
#>    ENTRY NAME
#> 1      1  CH1
#> 2      2  CH2
#> 3      3  CH3
#> 4      4  CH4
#> 5      5  CH5
#> 6      6   G6
#> 7      7   G7
#> 8      8   G8
#> 9      9   G9
#> 10    10  G10
#> 11    11  G11
#> 12    12  G12
tail(treatment_list, 12)
#>     ENTRY NAME
#> 714   714 G714
#> 715   715 G715
#> 716   716 G716
#> 717   717 G717
#> 718   718 G718
#> 719   719 G719
#> 720   720 G720
#> 721   721 G721
#> 722   722 G722
#> 723   723 G723
#> 724   724 G724
#> 725   725 G725
spatDB <- diagonal_arrangement(
  nrows = 30, 
  ncols = 26,
  checks = 5, 
  plotNumber = 1, 
  kindExpt = "DBUDC", 
  planter = "serpentine", 
  splitBy = "row", 
  blocks = c(150,155,95,200,120),
  data = treatment_list
 )
spatDB$infoDesign
#> $rows
#> [1] 30
#> 
#> $columns
#> [1] 26
#> 
#> $treatments
#> [1] 150 155  95 200 120
#> 
#> $checks
#> [1] 5
#> 
#> $entry_checks
#> [1] 1 2 3 4 5
#> 
#> $rep_checks
#> [1] 14 11 12 11 12
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $percent_checks
#> [1] "7.7%"
#> 
#> $fillers
#> [1] 0
#> 
#> $seed
#> [1] 1375.622
#> 
#> $id_design
#> [1] 15
#> 
spatDB$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  715    3  618  631  640  725  648  686  630   646   669   684   721
#> Row29  716  639  689  614  701    4  718  678  708   679   680   683   719
#> Row28  685  653  703  647  667  641  616  705  610     2   696   671   663
#> Row27    4  724  710  664  628  633  660  698  608   711   692   677   668
#> Row26  697  612  700  687    3  613  617  606  638   607   643   722   673
#> Row25  515  498  457  465  464  538  448  517    5   441   599   590   588
#> Row24  535  431  502  524  439  410  437  576  434   503   422   522     5
#> Row23  573  543  500    4  436  501  548  580  564   528   596   435   430
#> Row22  514  521  512  427  489  442  417    2  584   569   583   604   553
#> Row21  533  445  541  456  484  505  504  449  571   472   486     3   586
#> Row20  493  510    1  443  561  453  480  438  475   494   440   550   469
#> Row19  559  481  462  560  572  508    3  591  542   466   529   419   574
#> Row18  495  429  525  414  578  552  602  479  530   406     5   450   433
#> Row17  385    5  384  374  355  340  313  332  390   319   360   331   403
#> Row16  377  386  339  404  326    3  363  368  337   346   336   392   394
#> Row15  400  317  372  361  356  343  314  328  345     1   405   347   350
#> Row14    3  344  367  358  379  348  402  401  371   342   362   396   365
#> Row13  158  298  257  303    3  203  181  285  256   287   267   163   274
#> Row12  280  224  173  270  249  178  193  198    5   217   169   238   231
#> Row11  195  248  261  259  186  188  187  293  273   202   172   250     4
#> Row10  222  184  204    2  234  242  251  239  206   197   208   302   183
#> Row9   262  214  246  171  215  290  295    2  218   212   241   232   294
#> Row8   291  266  269  189  156  166  296  306  297   268   196     3   223
#> Row7    13   96    3  113   10  128  134  164  161   191   307   276   174
#> Row6    24   62  114  121   87  133    3   91  146    89   125   106    97
#> Row5    29   55   42  154   54  122  131  142   72    59     1    84    39
#> Row4   155    2   82   56  153   80   83   53    7    86   139   150    48
#> Row3    15   16  149  115   58    4   23   41   28    75    68    32   130
#> Row2   105  109   38    8   64  132   78   93  138     5    25    17   151
#> Row1     2  120   66   14  147   11   76  129  100    63   110    37    19
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row30   720     5   609   652   714   707   636   688   626   713   699   651
#> Row29   681   645   706   637   672     1   627   690   665   658   676   629
#> Row28   656   634   622   632   654   611   674   661   623     2   659   620
#> Row27     5   666   702   682   642   655   625   704   657   624   670   709
#> Row26   675   615   662   723     1   619   644   649   694   621   650   712
#> Row25   492   432   474   409   597   519   491   507     4   458   416   562
#> Row24   483   463   444   411   558   567   408   447   555   568   565   461
#> Row23   603   490   532     1   605   534   575   424   451   460   518   536
#> Row22   531   546   545   421   423   585   470     4   446   600   488   420
#> Row21   549   598   454   513   413   476   592   551   577   595   540     2
#> Row20   557   473     2   547   554   482   415   563   594   467   485   587
#> Row19   477   506   509   516   579   581     1   487   471   426   527   566
#> Row18   523   511   593   478   496   412   582   601   407   428     1   499
#> Row17   327     2   388   338   393   468   455   526   418   539   589   537
#> Row16   325   312   364   391   375     1   359   311   357   369   329   320
#> Row15   315   354   324   351   397   395   322   335   366     4   318   398
#> Row14     5   349   381   323   382   378   383   353   333   387   389   376
#> Row13   192   160   240   289     4   225   330   316   399   380   373   321
#> Row12   243   236   253   205   170   165   282   185     1   168   176   283
#> Row11   235   167   210   284   233   229   281   247   157   300   258   310
#> Row10   230   279   221     2   304   159   309   301   305   255   175   200
#> Row9    278   308   237   244   199   194   245     1   292   190   265   286
#> Row8    207   260   211   299   252   162   228   288   219   275   182     5
#> Row7    213   263     4   227   220   254   209   226   271   272   179   177
#> Row6     74    81    70   140     6    45     4    65    22   143    94    98
#> Row5    104    21    49    50    90    51    67    34   112   148     1    12
#> Row4     61     5    99   141    85    79    44   118   127    33    31    73
#> Row3     92   119   144    26   102     5   123   126     9    88    30   111
#> Row2     71   135   101    69   124    52   103   136    47     3   107    40
#> Row1      1   137    43    77    27   116   152    95   145    46    18    36
#>       Col26
#> Row30   693
#> Row29   635
#> Row28   691
#> Row27   695
#> Row26   717
#> Row25   556
#> Row24     3
#> Row23   570
#> Row22   520
#> Row21   544
#> Row20   452
#> Row19   459
#> Row18   425
#> Row17   497
#> Row16   334
#> Row15   370
#> Row14   352
#> Row13   341
#> Row12   264
#> Row11     1
#> Row10   180
#> Row9    277
#> Row8    201
#> Row7    216
#> Row6     57
#> Row5     20
#> Row4    117
#> Row3     35
#> Row2     60
#> Row1    108
#> 
spatDB$plotsNumber
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  780  779  778  777  776  775  774  773  772   771   770   769   768
#> Row29  729  730  731  732  733  734  735  736  737   738   739   740   741
#> Row28  728  727  726  725  724  723  722  721  720   719   718   717   716
#> Row27  677  678  679  680  681  682  683  684  685   686   687   688   689
#> Row26  676  675  674  673  672  671  670  669  668   667   666   665   664
#> Row25  625  626  627  628  629  630  631  632  633   634   635   636   637
#> Row24  624  623  622  621  620  619  618  617  616   615   614   613   612
#> Row23  573  574  575  576  577  578  579  580  581   582   583   584   585
#> Row22  572  571  570  569  568  567  566  565  564   563   562   561   560
#> Row21  521  522  523  524  525  526  527  528  529   530   531   532   533
#> Row20  520  519  518  517  516  515  514  513  512   511   510   509   508
#> Row19  469  470  471  472  473  474  475  476  477   478   479   480   481
#> Row18  468  467  466  465  464  463  462  461  460   459   458   457   456
#> Row17  417  418  419  420  421  422  423  424  425   426   427   428   429
#> Row16  416  415  414  413  412  411  410  409  408   407   406   405   404
#> Row15  365  366  367  368  369  370  371  372  373   374   375   376   377
#> Row14  364  363  362  361  360  359  358  357  356   355   354   353   352
#> Row13  313  314  315  316  317  318  319  320  321   322   323   324   325
#> Row12  312  311  310  309  308  307  306  305  304   303   302   301   300
#> Row11  261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row10  260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row9   209  210  211  212  213  214  215  216  217   218   219   220   221
#> Row8   208  207  206  205  204  203  202  201  200   199   198   197   196
#> Row7   157  158  159  160  161  162  163  164  165   166   167   168   169
#> Row6   156  155  154  153  152  151  150  149  148   147   146   145   144
#> Row5   105  106  107  108  109  110  111  112  113   114   115   116   117
#> Row4   104  103  102  101  100   99   98   97   96    95    94    93    92
#> Row3    53   54   55   56   57   58   59   60   61    62    63    64    65
#> Row2    52   51   50   49   48   47   46   45   44    43    42    41    40
#> Row1     1    2    3    4    5    6    7    8    9    10    11    12    13
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row30   767   766   765   764   763   762   761   760   759   758   757   756
#> Row29   742   743   744   745   746   747   748   749   750   751   752   753
#> Row28   715   714   713   712   711   710   709   708   707   706   705   704
#> Row27   690   691   692   693   694   695   696   697   698   699   700   701
#> Row26   663   662   661   660   659   658   657   656   655   654   653   652
#> Row25   638   639   640   641   642   643   644   645   646   647   648   649
#> Row24   611   610   609   608   607   606   605   604   603   602   601   600
#> Row23   586   587   588   589   590   591   592   593   594   595   596   597
#> Row22   559   558   557   556   555   554   553   552   551   550   549   548
#> Row21   534   535   536   537   538   539   540   541   542   543   544   545
#> Row20   507   506   505   504   503   502   501   500   499   498   497   496
#> Row19   482   483   484   485   486   487   488   489   490   491   492   493
#> Row18   455   454   453   452   451   450   449   448   447   446   445   444
#> Row17   430   431   432   433   434   435   436   437   438   439   440   441
#> Row16   403   402   401   400   399   398   397   396   395   394   393   392
#> Row15   378   379   380   381   382   383   384   385   386   387   388   389
#> Row14   351   350   349   348   347   346   345   344   343   342   341   340
#> Row13   326   327   328   329   330   331   332   333   334   335   336   337
#> Row12   299   298   297   296   295   294   293   292   291   290   289   288
#> Row11   274   275   276   277   278   279   280   281   282   283   284   285
#> Row10   247   246   245   244   243   242   241   240   239   238   237   236
#> Row9    222   223   224   225   226   227   228   229   230   231   232   233
#> Row8    195   194   193   192   191   190   189   188   187   186   185   184
#> Row7    170   171   172   173   174   175   176   177   178   179   180   181
#> Row6    143   142   141   140   139   138   137   136   135   134   133   132
#> Row5    118   119   120   121   122   123   124   125   126   127   128   129
#> Row4     91    90    89    88    87    86    85    84    83    82    81    80
#> Row3     66    67    68    69    70    71    72    73    74    75    76    77
#> Row2     39    38    37    36    35    34    33    32    31    30    29    28
#> Row1     14    15    16    17    18    19    20    21    22    23    24    25
#>       Col26
#> Row30   755
#> Row29   754
#> Row28   703
#> Row27   702
#> Row26   651
#> Row25   650
#> Row24   599
#> Row23   598
#> Row22   547
#> Row21   546
#> Row20   495
#> Row19   494
#> Row18   443
#> Row17   442
#> Row16   391
#> Row15   390
#> Row14   339
#> Row13   338
#> Row12   287
#> Row11   286
#> Row10   235
#> Row9    234
#> Row8    183
#> Row7    182
#> Row6    131
#> Row5    130
#> Row4     79
#> Row3     78
#> Row2     27
#> Row1     26
#> 
head(spatDB$fieldBook,12)
#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 Block1        1 2022    1   1      1      2     2       CH2
#> 2   2 Block1        1 2022    2   1      2      0   120      G120
#> 3   3 Block1        1 2022    3   1      3      0    66       G66
#> 4   4 Block1        1 2022    4   1      4      0    14       G14
#> 5   5 Block1        1 2022    5   1      5      0   147      G147
#> 6   6 Block1        1 2022    6   1      6      0    11       G11
#> 7   7 Block1        1 2022    7   1      7      0    76       G76
#> 8   8 Block1        1 2022    8   1      8      0   129      G129
#> 9   9 Block1        1 2022    9   1      9      0   100      G100
#> 10 10 Block1        1 2022   10   1     10      0    63       G63
#> 11 11 Block1        1 2022   11   1     11      0   110      G110
#> 12 12 Block1        1 2022   12   1     12      0    37       G37

# Example 3: Generates a spatial decision block diagonal arrangement design in one location
# with 270 treatments allocated in 3 experiments or blocks for a field with dimensions
# 20 rows x 15 cols in a serpentine arrangement. Which in turn is an augmented block (3 blocks).
spatAB <- diagonal_arrangement(nrows = 20, ncols = 15, lines = 270, 
                               checks = 4, 
                               plotNumber = c(1,1001,2001), 
                               kindExpt = "DBUDC", 
                               planter = "serpentine",
                               exptName = c("20WRA", "20WRB", "20WRC"), 
                               blocks = c(90, 90, 90),
                               splitBy = "column")
spatAB$infoDesign
#> $rows
#> [1] 20
#> 
#> $columns
#> [1] 15
#> 
#> $treatments
#> [1] 90 90 90
#> 
#> $checks
#> [1] 4
#> 
#> $entry_checks
#> [1] 1 2 3 4
#> 
#> $rep_checks
#> [1] 6 7 9 8
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $percent_checks
#> [1] "10%"
#> 
#> $fillers
#> [1] 0
#> 
#> $seed
#> [1] -6568.56
#> 
#> $id_design
#> [1] 15
#> 
spatAB$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20   18    3   17   69   48  100  142  120  162   180   218     3   258
#> Row19   65   70   93   83    4  104  112  116  137   171   196   215   248
#> Row18   35   25    5   42   39  109  172    3  183   140   189   247   222
#> Row17    1   15   51   19   53   96  145  133  122   170     1   224   257
#> Row16   24   16    8    3   49  178  152  176   99   166   243   214   191
#> Row15   57   84   38   71   27  174    4  129  157   169   197   241   261
#> Row14   59   31   23   36   30  134  139  123  107     1   188   236   192
#> Row13   76   10    3   14   75  153  147   98  175   136   190   193     4
#> Row12   86   64    6   41   54    3  182  177  161   126   185   201   212
#> Row11   21   33   92   26   46  151  167  164    2   146   200   204   273
#> Row10   72    4   81   63    7  117  106   95  159   132   206     4   259
#> Row9    47   62   50   44    1  149  127  173  103   156   262   253   220
#> Row8    74   78   77   66   56  163  105    1  154   135   195   274   265
#> Row7     2   40   28    9   12  130  114  101  155   124     3   202   211
#> Row6    90   79   82    4   94  110  184  115  118   119   223   255   233
#> Row5    52   85   29   89   34  131    4  113  179   168   221   226   213
#> Row4    61   37   45   88   73  165  158  143  121     2   271   227   205
#> Row3    20   55    2   22   91  108  160  128  141   144   210   242     2
#> Row2    68   80   58   60   87    3  138  148  181   111   207   217   239
#> Row1    43   32   67   13   11  125  150   97    2   102   235   268   267
#>       Col14 Col15
#> Row20   229   216
#> Row19   272     1
#> Row18   270   264
#> Row17   219   228
#> Row16     2   238
#> Row15   232   187
#> Row14   269   252
#> Row13   194   249
#> Row12   225   231
#> Row11   199   237
#> Row10   260   208
#> Row9    240     3
#> Row8    246   244
#> Row7    198   251
#> Row6      4   263
#> Row5    254   250
#> Row4    230   245
#> Row3    256   234
#> Row2    209   203
#> Row1    186   266
#> 
spatAB$plotsNumber
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  100   99   98   97   96 1100 1099 1098 1097  1096  2100  2099  2098
#> Row19   91   92   93   94   95 1091 1092 1093 1094  1095  2091  2092  2093
#> Row18   90   89   88   87   86 1090 1089 1088 1087  1086  2090  2089  2088
#> Row17   81   82   83   84   85 1081 1082 1083 1084  1085  2081  2082  2083
#> Row16   80   79   78   77   76 1080 1079 1078 1077  1076  2080  2079  2078
#> Row15   71   72   73   74   75 1071 1072 1073 1074  1075  2071  2072  2073
#> Row14   70   69   68   67   66 1070 1069 1068 1067  1066  2070  2069  2068
#> Row13   61   62   63   64   65 1061 1062 1063 1064  1065  2061  2062  2063
#> Row12   60   59   58   57   56 1060 1059 1058 1057  1056  2060  2059  2058
#> Row11   51   52   53   54   55 1051 1052 1053 1054  1055  2051  2052  2053
#> Row10   50   49   48   47   46 1050 1049 1048 1047  1046  2050  2049  2048
#> Row9    41   42   43   44   45 1041 1042 1043 1044  1045  2041  2042  2043
#> Row8    40   39   38   37   36 1040 1039 1038 1037  1036  2040  2039  2038
#> Row7    31   32   33   34   35 1031 1032 1033 1034  1035  2031  2032  2033
#> Row6    30   29   28   27   26 1030 1029 1028 1027  1026  2030  2029  2028
#> Row5    21   22   23   24   25 1021 1022 1023 1024  1025  2021  2022  2023
#> Row4    20   19   18   17   16 1020 1019 1018 1017  1016  2020  2019  2018
#> Row3    11   12   13   14   15 1011 1012 1013 1014  1015  2011  2012  2013
#> Row2    10    9    8    7    6 1010 1009 1008 1007  1006  2010  2009  2008
#> Row1     1    2    3    4    5 1001 1002 1003 1004  1005  2001  2002  2003
#>       Col14 Col15
#> Row20  2097  2096
#> Row19  2094  2095
#> Row18  2087  2086
#> Row17  2084  2085
#> Row16  2077  2076
#> Row15  2074  2075
#> Row14  2067  2066
#> Row13  2064  2065
#> Row12  2057  2056
#> Row11  2054  2055
#> Row10  2047  2046
#> Row9   2044  2045
#> Row8   2037  2036
#> Row7   2034  2035
#> Row6   2027  2026
#> Row5   2024  2025
#> Row4   2017  2016
#> Row3   2014  2015
#> Row2   2007  2006
#> Row1   2004  2005
#> 
head(spatAB$fieldBook,12)
#>    ID  EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20WRA        1 2022    1   1      1      0    43    Gen-43
#> 2   2 20WRA        1 2022    2   1      2      0    32    Gen-32
#> 3   3 20WRA        1 2022    3   1      3      0    67    Gen-67
#> 4   4 20WRA        1 2022    4   1      4      0    13    Gen-13
#> 5   5 20WRA        1 2022    5   1      5      0    11    Gen-11
#> 6   6 20WRB        1 2022 1001   1      6      0   125   Gen-125
#> 7   7 20WRB        1 2022 1002   1      7      0   150   Gen-150
#> 8   8 20WRB        1 2022 1003   1      8      0    97    Gen-97
#> 9   9 20WRB        1 2022 1004   1      9      2     2   Check-2
#> 10 10 20WRB        1 2022 1005   1     10      0   102   Gen-102
#> 11 11 20WRC        1 2022 2001   1     11      0   235   Gen-235
#> 12 12 20WRC        1 2022 2002   1     12      0   268   Gen-268