Spatial Un-replicated Diagonal Arrangement Design

Randomly generates an spatial un-replicated diagonal arrangement design.

diagonal_arrangement(
  nrows = NULL,
  ncols = NULL,
  lines = NULL,
  checks = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  kindExpt = "SUDC",
  splitBy = "row",
  seed = NULL,
  blocks = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows
Number of rows in the field.

ncols
Number of columns in the field.

lines
Number of genotypes, experimental lines or treatments.

checks
Integer number of genotypes checks or a numeric vector.

planter
Option for serpentine or cartesian plot arrangement. By default planter = 'serpentine'.

l
Number of locations or sites. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

kindExpt
Type of diagonal design, with single options: Single Un-replicated Diagonal Checks 'SUDC' and Decision Blocks Un-replicated Design with Diagonal Checks 'DBUDC' for multiple experiments. By default kindExpt = 'SUDC'.

splitBy
Option to split the field when kindExpt = 'DBUDC' is selected. By default splitBy = 'row'.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

blocks
Number of experiments or blocks to generate an DBUDC design. If kindExpt = 'DBUDC' and data is null, blocks are mandatory.

exptName
(optional) Name of the experiment.

locationNames
(optional) Names each location.

data
(optional) Data frame with 2 columns: ENTRY | NAME | BLOCK or only 2 columns ENTRY | NAME if kindExpt = 'SUDC'.

Value

A list with information on the design parameters.

A matrix with the randomization layout.

A matrix with the layout plot number.

A data frame with field book design. This includes the index (Row, Column).

References

Clarke, G. P. Y., & Stefanova, K. T. (2011). Optimal design for early-generation plant breeding trials with unreplicated or partially replicated test lines. Australian & New Zealand Journal of Statistics, 53(4), 461–480.

Examples

# Example 1: Generates a spatial single diagonal arrangement design in one location
# with 270 treatments and 30 check plots for a field with dimensions 15 rows x 20 cols
# in a serpentine arrangement.
spatd <- diagonal_arrangement(nrows = 15, ncols = 20, lines = 270, 
                              checks = 4, 
                              plotNumber = 101, 
                              kindExpt = "SUDC", 
                              planter = "serpentine", 
                              seed = 1987,
                              exptName = "20WRY1", 
                              locationNames = "MINOT")
spatd$infoDesign

#> $Lines
#> [1] 270
#> 
#> $checks
#> [1] 4
#> 
#> $RepChecks
#> 1 2 3 4 
#> 8 8 7 7 
#> 
#> $percentChecks
#> [1] "10%"
#> 
#> $Fillers
#> [1] 0
#> 
#> $seed
#> [1] 1987
#> 

spatd$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row15   59    4    9  271  139   92   87   37  136    74   113     2   270
#> Row14  146  267  248   12    3   88   82   47  210    66   153   122   211
#> Row13   42   75  173  178  253  107  181    4  150   131   105   124   249
#> Row12    2   19   96  201  117   29  121  215  242    34     1   265    24
#> Row11   35   91  164    2  126   30  161  224   11   255   194   246    17
#> Row10  205   73  158   84  127  244    3   62  137   157   106   195    80
#> Row9   167  133  251  145  168  204  116  119  247     4    39   123   256
#> Row8   238   57    1  162  193  190  142  101   26   239   222   227     1
#> Row7   100    8  261  208  172    2  202  184  120   219   218   196   177
#> Row6   129  134  138  245  221   79  191   46    2    72   169   152    61
#> Row5    76    3  216  258  180   67   55  250   18   236    52     4   140
#> Row4    50  132  254  272    3  109   68   53   70   264   103   212    28
#> Row3   187  148  135   43   41   49   21    2  217   104   213   237   186
#> Row2     1   36  154   38  234  233  160  252  240    98     4   108   197
#> Row1   185   31   94    1  274  200  163  207  232   257   203   118    25
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row15   189   183    33    78    54    20   214
#> Row14    99     2    85   102    71   175     5
#> Row13    83   260   130    44     4    32    65
#> Row12    48   206   259   235    90   263     6
#> Row11     3    27   165   174    81   223    69
#> Row10   166   182   110     1   111   155    77
#> Row9     14   229   188   230   241    45     4
#> Row8    170   225   268    23   209   143   144
#> Row7    171    51     2   141   112   243   192
#> Row6     95    10    60   151   220     3   262
#> Row5     89    93   156   179   115   176   199
#> Row4    266     1   149    63   128   226    16
#> Row3     13    15   125   228     1   198   114
#> Row2     64    58    40     7    22    97   147
#> Row1      3   159   273   269    56    86   231

spatd$plotsNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row15  381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row14  380  379  378  377  376  375  374  373  372   371   370   369   368
#> Row13  341  342  343  344  345  346  347  348  349   350   351   352   353
#> Row12  340  339  338  337  336  335  334  333  332   331   330   329   328
#> Row11  301  302  303  304  305  306  307  308  309   310   311   312   313
#> Row10  300  299  298  297  296  295  294  293  292   291   290   289   288
#> Row9   261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row8   260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row7   221  222  223  224  225  226  227  228  229   230   231   232   233
#> Row6   220  219  218  217  216  215  214  213  212   211   210   209   208
#> Row5   181  182  183  184  185  186  187  188  189   190   191   192   193
#> Row4   180  179  178  177  176  175  174  173  172   171   170   169   168
#> Row3   141  142  143  144  145  146  147  148  149   150   151   152   153
#> Row2   140  139  138  137  136  135  134  133  132   131   130   129   128
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row15   394   395   396   397   398   399   400
#> Row14   367   366   365   364   363   362   361
#> Row13   354   355   356   357   358   359   360
#> Row12   327   326   325   324   323   322   321
#> Row11   314   315   316   317   318   319   320
#> Row10   287   286   285   284   283   282   281
#> Row9    274   275   276   277   278   279   280
#> Row8    247   246   245   244   243   242   241
#> Row7    234   235   236   237   238   239   240
#> Row6    207   206   205   204   203   202   201
#> Row5    194   195   196   197   198   199   200
#> Row4    167   166   165   164   163   162   161
#> Row3    154   155   156   157   158   159   160
#> Row2    127   126   125   124   123   122   121
#> Row1    114   115   116   117   118   119   120

head(spatd$fieldBook, 12)

#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20WRY1    MINOT 2021  101   1      1      0   185   gen 185
#> 2   2 20WRY1    MINOT 2021  102   1      2      0    31    gen 31
#> 3   3 20WRY1    MINOT 2021  103   1      3      0    94    gen 94
#> 4   4 20WRY1    MINOT 2021  104   1      4      1     1   Check 1
#> 5   5 20WRY1    MINOT 2021  105   1      5      0   274   gen 274
#> 6   6 20WRY1    MINOT 2021  106   1      6      0   200   gen 200
#> 7   7 20WRY1    MINOT 2021  107   1      7      0   163   gen 163
#> 8   8 20WRY1    MINOT 2021  108   1      8      0   207   gen 207
#> 9   9 20WRY1    MINOT 2021  109   1      9      0   232   gen 232
#> 10 10 20WRY1    MINOT 2021  110   1     10      0   257   gen 257
#> 11 11 20WRY1    MINOT 2021  111   1     11      0   203   gen 203
#> 12 12 20WRY1    MINOT 2021  112   1     12      0   118   gen 118

# Example 2: Generates a spatial decision block diagonal arrangement design in one location
# with 720 treatments allocated in 5 experiments or blocks for a field with dimensions
# 30 rows x 26 cols in a serpentine arrangement. In this case, we show how to set up the data 
# option with the entries list.
checks <- 5;expts <- 5
list_checks <- paste("CH", 1:checks, sep = "")
treatments <- paste("G", 6:725, sep = "")
BLOCK <- c(rep("ALL", checks), rep(1:expts, c(150,155,95,200,120)))
treatment_list <- data.frame(list(ENTRY = 1:725, NAME = c(list_checks, treatments), BLOCK = BLOCK))
head(treatment_list, 12) 

#>    ENTRY NAME BLOCK
#> 1      1  CH1   ALL
#> 2      2  CH2   ALL
#> 3      3  CH3   ALL
#> 4      4  CH4   ALL
#> 5      5  CH5   ALL
#> 6      6   G6     1
#> 7      7   G7     1
#> 8      8   G8     1
#> 9      9   G9     1
#> 10    10  G10     1
#> 11    11  G11     1
#> 12    12  G12     1

tail(treatment_list, 12)

#>     ENTRY NAME BLOCK
#> 714   714 G714     5
#> 715   715 G715     5
#> 716   716 G716     5
#> 717   717 G717     5
#> 718   718 G718     5
#> 719   719 G719     5
#> 720   720 G720     5
#> 721   721 G721     5
#> 722   722 G722     5
#> 723   723 G723     5
#> 724   724 G724     5
#> 725   725 G725     5

spatDB <- diagonal_arrangement(nrows = 30, ncols = 26,
                               checks = 5, 
                               plotNumber = 1, 
                               kindExpt = "DBUDC", 
                               planter = "serpentine", 
                               splitBy = "row", 
                               data = treatment_list)

#> [1] 725
#> [1] 725
#> [1] 725
#> [1] 725
#> [1] 725
#> [1] 725
#> [1] 725
#> [1] 725
#> [1] 725
#> [1] 725
#> [1] 725

spatDB$infoDesign

#> $Lines
#> [1] 150 155  95 200 120
#> 
#> $checks
#> [1] 5
#> 
#> $RepChecks
#>  1  2  3  4  5 
#> 12 11 12 13 12 
#> 
#> $percentChecks
#> [1] "7.7%"
#> 
#> $Fillers
#> [1] 0
#> 
#> $seed
#> [1] 3859.465
#> 

spatDB$layoutRandom

#>       Col1  Col2  Col3  Col4  Col5  Col6  Col7  Col8  Col9  Col10 Col11 Col12
#> Row30 "675" "3"   "638" "639" "635" "664" "665" "650" "687" "619" "656" "670"
#> Row29 "720" "718" "714" "723" "711" "2"   "682" "717" "652" "695" "719" "662"
#> Row28 "610" "611" "622" "625" "697" "628" "673" "707" "706" "4"   "627" "712"
#> Row27 "1"   "641" "661" "689" "632" "646" "686" "614" "616" "631" "626" "637"
#> Row26 "647" "724" "615" "693" "3"   "624" "722" "667" "716" "663" "698" "633"
#> Row25 "476" "586" "559" "416" "466" "482" "411" "515" "5"   "524" "556" "489"
#> Row24 "461" "526" "530" "475" "565" "542" "469" "440" "527" "425" "597" "481"
#> Row23 "427" "457" "444" "4"   "589" "582" "436" "451" "561" "415" "460" "496"
#> Row22 "529" "569" "454" "501" "465" "510" "494" "5"   "554" "599" "468" "412"
#> Row21 "453" "528" "513" "478" "540" "495" "502" "413" "576" "595" "601" "1"  
#> Row20 "517" "560" "4"   "578" "549" "484" "488" "426" "442" "545" "491" "587"
#> Row19 "574" "508" "503" "446" "479" "512" "1"   "583" "470" "593" "439" "441"
#> Row18 "420" "585" "550" "509" "538" "449" "452" "544" "584" "474" "2"   "410"
#> Row17 "404" "2"   "394" "349" "361" "358" "377" "355" "323" "322" "321" "376"
#> Row16 "400" "375" "366" "338" "347" "1"   "335" "388" "343" "385" "331" "401"
#> Row15 "379" "403" "330" "360" "392" "374" "312" "398" "393" "4"   "317" "353"
#> Row14 "3"   "363" "352" "328" "372" "378" "402" "370" "333" "329" "354" "324"
#> Row13 "296" "239" "291" "158" "3"   "273" "310" "181" "200" "178" "174" "263"
#> Row12 "238" "255" "188" "261" "262" "161" "300" "279" "3"   "198" "172" "210"
#> Row11 "159" "196" "187" "301" "269" "180" "270" "228" "265" "230" "253" "268"
#> Row10 "166" "251" "171" "5"   "215" "237" "288" "177" "302" "164" "189" "185"
#> Row9  "169" "157" "259" "236" "243" "280" "241" "5"   "247" "256" "309" "219"
#> Row8  "199" "223" "202" "283" "175" "208" "203" "193" "233" "293" "258" "5"  
#> Row7  "93"  "148" "1"   "102" "22"  "33"  "135" "162" "195" "290" "183" "281"
#> Row6  "98"  "56"  "62"  "72"  "23"  "48"  "2"   "15"  "12"  "52"  "109" "107"
#> Row5  "117" "87"  "10"  "18"  "137" "74"  "75"  "81"  "63"  "73"  "4"   "66" 
#> Row4  "65"  "5"   "95"  "51"  "91"  "105" "111" "122" "86"  "54"  "143" "59" 
#> Row3  "40"  "132" "90"  "145" "57"  "1"   "94"  "139" "100" "13"  "29"  "38" 
#> Row2  "154" "41"  "53"  "134" "34"  "125" "83"  "106" "115" "1"   "127" "24" 
#> Row1  "2"   "129" "61"  "124" "80"  "146" "17"  "25"  "121" "152" "131" "119"
#>       Col13 Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24
#> Row30 "679" "713" "3"   "634" "645" "643" "710" "725" "709" "606" "691" "715"
#> Row29 "692" "642" "680" "629" "690" "659" "1"   "681" "658" "674" "701" "676"
#> Row28 "683" "696" "657" "703" "655" "640" "684" "694" "651" "612" "4"   "649"
#> Row27 "671" "2"   "608" "613" "721" "669" "699" "704" "644" "688" "668" "672"
#> Row26 "677" "617" "702" "678" "700" "2"   "654" "630" "648" "607" "705" "666"
#> Row25 "431" "417" "536" "591" "471" "467" "443" "573" "423" "5"   "532" "490"
#> Row24 "4"   "534" "604" "520" "462" "553" "428" "499" "551" "602" "571" "507"
#> Row23 "568" "533" "477" "594" "5"   "605" "577" "596" "543" "505" "418" "514"
#> Row22 "473" "487" "421" "435" "575" "567" "408" "486" "3"   "438" "497" "555"
#> Row21 "437" "430" "447" "547" "511" "414" "598" "564" "600" "448" "492" "455"
#> Row20 "493" "463" "557" "3"   "472" "546" "464" "432" "603" "570" "406" "590"
#> Row19 "535" "458" "459" "434" "563" "522" "537" "2"   "548" "409" "592" "506"
#> Row18 "552" "519" "424" "588" "429" "422" "498" "450" "581" "541" "504" "1"  
#> Row17 "381" "362" "5"   "326" "345" "339" "580" "566" "480" "539" "500" "562"
#> Row16 "391" "344" "318" "369" "395" "311" "1"   "356" "341" "315" "340" "396"
#> Row15 "316" "332" "334" "325" "405" "350" "387" "359" "380" "397" "5"   "342"
#> Row14 "390" "4"   "373" "384" "365" "386" "346" "313" "367" "351" "336" "357"
#> Row13 "272" "276" "266" "282" "234" "1"   "220" "371" "399" "383" "314" "389"
#> Row12 "201" "275" "160" "222" "205" "242" "307" "218" "226" "2"   "245" "305"
#> Row11 "4"   "212" "284" "197" "254" "182" "248" "246" "191" "252" "240" "186"
#> Row10 "294" "214" "295" "299" "4"   "298" "184" "286" "204" "213" "209" "227"
#> Row9  "289" "176" "156" "250" "170" "231" "274" "297" "4"   "163" "264" "229"
#> Row8  "224" "287" "221" "179" "303" "271" "285" "235" "217" "278" "211" "216"
#> Row7  "308" "207" "190" "1"   "244" "173" "232" "306" "225" "304" "167" "192"
#> Row6  "47"  "82"  "36"  "110" "50"  "138" "133" "2"   "9"   "35"  "6"   "60" 
#> Row5  "101" "155" "49"  "45"  "67"  "153" "11"  "78"  "8"   "108" "140" "5"  
#> Row4  "77"  "32"  "3"   "16"  "92"  "70"  "123" "44"  "136" "99"  "141" "79" 
#> Row3  "89"  "112" "69"  "120" "19"  "88"  "2"   "76"  "97"  "28"  "55"  "42" 
#> Row2  "96"  "147" "116" "20"  "130" "104" "113" "21"  "103" "68"  "3"   "126"
#> Row1  "142" "4"   "114" "85"  "30"  "27"  "58"  "71"  "31"  "64"  "14"  "26" 
#>       Col25 Col26
#> Row30 "620" "609"
#> Row29 "618" "708"
#> Row28 "660" "685"
#> Row27 "636" "653"
#> Row26 "621" "623"
#> Row25 "558" "433"
#> Row24 "518" "5"  
#> Row23 "579" "445"
#> Row22 "516" "407"
#> Row21 "3"   "523"
#> Row20 "525" "485"
#> Row19 "456" "483"
#> Row18 "521" "572"
#> Row17 "419" "531"
#> Row16 "320" "327"
#> Row15 "319" "382"
#> Row14 "364" "368"
#> Row13 "337" "348"
#> Row12 "277" "206"
#> Row11 "249" "3"  
#> Row10 "267" "292"
#> Row9  "260" "194"
#> Row8  "4"   "257"
#> Row7  "168" "165"
#> Row6  "151" "150"
#> Row5  "149" "46" 
#> Row4  "37"  "144"
#> Row3  "39"  "43" 
#> Row2  "7"   "84" 
#> Row1  "118" "128"

spatDB$plotsNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  780  779  778  777  776  775  774  773  772   771   770   769   768
#> Row29  729  730  731  732  733  734  735  736  737   738   739   740   741
#> Row28  728  727  726  725  724  723  722  721  720   719   718   717   716
#> Row27  677  678  679  680  681  682  683  684  685   686   687   688   689
#> Row26  676  675  674  673  672  671  670  669  668   667   666   665   664
#> Row25  625  626  627  628  629  630  631  632  633   634   635   636   637
#> Row24  624  623  622  621  620  619  618  617  616   615   614   613   612
#> Row23  573  574  575  576  577  578  579  580  581   582   583   584   585
#> Row22  572  571  570  569  568  567  566  565  564   563   562   561   560
#> Row21  521  522  523  524  525  526  527  528  529   530   531   532   533
#> Row20  520  519  518  517  516  515  514  513  512   511   510   509   508
#> Row19  469  470  471  472  473  474  475  476  477   478   479   480   481
#> Row18  468  467  466  465  464  463  462  461  460   459   458   457   456
#> Row17  417  418  419  420  421  422  423  424  425   426   427   428   429
#> Row16  416  415  414  413  412  411  410  409  408   407   406   405   404
#> Row15  365  366  367  368  369  370  371  372  373   374   375   376   377
#> Row14  364  363  362  361  360  359  358  357  356   355   354   353   352
#> Row13  313  314  315  316  317  318  319  320  321   322   323   324   325
#> Row12  312  311  310  309  308  307  306  305  304   303   302   301   300
#> Row11  261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row10  260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row9   209  210  211  212  213  214  215  216  217   218   219   220   221
#> Row8   208  207  206  205  204  203  202  201  200   199   198   197   196
#> Row7   157  158  159  160  161  162  163  164  165   166   167   168   169
#> Row6   156  155  154  153  152  151  150  149  148   147   146   145   144
#> Row5   105  106  107  108  109  110  111  112  113   114   115   116   117
#> Row4   104  103  102  101  100   99   98   97   96    95    94    93    92
#> Row3    53   54   55   56   57   58   59   60   61    62    63    64    65
#> Row2    52   51   50   49   48   47   46   45   44    43    42    41    40
#> Row1     1    2    3    4    5    6    7    8    9    10    11    12    13
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row30   767   766   765   764   763   762   761   760   759   758   757   756
#> Row29   742   743   744   745   746   747   748   749   750   751   752   753
#> Row28   715   714   713   712   711   710   709   708   707   706   705   704
#> Row27   690   691   692   693   694   695   696   697   698   699   700   701
#> Row26   663   662   661   660   659   658   657   656   655   654   653   652
#> Row25   638   639   640   641   642   643   644   645   646   647   648   649
#> Row24   611   610   609   608   607   606   605   604   603   602   601   600
#> Row23   586   587   588   589   590   591   592   593   594   595   596   597
#> Row22   559   558   557   556   555   554   553   552   551   550   549   548
#> Row21   534   535   536   537   538   539   540   541   542   543   544   545
#> Row20   507   506   505   504   503   502   501   500   499   498   497   496
#> Row19   482   483   484   485   486   487   488   489   490   491   492   493
#> Row18   455   454   453   452   451   450   449   448   447   446   445   444
#> Row17   430   431   432   433   434   435   436   437   438   439   440   441
#> Row16   403   402   401   400   399   398   397   396   395   394   393   392
#> Row15   378   379   380   381   382   383   384   385   386   387   388   389
#> Row14   351   350   349   348   347   346   345   344   343   342   341   340
#> Row13   326   327   328   329   330   331   332   333   334   335   336   337
#> Row12   299   298   297   296   295   294   293   292   291   290   289   288
#> Row11   274   275   276   277   278   279   280   281   282   283   284   285
#> Row10   247   246   245   244   243   242   241   240   239   238   237   236
#> Row9    222   223   224   225   226   227   228   229   230   231   232   233
#> Row8    195   194   193   192   191   190   189   188   187   186   185   184
#> Row7    170   171   172   173   174   175   176   177   178   179   180   181
#> Row6    143   142   141   140   139   138   137   136   135   134   133   132
#> Row5    118   119   120   121   122   123   124   125   126   127   128   129
#> Row4     91    90    89    88    87    86    85    84    83    82    81    80
#> Row3     66    67    68    69    70    71    72    73    74    75    76    77
#> Row2     39    38    37    36    35    34    33    32    31    30    29    28
#> Row1     14    15    16    17    18    19    20    21    22    23    24    25
#>       Col26
#> Row30   755
#> Row29   754
#> Row28   703
#> Row27   702
#> Row26   651
#> Row25   650
#> Row24   599
#> Row23   598
#> Row22   547
#> Row21   546
#> Row20   495
#> Row19   494
#> Row18   443
#> Row17   442
#> Row16   391
#> Row15   390
#> Row14   339
#> Row13   338
#> Row12   287
#> Row11   286
#> Row10   235
#> Row9    234
#> Row8    183
#> Row7    182
#> Row6    131
#> Row5    130
#> Row4     79
#> Row3     78
#> Row2     27
#> Row1     26

head(spatDB$fieldBook,12)

#>    ID  EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 Expt1        1 2021    1   1      1      2     2       CH2
#> 2   2 Expt1        1 2021    2   1      2      0   129      G129
#> 3   3 Expt1        1 2021    3   1      3      0    61       G61
#> 4   4 Expt1        1 2021    4   1      4      0   124      G124
#> 5   5 Expt1        1 2021    5   1      5      0    80       G80
#> 6   6 Expt1        1 2021    6   1      6      0   146      G146
#> 7   7 Expt1        1 2021    7   1      7      0    17       G17
#> 8   8 Expt1        1 2021    8   1      8      0    25       G25
#> 9   9 Expt1        1 2021    9   1      9      0   121      G121
#> 10 10 Expt1        1 2021   10   1     10      0   152      G152
#> 11 11 Expt1        1 2021   11   1     11      0   131      G131
#> 12 12 Expt1        1 2021   12   1     12      0   119      G119

# Example 3: Generates a spatial decision block diagonal arrangement design in one location
# with 270 treatments allocated in 3 experiments or blocks for a field with dimensions
# 20 rows x 15 cols in a serpentine arrangement. Which in turn is an augmented block (3 blocks).
spatAB <- diagonal_arrangement(nrows = 20, ncols = 15, lines = 270, 
                               checks = 4, 
                               plotNumber = c(1,1001,2001), 
                               kindExpt = "DBUDC", 
                               planter = "serpentine",
                               exptName = c("20WRA", "20WRB", "20WRC"), 
                               blocks = c(90, 90, 90),
                               splitBy = "column")
spatAB$infoDesign

#> $Lines
#> [1] 90 90 90
#> 
#> $checks
#> [1] 4
#> 
#> $RepChecks
#> 1 2 3 4 
#> 9 8 7 6 
#> 
#> $percentChecks
#> [1] "10%"
#> 
#> $Fillers
#> [1] 0
#> 
#> $seed
#> [1] -9263.135
#> 

spatAB$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20   15    3   24    9   93  126  141  119  142   162   247     4   231
#> Row19   72   77   91   83    4  165  139  106  128   161   223   228   274
#> Row18   17   56   29   47   27  103   99    4  152   108   193   187   230
#> Row17    1   18   34   16   38  174  147  130  104   173     3   217   201
#> Row16   44   64   31    3   70  138  176  169  150   166   251   200   241
#> Row15   89   21   51   79    7  151    1  172  180   164   242   202   203
#> Row14   39   74   78   14   61  129  178  155  107     4   269   232   243
#> Row13   85   45    1   65   75  143  123  184  134   156   233   263     2
#> Row12   76   50   37   90   20    1  148  159  145   179   226   245   185
#> Row11   35   94   23   43   63  136   95  158    2   102   207   239   198
#> Row10   92    4   58   41   46  111  146  171  157   132   209     1   196
#> Row9    62   11   10   67    3  124  149   98  163   177   254   262   240
#> Row8    49   52   71    5   68  182  175    2  125   183   186   221   266
#> Row7     2   84   86   73   81  115  116  154  167   101     1   248   213
#> Row6     8   53   32    2   66  168  117  112  170   120   237   222   257
#> Row5    55   59   69   22   88   97    3  113  144   160   194   259   212
#> Row4    36   13   26    6   57  127  105  121  109     1   253   188   214
#> Row3    12   28    1   60   33  131  133  118  181   137   225   256     4
#> Row2    42   40   48   80   25    3  153  122  100   140   192   264   211
#> Row1    82   54   30   87   19   96  135  110    2   114   190   199   270
#>       Col14 Col15
#> Row20   258   234
#> Row19   205     3
#> Row18   204   249
#> Row17   220   219
#> Row16     1   260
#> Row15   235   189
#> Row14   268   236
#> Row13   224   255
#> Row12   206   261
#> Row11   246   272
#> Row10   265   267
#> Row9    273     2
#> Row8    197   215
#> Row7    195   250
#> Row6      2   216
#> Row5    227   218
#> Row4    252   271
#> Row3    208   229
#> Row2    210   238
#> Row1    191   244

spatAB$plotsNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  100   99   98   97   96 1100 1099 1098 1097  1096  2100  2099  2098
#> Row19   91   92   93   94   95 1091 1092 1093 1094  1095  2091  2092  2093
#> Row18   90   89   88   87   86 1090 1089 1088 1087  1086  2090  2089  2088
#> Row17   81   82   83   84   85 1081 1082 1083 1084  1085  2081  2082  2083
#> Row16   80   79   78   77   76 1080 1079 1078 1077  1076  2080  2079  2078
#> Row15   71   72   73   74   75 1071 1072 1073 1074  1075  2071  2072  2073
#> Row14   70   69   68   67   66 1070 1069 1068 1067  1066  2070  2069  2068
#> Row13   61   62   63   64   65 1061 1062 1063 1064  1065  2061  2062  2063
#> Row12   60   59   58   57   56 1060 1059 1058 1057  1056  2060  2059  2058
#> Row11   51   52   53   54   55 1051 1052 1053 1054  1055  2051  2052  2053
#> Row10   50   49   48   47   46 1050 1049 1048 1047  1046  2050  2049  2048
#> Row9    41   42   43   44   45 1041 1042 1043 1044  1045  2041  2042  2043
#> Row8    40   39   38   37   36 1040 1039 1038 1037  1036  2040  2039  2038
#> Row7    31   32   33   34   35 1031 1032 1033 1034  1035  2031  2032  2033
#> Row6    30   29   28   27   26 1030 1029 1028 1027  1026  2030  2029  2028
#> Row5    21   22   23   24   25 1021 1022 1023 1024  1025  2021  2022  2023
#> Row4    20   19   18   17   16 1020 1019 1018 1017  1016  2020  2019  2018
#> Row3    11   12   13   14   15 1011 1012 1013 1014  1015  2011  2012  2013
#> Row2    10    9    8    7    6 1010 1009 1008 1007  1006  2010  2009  2008
#> Row1     1    2    3    4    5 1001 1002 1003 1004  1005  2001  2002  2003
#>       Col14 Col15
#> Row20  2097  2096
#> Row19  2094  2095
#> Row18  2087  2086
#> Row17  2084  2085
#> Row16  2077  2076
#> Row15  2074  2075
#> Row14  2067  2066
#> Row13  2064  2065
#> Row12  2057  2056
#> Row11  2054  2055
#> Row10  2047  2046
#> Row9   2044  2045
#> Row8   2037  2036
#> Row7   2034  2035
#> Row6   2027  2026
#> Row5   2024  2025
#> Row4   2017  2016
#> Row3   2014  2015
#> Row2   2007  2006
#> Row1   2004  2005

head(spatAB$fieldBook,12)

#>    ID  EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20WRA        1 2021    1   1      1      0    82    gen 82
#> 2   2 20WRA        1 2021    2   1      2      0    54    gen 54
#> 3   3 20WRA        1 2021    3   1      3      0    30    gen 30
#> 4   4 20WRA        1 2021    4   1      4      0    87    gen 87
#> 5   5 20WRA        1 2021    5   1      5      0    19    gen 19
#> 6   6 20WRB        1 2021 1001   1      6      0    96    gen 96
#> 7   7 20WRB        1 2021 1002   1      7      0   135   gen 135
#> 8   8 20WRB        1 2021 1003   1      8      0   110   gen 110
#> 9   9 20WRB        1 2021 1004   1      9      2     2   Check 2
#> 10 10 20WRB        1 2021 1005   1     10      0   114   gen 114
#> 11 11 20WRC        1 2021 2001   1     11      0   190   gen 190
#> 12 12 20WRC        1 2021 2002   1     12      0   199   gen 199