Spatial Un-replicated Diagonal Arrangement Design

Randomly generates an spatial un-replicated diagonal arrangement design.

Usage

diagonal_arrangement(
  nrows = NULL,
  ncols = NULL,
  lines = NULL,
  checks = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  kindExpt = "SUDC",
  splitBy = "row",
  seed = NULL,
  blocks = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows

Number of rows in the field.

ncols

Number of columns in the field.

lines

Number of genotypes, experimental lines or treatments.

checks

Number of genotypes checks.

planter

Option for serpentine or cartesian plot arrangement. By default planter = 'serpentine'.

l

Number of locations or sites. By default l = 1.

plotNumber

Numeric vector with the starting plot number for each location. By default plotNumber = 101.

kindExpt

Type of diagonal design, with single options: Single Un-replicated Diagonal Checks 'SUDC' and Decision Blocks Un-replicated Design with Diagonal Checks 'DBUDC' for multiple experiments. By default kindExpt = 'SUDC'.

splitBy

Option to split the field when kindExpt = 'DBUDC' is selected. By default splitBy = 'row'.

seed

(optional) Real number that specifies the starting seed to obtain reproducible designs.

blocks

Number of experiments or blocks to generate an DBUDC design. If kindExpt = 'DBUDC' and data is null, blocks are mandatory.

exptName

(optional) Name of the experiment.

locationNames

(optional) Names each location.

data

(optional) Data frame with 2 columns: ENTRY | NAME .

Value

A list with five elements.

infoDesign is a list with information on the design parameters.

layoutRandom is a matrix with the randomization layout.

plotsNumber is a matrix with the layout plot number.

data_entry is a data frame with the data input.

fieldBook is a data frame with field book design. This includes the index (Row, Column).

References

Clarke, G. P. Y., & Stefanova, K. T. (2011). Optimal design for early-generation plant breeding trials with unreplicated or partially replicated test lines. Australian & New Zealand Journal of Statistics, 53(4), 461–480.

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a spatial single diagonal arrangement design in one location
# with 270 treatments and 30 check plots for a field with dimensions 15 rows x 20 cols
# in a serpentine arrangement.
spatd <- diagonal_arrangement(nrows = 15, ncols = 20, lines = 270, 
                              checks = 4, 
                              plotNumber = 101, 
                              kindExpt = "SUDC", 
                              planter = "serpentine", 
                              seed = 1987,
                              exptName = "20WRY1", 
                              locationNames = "MINOT")
spatd$infoDesign
#> $rows
#> [1] 15
#> 
#> $columns
#> [1] 20
#> 
#> $treatments
#> [1] 270
#> 
#> $checks
#> [1] 4
#> 
#> $entry_checks
#> [1] 1 2 3 4
#> 
#> $rep_checks
#> [1] 8 7 7 8
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $percent_checks
#> [1] "10%"
#> 
#> $fillers
#> [1] 0
#> 
#> $seed
#> [1] 1987
#> 
#> $id_design
#> [1] 15
#> 
spatd$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row15   98    4  217  213  139   59  247  208  272    24   232     1   222
#> Row14  221   34  243   69    2  145   37  177  233    67    84    78   239
#> Row13  111  237  254  249   33  142  137    4   39   156   101   159    93
#> Row12    4   27  135   23  191  167   71  244  263   176     4   148   227
#> Row11   89  203  269    2   94  205   43  199  228    95   206   224    10
#> Row10  113  151   64  179   79  184    3  165  220     5    49   183    26
#> Row9   115  138  274   51  264  103  105   82  152     2   158   129   119
#> Row8   147  157    3   40  120  190   66   96   13   163   207   127     4
#> Row7   259  255   92   16    9    1   45  180  154    76   189   144    42
#> Row6   245  114  195   15  173  223  117   60    2   171    87   262    62
#> Row5   108    3  230  186   81  141  219  268  132    30   197     1    97
#> Row4   172  118   61   58    3   80    8  218   14    47   194    54   240
#> Row3    12  116  212   75  209   90  193    3  126   131   169   168   133
#> Row2     3  198   28  210  236   25   53  178  130    17     4   238   146
#> Row1   122  231   77    2  175   19   86  136   57   192   258   106   229
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row15   246   261   100   266   160   162    52
#> Row14    56     1     7   270   250    22   235
#> Row13   174   204    20   134     4   155   112
#> Row12   150    36   164    63    70   153   256
#> Row11     1    29   234    65   248    99     6
#> Row10   196    55   102     1   170    83   187
#> Row9     68   265    72   200    21   201     4
#> Row8    109    73   128   241   253   202    11
#> Row7    140   226     2   225   271   166   161
#> Row6    215   273    35   242   124     3    74
#> Row5    188    18   121   260    41   216    91
#> Row4    143     2    48   252    85    44   211
#> Row3    182   185    88    50     1    32   104
#> Row2    257    31   149   125   110   181    46
#> Row1      1   123    38   107   214   251   267
#> 
spatd$plotsNumber
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row15  381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row14  380  379  378  377  376  375  374  373  372   371   370   369   368
#> Row13  341  342  343  344  345  346  347  348  349   350   351   352   353
#> Row12  340  339  338  337  336  335  334  333  332   331   330   329   328
#> Row11  301  302  303  304  305  306  307  308  309   310   311   312   313
#> Row10  300  299  298  297  296  295  294  293  292   291   290   289   288
#> Row9   261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row8   260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row7   221  222  223  224  225  226  227  228  229   230   231   232   233
#> Row6   220  219  218  217  216  215  214  213  212   211   210   209   208
#> Row5   181  182  183  184  185  186  187  188  189   190   191   192   193
#> Row4   180  179  178  177  176  175  174  173  172   171   170   169   168
#> Row3   141  142  143  144  145  146  147  148  149   150   151   152   153
#> Row2   140  139  138  137  136  135  134  133  132   131   130   129   128
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row15   394   395   396   397   398   399   400
#> Row14   367   366   365   364   363   362   361
#> Row13   354   355   356   357   358   359   360
#> Row12   327   326   325   324   323   322   321
#> Row11   314   315   316   317   318   319   320
#> Row10   287   286   285   284   283   282   281
#> Row9    274   275   276   277   278   279   280
#> Row8    247   246   245   244   243   242   241
#> Row7    234   235   236   237   238   239   240
#> Row6    207   206   205   204   203   202   201
#> Row5    194   195   196   197   198   199   200
#> Row4    167   166   165   164   163   162   161
#> Row3    154   155   156   157   158   159   160
#> Row2    127   126   125   124   123   122   121
#> Row1    114   115   116   117   118   119   120
#> 
head(spatd$fieldBook, 12)
#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20WRY1    MINOT 2022  101   1      1      0   122   Gen-122
#> 2   2 20WRY1    MINOT 2022  102   1      2      0   231   Gen-231
#> 3   3 20WRY1    MINOT 2022  103   1      3      0    77    Gen-77
#> 4   4 20WRY1    MINOT 2022  104   1      4      2     2   Check-2
#> 5   5 20WRY1    MINOT 2022  105   1      5      0   175   Gen-175
#> 6   6 20WRY1    MINOT 2022  106   1      6      0    19    Gen-19
#> 7   7 20WRY1    MINOT 2022  107   1      7      0    86    Gen-86
#> 8   8 20WRY1    MINOT 2022  108   1      8      0   136   Gen-136
#> 9   9 20WRY1    MINOT 2022  109   1      9      0    57    Gen-57
#> 10 10 20WRY1    MINOT 2022  110   1     10      0   192   Gen-192
#> 11 11 20WRY1    MINOT 2022  111   1     11      0   258   Gen-258
#> 12 12 20WRY1    MINOT 2022  112   1     12      0   106   Gen-106

# Example 2: Generates a spatial decision block diagonal arrangement design in one location
# with 720 treatments allocated in 5 experiments or blocks for a field with dimensions
# 30 rows x 26 cols in a serpentine arrangement. In this case, we show how to set up the data 
# option with the entries list.
checks <- 5;expts <- 5
list_checks <- paste("CH", 1:checks, sep = "")
treatments <- paste("G", 6:725, sep = "")
BLOCK <- c(rep("ALL", checks), rep(1:expts, c(150,155,95,200,120)))
treatment_list <- data.frame(list(ENTRY = 1:725, NAME = c(list_checks, treatments), BLOCK = BLOCK))
head(treatment_list, 12) 
#>    ENTRY NAME BLOCK
#> 1      1  CH1   ALL
#> 2      2  CH2   ALL
#> 3      3  CH3   ALL
#> 4      4  CH4   ALL
#> 5      5  CH5   ALL
#> 6      6   G6     1
#> 7      7   G7     1
#> 8      8   G8     1
#> 9      9   G9     1
#> 10    10  G10     1
#> 11    11  G11     1
#> 12    12  G12     1
tail(treatment_list, 12)
#>     ENTRY NAME BLOCK
#> 714   714 G714     5
#> 715   715 G715     5
#> 716   716 G716     5
#> 717   717 G717     5
#> 718   718 G718     5
#> 719   719 G719     5
#> 720   720 G720     5
#> 721   721 G721     5
#> 722   722 G722     5
#> 723   723 G723     5
#> 724   724 G724     5
#> 725   725 G725     5
spatDB <- diagonal_arrangement(nrows = 30, 
                               ncols = 26,
                               checks = 5, 
                               plotNumber = 1, 
                               kindExpt = "DBUDC", 
                               planter = "serpentine", 
                               splitBy = "row", 
                               data = treatment_list)
spatDB$infoDesign
#> $rows
#> [1] 30
#> 
#> $columns
#> [1] 26
#> 
#> $treatments
#> [1] 150 155  95 200 120
#> 
#> $checks
#> [1] 5
#> 
#> $entry_checks
#> [1] 1 2 3 4 5
#> 
#> $rep_checks
#> [1] 12 10 13 13 12
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $percent_checks
#> [1] "7.7%"
#> 
#> $fillers
#> [1] 0
#> 
#> $seed
#> [1] -8289.062
#> 
#> $id_design
#> [1] 15
#> 
spatDB$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  695    2  609  692  709  696  617  722  616   651   633   661   646
#> Row29  701  679  718  641  656    2  684  665  687   669   668   655   720
#> Row28  639  694  660  711  640  645  702  673  623     4   647   685   690
#> Row27    5  708  620  666  626  667  705  618  707   662   716   619   672
#> Row26  654  683  612  622    1  659  710  627  608   680   628   637   713
#> Row25  411  488  585  537  589  471  570  462    5   527   414   582   533
#> Row24  558  529  536  481  531  541  511  460  435   476   473   546     2
#> Row23  530  494  455    4  605  504  485  418  574   507   407   501   505
#> Row22  434  597  461  535  593  465  520    4  555   599   427   542   443
#> Row21  463  449  509  467  567  413  514  500  601   412   572     2   594
#> Row20  433  551    2  452  469  552  442  426  604   595   523   579   446
#> Row19  562  590  524  508  560  416    4  526  534   539   453   487   466
#> Row18  423  568  602  588  513  515  503  522  489   580     5   532   492
#> Row17  352    2  356  370  368  376  388  341  334   343   374   330   347
#> Row16  323  335  361  357  316    5  321  387  365   327   332   375   336
#> Row15  390  338  364  378  358  311  366  401  329     1   355   385   397
#> Row14    1  377  350  394  345  328  317  367  319   382   389   362   337
#> Row13  258  196  173  261    3  247  159  262  308   289   202   166   235
#> Row12  275  210  281  240  215  195  273  180    4   224   176   199   263
#> Row11  292  217  243  205  225  254  172  161  190   232   303   244     1
#> Row10  175  256  267    5  208  242  245  193  259   257   293   198   296
#> Row9   184  290  188  309  239  291  170    1  182   187   212   265   160
#> Row8   238  213  204  168  201  301  220  206  307   229   216     3   288
#> Row7    44  155    2  146  131   56   35  287  178   278   300   162   222
#> Row6   135  109   48   28  125  114    1   53   86   136    98    99    72
#> Row5   122  108   66   93  143   88   81   11   92    87     1   127    46
#> Row4   152    5  144   10  141   17    9   64  106     8   151    96   142
#> Row3   113   77  139   78   67    5  145  123   45    22   132   102    58
#> Row2    74   21   13   31    6   60   20   25  101     5   130    61   154
#> Row1     2  117   47  124   43   51   57   42   95    16    54   126   111
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row30   663     5   613   658   676   698   725   630   715   686   629   652
#> Row29   653   642   682   634   607     3   621   703   671   699   677   670
#> Row28   649   719   704   631   700   611   644   681   678     4   610   624
#> Row27     1   614   632   674   689   657   706   691   723   648   712   625
#> Row26   615   675   643   638     3   693   724   721   606   664   688   717
#> Row25   409   477   528   525   464   496   591   600     3   450   561   516
#> Row24   510   521   592   419   598   564   543   550   577   420   441   424
#> Row23   451   547   475     1   553   603   544   425   596   538   565   586
#> Row22   517   444   587   482   573   486   518     4   581   470   512   415
#> Row21   437   478   493   459   576   440   480   557   438   436   417     1
#> Row20   454   406     1   474   472   549   491   490   575   519   584   421
#> Row19   506   483   559   410   498   554     5   569   502   430   429   457
#> Row18   445   408   458   447   468   497   495   448   456   422     3   566
#> Row17   340     4   398   353   381   432   548   484   499   479   571   578
#> Row16   318   313   322   402   314     3   312   373   354   403   372   383
#> Row15   331   344   363   395   348   393   324   326   399     3   380   359
#> Row14     5   349   346   351   360   379   404   369   325   391   384   333
#> Row13   158   295   164   280     4   310   342   400   386   320   315   371
#> Row12   171   233   286   192   284   186   163   294     3   306   283   194
#> Row11   237   246   302   274   230   185   183   218   282   174   255   223
#> Row10   270   299   285     2   264   157   167   219   248   197   250   297
#> Row9    276   272   179   271   260   304   249     2   226   298   181   214
#> Row8    253   221   227   165   277   231   236   241   156   177   269     4
#> Row7    189   268     4   251   191   234   266   211   228   252   279   207
#> Row6    116    52   104    55   105    90     4    14    18   100    82   115
#> Row5    150    84    70   128    39    23    65    80    37    68     1    38
#> Row4    147     4   134   103    32    19     7   140    12    30    24    26
#> Row3     59   119    85    89    33     3    36   118    34    63   138    76
#> Row2     91    73   137    94    27   120   110    79    40     3    50    15
#> Row1      3   112   121   107   129    71    97    29   153   148    49   133
#>       Col26
#> Row30   635
#> Row29   697
#> Row28   714
#> Row27   650
#> Row26   636
#> Row25   583
#> Row24     3
#> Row23   439
#> Row22   540
#> Row21   563
#> Row20   428
#> Row19   545
#> Row18   431
#> Row17   556
#> Row16   396
#> Row15   392
#> Row14   405
#> Row13   339
#> Row12   203
#> Row11     5
#> Row10   209
#> Row9    305
#> Row8    200
#> Row7    169
#> Row6     69
#> Row5     41
#> Row4     83
#> Row3    149
#> Row2     75
#> Row1     62
#> 
spatDB$plotsNumber
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  780  779  778  777  776  775  774  773  772   771   770   769   768
#> Row29  729  730  731  732  733  734  735  736  737   738   739   740   741
#> Row28  728  727  726  725  724  723  722  721  720   719   718   717   716
#> Row27  677  678  679  680  681  682  683  684  685   686   687   688   689
#> Row26  676  675  674  673  672  671  670  669  668   667   666   665   664
#> Row25  625  626  627  628  629  630  631  632  633   634   635   636   637
#> Row24  624  623  622  621  620  619  618  617  616   615   614   613   612
#> Row23  573  574  575  576  577  578  579  580  581   582   583   584   585
#> Row22  572  571  570  569  568  567  566  565  564   563   562   561   560
#> Row21  521  522  523  524  525  526  527  528  529   530   531   532   533
#> Row20  520  519  518  517  516  515  514  513  512   511   510   509   508
#> Row19  469  470  471  472  473  474  475  476  477   478   479   480   481
#> Row18  468  467  466  465  464  463  462  461  460   459   458   457   456
#> Row17  417  418  419  420  421  422  423  424  425   426   427   428   429
#> Row16  416  415  414  413  412  411  410  409  408   407   406   405   404
#> Row15  365  366  367  368  369  370  371  372  373   374   375   376   377
#> Row14  364  363  362  361  360  359  358  357  356   355   354   353   352
#> Row13  313  314  315  316  317  318  319  320  321   322   323   324   325
#> Row12  312  311  310  309  308  307  306  305  304   303   302   301   300
#> Row11  261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row10  260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row9   209  210  211  212  213  214  215  216  217   218   219   220   221
#> Row8   208  207  206  205  204  203  202  201  200   199   198   197   196
#> Row7   157  158  159  160  161  162  163  164  165   166   167   168   169
#> Row6   156  155  154  153  152  151  150  149  148   147   146   145   144
#> Row5   105  106  107  108  109  110  111  112  113   114   115   116   117
#> Row4   104  103  102  101  100   99   98   97   96    95    94    93    92
#> Row3    53   54   55   56   57   58   59   60   61    62    63    64    65
#> Row2    52   51   50   49   48   47   46   45   44    43    42    41    40
#> Row1     1    2    3    4    5    6    7    8    9    10    11    12    13
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row30   767   766   765   764   763   762   761   760   759   758   757   756
#> Row29   742   743   744   745   746   747   748   749   750   751   752   753
#> Row28   715   714   713   712   711   710   709   708   707   706   705   704
#> Row27   690   691   692   693   694   695   696   697   698   699   700   701
#> Row26   663   662   661   660   659   658   657   656   655   654   653   652
#> Row25   638   639   640   641   642   643   644   645   646   647   648   649
#> Row24   611   610   609   608   607   606   605   604   603   602   601   600
#> Row23   586   587   588   589   590   591   592   593   594   595   596   597
#> Row22   559   558   557   556   555   554   553   552   551   550   549   548
#> Row21   534   535   536   537   538   539   540   541   542   543   544   545
#> Row20   507   506   505   504   503   502   501   500   499   498   497   496
#> Row19   482   483   484   485   486   487   488   489   490   491   492   493
#> Row18   455   454   453   452   451   450   449   448   447   446   445   444
#> Row17   430   431   432   433   434   435   436   437   438   439   440   441
#> Row16   403   402   401   400   399   398   397   396   395   394   393   392
#> Row15   378   379   380   381   382   383   384   385   386   387   388   389
#> Row14   351   350   349   348   347   346   345   344   343   342   341   340
#> Row13   326   327   328   329   330   331   332   333   334   335   336   337
#> Row12   299   298   297   296   295   294   293   292   291   290   289   288
#> Row11   274   275   276   277   278   279   280   281   282   283   284   285
#> Row10   247   246   245   244   243   242   241   240   239   238   237   236
#> Row9    222   223   224   225   226   227   228   229   230   231   232   233
#> Row8    195   194   193   192   191   190   189   188   187   186   185   184
#> Row7    170   171   172   173   174   175   176   177   178   179   180   181
#> Row6    143   142   141   140   139   138   137   136   135   134   133   132
#> Row5    118   119   120   121   122   123   124   125   126   127   128   129
#> Row4     91    90    89    88    87    86    85    84    83    82    81    80
#> Row3     66    67    68    69    70    71    72    73    74    75    76    77
#> Row2     39    38    37    36    35    34    33    32    31    30    29    28
#> Row1     14    15    16    17    18    19    20    21    22    23    24    25
#>       Col26
#> Row30   755
#> Row29   754
#> Row28   703
#> Row27   702
#> Row26   651
#> Row25   650
#> Row24   599
#> Row23   598
#> Row22   547
#> Row21   546
#> Row20   495
#> Row19   494
#> Row18   443
#> Row17   442
#> Row16   391
#> Row15   390
#> Row14   339
#> Row13   338
#> Row12   287
#> Row11   286
#> Row10   235
#> Row9    234
#> Row8    183
#> Row7    182
#> Row6    131
#> Row5    130
#> Row4     79
#> Row3     78
#> Row2     27
#> Row1     26
#> 
head(spatDB$fieldBook,12)
#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 Block1        1 2022    1   1      1      2     2       CH2
#> 2   2 Block1        1 2022    2   1      2      0   117      G117
#> 3   3 Block1        1 2022    3   1      3      0    47       G47
#> 4   4 Block1        1 2022    4   1      4      0   124      G124
#> 5   5 Block1        1 2022    5   1      5      0    43       G43
#> 6   6 Block1        1 2022    6   1      6      0    51       G51
#> 7   7 Block1        1 2022    7   1      7      0    57       G57
#> 8   8 Block1        1 2022    8   1      8      0    42       G42
#> 9   9 Block1        1 2022    9   1      9      0    95       G95
#> 10 10 Block1        1 2022   10   1     10      0    16       G16
#> 11 11 Block1        1 2022   11   1     11      0    54       G54
#> 12 12 Block1        1 2022   12   1     12      0   126      G126

# Example 3: Generates a spatial decision block diagonal arrangement design in one location
# with 270 treatments allocated in 3 experiments or blocks for a field with dimensions
# 20 rows x 15 cols in a serpentine arrangement. Which in turn is an augmented block (3 blocks).
spatAB <- diagonal_arrangement(nrows = 20, ncols = 15, lines = 270, 
                               checks = 4, 
                               plotNumber = c(1,1001,2001), 
                               kindExpt = "DBUDC", 
                               planter = "serpentine",
                               exptName = c("20WRA", "20WRB", "20WRC"), 
                               blocks = c(90, 90, 90),
                               splitBy = "column")
spatAB$infoDesign
#> $rows
#> [1] 20
#> 
#> $columns
#> [1] 15
#> 
#> $treatments
#> [1] 90 90 90
#> 
#> $checks
#> [1] 4
#> 
#> $entry_checks
#> [1] 1 2 3 4
#> 
#> $rep_checks
#> [1] 8 8 7 7
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $percent_checks
#> [1] "10%"
#> 
#> $fillers
#> [1] 0
#> 
#> $seed
#> [1] 6966.657
#> 
#> $id_design
#> [1] 15
#> 
spatAB$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20   94    2   25   33   50  177  150  139  168   128   196     2   269
#> Row19   18   30   73   66    3  184   95  146  167   145   256   271   214
#> Row18   34   20   82   13   59  153   96    1  109   169   218   230   234
#> Row17    1   10   58   71   91  105  130  138  148   134     1   237   232
#> Row16   89   31   45    4   28  175  137  102  159   178   272   241   211
#> Row15   68   83   54   76   44  182    4  129  131   132   250   190   226
#> Row14   90   80   53   22   63  181  140  112  133     4   229   222   205
#> Row13   27   12    2   29   11  180   99  173  165   119   270   259     3
#> Row12   24   36   56    5   79    3  125  141  118   149   199   247   225
#> Row11   52   46   17   92   62  160  176  120    2   136   267   189   204
#> Row10    7    1   74   60   67  151  103  110  152   156   268     4   217
#> Row9    15   85   41   69    4  121  135  161  158   123   213   265   203
#> Row8    47   64   61   55   88  111  113    2   98   172   255   260   228
#> Row7     3   48   14   38   37  144  174  170  106   100     3   220   224
#> Row6    43   65   35    1   77  166  126  122  127   124   212   248   194
#> Row5    87   81   70   49   86  143    1  179  117   104   262   195   266
#> Row4    51   19   57   39   93  107  155  154  183     3   219   245   264
#> Row3    16   32    3   40   23  114  108  157  171   142   197   249     4
#> Row2    75   26   42    9    8    2  115  162  163   116   253   193   207
#> Row1     6   72   21   84   78   97  164  147    1   101   273   231   208
#>       Col14 Col15
#> Row20   227   202
#> Row19   198     2
#> Row18   261   221
#> Row17   244   200
#> Row16     2   192
#> Row15   187   233
#> Row14   251   210
#> Row13   186   206
#> Row12   258   235
#> Row11   201   188
#> Row10   252   209
#> Row9    223     1
#> Row8    185   263
#> Row7    246   257
#> Row6      4   254
#> Row5    239   236
#> Row4    191   238
#> Row3    216   242
#> Row2    215   274
#> Row1    240   243
#> 
spatAB$plotsNumber
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  100   99   98   97   96 1100 1099 1098 1097  1096  2100  2099  2098
#> Row19   91   92   93   94   95 1091 1092 1093 1094  1095  2091  2092  2093
#> Row18   90   89   88   87   86 1090 1089 1088 1087  1086  2090  2089  2088
#> Row17   81   82   83   84   85 1081 1082 1083 1084  1085  2081  2082  2083
#> Row16   80   79   78   77   76 1080 1079 1078 1077  1076  2080  2079  2078
#> Row15   71   72   73   74   75 1071 1072 1073 1074  1075  2071  2072  2073
#> Row14   70   69   68   67   66 1070 1069 1068 1067  1066  2070  2069  2068
#> Row13   61   62   63   64   65 1061 1062 1063 1064  1065  2061  2062  2063
#> Row12   60   59   58   57   56 1060 1059 1058 1057  1056  2060  2059  2058
#> Row11   51   52   53   54   55 1051 1052 1053 1054  1055  2051  2052  2053
#> Row10   50   49   48   47   46 1050 1049 1048 1047  1046  2050  2049  2048
#> Row9    41   42   43   44   45 1041 1042 1043 1044  1045  2041  2042  2043
#> Row8    40   39   38   37   36 1040 1039 1038 1037  1036  2040  2039  2038
#> Row7    31   32   33   34   35 1031 1032 1033 1034  1035  2031  2032  2033
#> Row6    30   29   28   27   26 1030 1029 1028 1027  1026  2030  2029  2028
#> Row5    21   22   23   24   25 1021 1022 1023 1024  1025  2021  2022  2023
#> Row4    20   19   18   17   16 1020 1019 1018 1017  1016  2020  2019  2018
#> Row3    11   12   13   14   15 1011 1012 1013 1014  1015  2011  2012  2013
#> Row2    10    9    8    7    6 1010 1009 1008 1007  1006  2010  2009  2008
#> Row1     1    2    3    4    5 1001 1002 1003 1004  1005  2001  2002  2003
#>       Col14 Col15
#> Row20  2097  2096
#> Row19  2094  2095
#> Row18  2087  2086
#> Row17  2084  2085
#> Row16  2077  2076
#> Row15  2074  2075
#> Row14  2067  2066
#> Row13  2064  2065
#> Row12  2057  2056
#> Row11  2054  2055
#> Row10  2047  2046
#> Row9   2044  2045
#> Row8   2037  2036
#> Row7   2034  2035
#> Row6   2027  2026
#> Row5   2024  2025
#> Row4   2017  2016
#> Row3   2014  2015
#> Row2   2007  2006
#> Row1   2004  2005
#> 
head(spatAB$fieldBook,12)
#>    ID  EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20WRA        1 2022    1   1      1      0     6     Gen-6
#> 2   2 20WRA        1 2022    2   1      2      0    72    Gen-72
#> 3   3 20WRA        1 2022    3   1      3      0    21    Gen-21
#> 4   4 20WRA        1 2022    4   1      4      0    84    Gen-84
#> 5   5 20WRA        1 2022    5   1      5      0    78    Gen-78
#> 6   6 20WRB        1 2022 1001   1      6      0    97    Gen-97
#> 7   7 20WRB        1 2022 1002   1      7      0   164   Gen-164
#> 8   8 20WRB        1 2022 1003   1      8      0   147   Gen-147
#> 9   9 20WRB        1 2022 1004   1      9      1     1   Check-1
#> 10 10 20WRB        1 2022 1005   1     10      0   101   Gen-101
#> 11 11 20WRC        1 2022 2001   1     11      0   273   Gen-273
#> 12 12 20WRC        1 2022 2002   1     12      0   231   Gen-231