Generates a Full Factorial Design

It randomly generates a full factorial design across locations.

full_factorial(
  setfactors = NULL,
  reps = NULL,
  l = 1,
  type = 2,
  plotNumber = 101,
  continuous = FALSE,
  planter = "serpentine",
  seed = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

setfactors
Numeric vector with levels of each factor.

reps
Number of replicates (full blocks).

l
Number of locations. By default l = 1.

type
Option for CRD or RCBD designs. Values are type = 1 (CRD) or type = 2 (RCBD). By default type = 2.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

continuous
Logical fpr plot number continuous or not. By default continuous = FALSE.

planter
Option for serpentine or cartesian plot arrangement. By default planter = 'serpentine'.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

locationNames
(optional) Names for each location.

data
(optional) Data frame with the labels of factors.

Value

A list with information on the design parameters.

Data frame with the full factorial field book.

References

Design and Analysis of Experiments, Volume 1, Introduction to Experimental Design. Second Edition. Klaus Hinkelmann & Oscar Kempthorne.John Wiley & Sons, Inc., Hoboken, New Jersey.

Examples

# Example 1: Generates a full factorial with 3 factors each with 2 levels. 
# This in an RCBD arrangement with 3 reps.
fullFact1 <- full_factorial(setfactors = c(2,2,2), reps = 3, l = 1, type = 2, 
                            plotNumber = 101, 
                            continuous = TRUE, 
                            planter = "serpentine", 
                            seed = 325, 
                            locationNames = "FARGO")
fullFact1$infoDesign

#> $factors
#> [1] "A" "B" "C"
#> 
#> $levels
#> $levels[[1]]
#> [1] 1 2
#> 
#> $levels[[2]]
#> [1] 1 2
#> 
#> $levels[[3]]
#> [1] 1 2
#> 
#> 
#> $runs
#> [1] 8
#> 
#> $alltreatments
#>   A B C
#> 1 1 1 1
#> 2 2 1 1
#> 3 1 2 1
#> 4 2 2 1
#> 5 1 1 2
#> 6 2 1 2
#> 7 1 2 2
#> 8 2 2 2
#> 
#> $Reps
#> [1] 3
#> 
#> $Locations
#> [1] 1
#> 
#> $locationNames
#> [1] "FARGO"
#> 
#> $kind
#> [1] "RCBD"
#> 

head(fullFact1$fieldBook,10)

#>    LOCATION PLOT REP FACTOR_A FACTOR_B FACTOR_C TRT_COMB
#> 1     FARGO  101   1        2        1        2    2*1*2
#> 2     FARGO  102   1        2        2        1    2*2*1
#> 3     FARGO  103   1        2        1        1    2*1*1
#> 4     FARGO  104   1        1        1        1    1*1*1
#> 5     FARGO  105   1        1        2        1    1*2*1
#> 6     FARGO  106   1        1        2        2    1*2*2
#> 7     FARGO  107   1        2        2        2    2*2*2
#> 8     FARGO  108   1        1        1        2    1*1*2
#> 16    FARGO  109   2        2        2        1    2*2*1
#> 15    FARGO  110   2        1        1        2    1*1*2

# Example 2: Generates a full factorial with 3 factors and each with levels: 2,3, 
# and 2, respectively. 
# This in an RCBD arrangement with 5 reps in 3 locations.
fullFact2 <- full_factorial(setfactors = c(2,3,2), reps = 5, l = 3, type = 2, 
                            plotNumber = c(101,1001,2001), 
                            continuous = FALSE, 
                            planter = "serpentine", 
                            seed = 326, 
                            locationNames = c("Loc1","Loc2","Loc3"))
fullFact2$infoDesign

#> $factors
#> [1] "A" "B" "C"
#> 
#> $levels
#> $levels[[1]]
#> [1] 1 2
#> 
#> $levels[[2]]
#> [1] 1 2 3
#> 
#> $levels[[3]]
#> [1] 1 2
#> 
#> 
#> $runs
#> [1] 12
#> 
#> $alltreatments
#>    A B C
#> 1  1 1 1
#> 2  2 1 1
#> 3  1 2 1
#> 4  2 2 1
#> 5  1 3 1
#> 6  2 3 1
#> 7  1 1 2
#> 8  2 1 2
#> 9  1 2 2
#> 10 2 2 2
#> 11 1 3 2
#> 12 2 3 2
#> 
#> $Reps
#> [1] 5
#> 
#> $Locations
#> [1] 3
#> 
#> $locationNames
#> [1] "Loc1" "Loc2" "Loc3"
#> 
#> $kind
#> [1] "RCBD"
#> 

head(fullFact2$fieldBook,10)

#>    LOCATION PLOT REP FACTOR_A FACTOR_B FACTOR_C TRT_COMB
#> 1      Loc1  101   1        1        3        2    1*3*2
#> 2      Loc1  102   1        2        2        2    2*2*2
#> 3      Loc1  103   1        2        3        1    2*3*1
#> 4      Loc1  104   1        2        3        2    2*3*2
#> 5      Loc1  105   1        2        1        2    2*1*2
#> 6      Loc1  106   1        1        1        1    1*1*1
#> 7      Loc1  107   1        2        2        1    2*2*1
#> 8      Loc1  108   1        1        1        2    1*1*2
#> 9      Loc1  109   1        1        3        1    1*3*1
#> 10     Loc1  110   1        2        1        1    2*1*1