Generates an Spatial Un-replicated Optimized Arrangement Design

Randomly generates a spatial un-replicated optimized arrangement design, where the distance between checks is maximized in such a way that each row and column have control plots. Note that design generation needs the dimension of the field (number of rows and columns).

optimized_arrangement(
  nrows = NULL,
  ncols = NULL,
  lines = NULL,
  amountChecks = NULL,
  checks = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  seed = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows
Number of rows in the field.

ncols
Number of columns in the field.

lines
Number of genotypes, experimental lines or treatments.

amountChecks
Integer with the amount total of checks or a numeric vector with the replicates of each check label.

checks
Number of genotypes as checks.

planter
Option for serpentine or cartesian arrangement. By default planter = 'serpentine'.

l
Number of locations. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

exptName
(optional) Name of the experiment.

locationNames
(optional) Name for each location.

data
(optional) Data frame with 3 columns: ENTRY | NAME | REPS.

Value

A list with information on the design parameters.

A matrix with the randomization layout.

A matrix with the layout plot number.

A data frame with fieldBook design. This includes the index (Row, Column).

References

Clarke, G. P. Y., & Stefanova, K. T. (2011). Optimal design for early-generation plant breeding trials with unreplicated or partially replicated test lines. Australian & New Zealand Journal of Statistics, 53(4), 461–480.

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a spatial unreplicated optimized arrangement design in one location
# with 362 genotypes + 38 check plots (5 checks) for a field with dimension 20 rows x 20 cols.
OptimAd1 <- optimized_arrangement(nrows = 20, ncols = 20, lines = 362, 
                                  amountChecks = 38, 
                                  checks = 1:5,
                                  planter = "cartesian", 
                                  plotNumber = 101,
                                  seed = 14,
                                  exptName = "20RW1",
                                  locationNames = "CASSELTON")
OptimAd1$infoDesign

#> $Lines
#> [1] 362
#> 
#> $checks
#> [1] 1 2 3 4 5
#> 
#> $RepChecks
#> [1]  7  7 10  7  7
#> 

OptimAd1$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  182  109   64   18  168   81    1  220  125    13     2   360    90
#> Row19  119   11  130  181  206  146  107  357   33    93   304   189    37
#> Row18    8  163  286  334  297   26  180  233   49   250    69   246   208
#> Row17  201  229  345    3   42  113  353   83  351   177   270    77   259
#> Row16   40   24   47  151  253   61  131  197    3   335    48     2   133
#> Row15  364  187  196  188    4  331  203  326  165   350   238   325   292
#> Row14    2  161  358  309    1  311  210  274  293    31    76   317   103
#> Row13  135  154  265  247  192  362  150   58  251     3   256   367   244
#> Row12  200  280  144  121  342  100   65  281   62   243    16   321   359
#> Row11  354  205  300  279  278   70   68  271  139   193    60   242    75
#> Row10  337  212   85    5  336   38   32  172   22   277    78   255   141
#> Row9    27  263   35  110  148  234  285  258  136   183   137     3   262
#> Row8   237  175    7  145   21   84  164  301    1   194   252   344    34
#> Row7   122  264    1   17   82  282  173  217  149   112   312   195   241
#> Row6   318    2  209  166   19  179  178  268  152    88   267   249   329
#> Row5   134  272    5  223  294   43  330   51   25   124     4   346   170
#> Row4     6  227  106  199   12    2    9    5   94    36   158   296   275
#> Row3   202  287  332  211  108    3  320  302  254   322   132   236     4
#> Row2     1   91   92   52  288  128  114  245  290   235    71   361     1
#> Row1   185   96  104  216  215  169    2    4  226   248   126    74   156
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row20    20    63   339   306    45    10   153
#> Row19     2   123   129   231     3   221   105
#> Row18   295     1   328   355    44     5   310
#> Row17   313   160   102     2   352   118    54
#> Row16    29   184   327   116   239   319   228
#> Row15     2   298   365   289   366   232   303
#> Row14   120   224   261   101    55    23   174
#> Row13   284   190   276   307   315    28   305
#> Row12    95   260   347    46   162     5     5
#> Row11   343     4   348     2    89    86   266
#> Row10   299   157     5   207    50    57    14
#> Row9     98    73     3   138   143    97   159
#> Row8    225   176   147    41     4    30   111
#> Row7    308   314   316    53   341   257     4
#> Row6     99   186   333   191   115   230    59
#> Row5    222   273    56   140   338   219   155
#> Row4    214    66    67   218   198   171   269
#> Row3    142    15   340    39   291   204    72
#> Row2    213   240   283   127    79   363   349
#> Row1    117    87   167   324   323   356    80

OptimAd1$plotNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  481  482  483  484  485  486  487  488  489   490   491   492   493
#> Row19  461  462  463  464  465  466  467  468  469   470   471   472   473
#> Row18  441  442  443  444  445  446  447  448  449   450   451   452   453
#> Row17  421  422  423  424  425  426  427  428  429   430   431   432   433
#> Row16  401  402  403  404  405  406  407  408  409   410   411   412   413
#> Row15  381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row14  361  362  363  364  365  366  367  368  369   370   371   372   373
#> Row13  341  342  343  344  345  346  347  348  349   350   351   352   353
#> Row12  321  322  323  324  325  326  327  328  329   330   331   332   333
#> Row11  301  302  303  304  305  306  307  308  309   310   311   312   313
#> Row10  281  282  283  284  285  286  287  288  289   290   291   292   293
#> Row9   261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row8   241  242  243  244  245  246  247  248  249   250   251   252   253
#> Row7   221  222  223  224  225  226  227  228  229   230   231   232   233
#> Row6   201  202  203  204  205  206  207  208  209   210   211   212   213
#> Row5   181  182  183  184  185  186  187  188  189   190   191   192   193
#> Row4   161  162  163  164  165  166  167  168  169   170   171   172   173
#> Row3   141  142  143  144  145  146  147  148  149   150   151   152   153
#> Row2   121  122  123  124  125  126  127  128  129   130   131   132   133
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row20   494   495   496   497   498   499   500
#> Row19   474   475   476   477   478   479   480
#> Row18   454   455   456   457   458   459   460
#> Row17   434   435   436   437   438   439   440
#> Row16   414   415   416   417   418   419   420
#> Row15   394   395   396   397   398   399   400
#> Row14   374   375   376   377   378   379   380
#> Row13   354   355   356   357   358   359   360
#> Row12   334   335   336   337   338   339   340
#> Row11   314   315   316   317   318   319   320
#> Row10   294   295   296   297   298   299   300
#> Row9    274   275   276   277   278   279   280
#> Row8    254   255   256   257   258   259   260
#> Row7    234   235   236   237   238   239   240
#> Row6    214   215   216   217   218   219   220
#> Row5    194   195   196   197   198   199   200
#> Row4    174   175   176   177   178   179   180
#> Row3    154   155   156   157   158   159   160
#> Row2    134   135   136   137   138   139   140
#> Row1    114   115   116   117   118   119   120

head(OptimAd1$fieldBook,12)

#>    ID  EXPT  LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20RW1 CASSELTON 2021  101   1      1      0   185      G185
#> 2   2 20RW1 CASSELTON 2021  102   1      2      0    96       G96
#> 3   3 20RW1 CASSELTON 2021  103   1      3      0   104      G104
#> 4   4 20RW1 CASSELTON 2021  104   1      4      0   216      G216
#> 5   5 20RW1 CASSELTON 2021  105   1      5      0   215      G215
#> 6   6 20RW1 CASSELTON 2021  106   1      6      0   169      G169
#> 7   7 20RW1 CASSELTON 2021  107   1      7      1     2       CH2
#> 20  8 20RW1 CASSELTON 2021  108   1      8      1     4       CH4
#> 24  9 20RW1 CASSELTON 2021  109   1      9      0   226      G226
#> 25 10 20RW1 CASSELTON 2021  110   1     10      0   248      G248
#> 26 11 20RW1 CASSELTON 2021  111   1     11      0   126      G126
#> 27 12 20RW1 CASSELTON 2021  112   1     12      0    74       G74

                  
# Example 2: Generates a spatial unreplicated optimized arrangement design in one location
# with 635 genotypes + 65 check plots (4 checks) for a field with dimension 20 rows x 35 cols.
# As example, we set up the data option with the entries list.
checks <- 4
list_checks <- paste("CH", 1:checks, sep = "")
treatments <- paste("G", 5:639, sep = "")
REPS <- c(17, 16, 16, 16, rep(1, 635))
treatment_list <- data.frame(list(ENTRY = 1:639, NAME = c(list_checks, treatments), REPS = REPS))
head(treatment_list, 12) 

#>    ENTRY NAME REPS
#> 1      1  CH1   17
#> 2      2  CH2   16
#> 3      3  CH3   16
#> 4      4  CH4   16
#> 5      5   G5    1
#> 6      6   G6    1
#> 7      7   G7    1
#> 8      8   G8    1
#> 9      9   G9    1
#> 10    10  G10    1
#> 11    11  G11    1
#> 12    12  G12    1

tail(treatment_list, 12)

#>     ENTRY NAME REPS
#> 628   628 G628    1
#> 629   629 G629    1
#> 630   630 G630    1
#> 631   631 G631    1
#> 632   632 G632    1
#> 633   633 G633    1
#> 634   634 G634    1
#> 635   635 G635    1
#> 636   636 G636    1
#> 637   637 G637    1
#> 638   638 G638    1
#> 639   639 G639    1

OptimAd2 <- optimized_arrangement(nrows = 20, ncols = 35, 
                                  planter = "serpentine", 
                                  plotNumber = 101,
                                  seed = 12,
                                  exptName = "20YWA2",
                                  locationNames = "MINOT",
                                  data = treatment_list)
OptimAd2$infoDesign

#> $Lines
#> [1] 635
#> 
#> $checks
#> [1] 1 2 3 4
#> 
#> $RepChecks
#> [1] 17 16 16 16
#> 

OptimAd2$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  632  233  275    2  109  429  526   64  388   356   180   340   217
#> Row19  289  128  362  430  373  513  393    2  382   181   569   389   406
#> Row18  578  498    2  564   27   63  441   45  182   376   522   198   585
#> Row17  101  334  558  627    2  270  146  106  537    61   177    16   445
#> Row16  597  617   15  369  478  200  390   36  166   223     1   481   462
#> Row15  322  500  409   52   21  516  421  132  282   514   505   125   354
#> Row14  366  168  119  465  621    4   32    2  497     1   544   176   215
#> Row13    1  407  108   93  427  589   81  380  309   556    90     2   566
#> Row12  620  294  375  280  574  104   12  379  450    14   102     4    10
#> Row11  413    2  239    1  346   28  602  169   19   263   494   467   515
#> Row10  145  512  364   86   22  288    3  614  531   254   344    69     4
#> Row9   248  472    6  100  349  357   55  571  232     3   359   553     1
#> Row8     3  588  463  629  351  613  123  243  178   378   591   587   110
#> Row7   129  581    3  608  368  170  199   11  570   484   404   291   222
#> Row6   464    4  528  384  371  547  436  133  144   579     2    50   451
#> Row5    20  171  319  140  184    3  308  147    1    49   205   245   572
#> Row4   535  523  201  473  636  314  607  116   91   634    92    75   204
#> Row3    95  580  127  536  563  331  534  228   67   399   453   477   342
#> Row2   599   72   56  598  615  391  565  412  350     5   361   532   509
#> Row1   164  419  210  605   94  358  387  552  297   437   143   352   583
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row20   370   360    39   618   296   186   235   142     1   279     7   631
#> Row19    42   122    78    41     1   284     3   555   475   543   541   207
#> Row18     2   157   551   616   112   400   521   506   300   559   343   435
#> Row17   538   341    48   459   424   619   492   325   600   221     4     2
#> Row16   576   542   355   230     2   567   172   487   417   273   160   241
#> Row15    13    98     2     2   130   272   550   557   483   226   449   267
#> Row14   161   256   258   154   495   452   259   310   212   103   320   432
#> Row13   251   633   293   305   165   156   622   192   312    71   196    26
#> Row12   268   218   247   338    30   183   414     1   134   316     3   611
#> Row11   630     4   394   253    74    88   115   410   403   540   137   236
#> Row10   120   336   302   203   246   329   486   493   596    96   496   269
#> Row9      1   281   396   624   327    37   367   491   471   167    84   539
#> Row8    238   593     1   460    51   260   529    59   518     3   335   485
#> Row7    290   213    87     3   604    76   252     1   304   298   345   261
#> Row6    638   313   214   173   117    58   188    60   138   216   348   229
#> Row5    262    73   426    33   265     1   577    77   250   187   625    43
#> Row4    527   639   150   174   415   431   508   135   139   439   422    38
#> Row3    386   163   383   458   392     4   479   448    85   525    29   402
#> Row2     44    24   517   433    23   420   337   398   482   131   141     3
#> Row1    197     4   480   189   255   562   299   507     3     1   105   332
#>       Col26 Col27 Col28 Col29 Col30 Col31 Col32 Col33 Col34 Col35
#> Row20   330   190     4   152   121    53   159   573     1   219
#> Row19   501   635   321   418   328   202   488   423   287   326
#> Row18   510    18   610   590   603   292     2   476   575   637
#> Row17   612    25   520   244   149   446   401   466    65   405
#> Row16     2   286   277   554   306     3    54   560    66   623
#> Row15   249   333   126   237   155     4   499   546   266   311
#> Row14   601   586   271     4     8    31   220   374   208   469
#> Row13    57   594     3   278   113   365    35   148   408   324
#> Row12   455   285    68   153   225   561    99   474   303   231
#> Row11    70   276   162   568     4   257   339   490   626   377
#> Row10   372   468   470   411   315   179    40     1    46   224
#> Row9     80   347   114   592   428    62   548   175   234   136
#> Row8    318   381   442   549   227    97    17    34   195     3
#> Row7    206   118   503   416   440    79    47   438   194   385
#> Row6    489   124   191   530     4    89   461   317   457   502
#> Row5    395   307   158   582    82   397   211   434   240   107
#> Row4      3   606   584     4   301   425   209   595    83     4
#> Row3    456     1   323   274   151   504   443   628     4   242
#> Row2    511     2   264   185   447   519     3   609   545   295
#> Row1    444   454   111   363   533   193   353   524   283     9

OptimAd2$plotNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  800  799  798  797  796  795  794  793  792   791   790   789   788
#> Row19  731  732  733  734  735  736  737  738  739   740   741   742   743
#> Row18  730  729  728  727  726  725  724  723  722   721   720   719   718
#> Row17  661  662  663  664  665  666  667  668  669   670   671   672   673
#> Row16  660  659  658  657  656  655  654  653  652   651   650   649   648
#> Row15  591  592  593  594  595  596  597  598  599   600   601   602   603
#> Row14  590  589  588  587  586  585  584  583  582   581   580   579   578
#> Row13  521  522  523  524  525  526  527  528  529   530   531   532   533
#> Row12  520  519  518  517  516  515  514  513  512   511   510   509   508
#> Row11  451  452  453  454  455  456  457  458  459   460   461   462   463
#> Row10  450  449  448  447  446  445  444  443  442   441   440   439   438
#> Row9   381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row8   380  379  378  377  376  375  374  373  372   371   370   369   368
#> Row7   311  312  313  314  315  316  317  318  319   320   321   322   323
#> Row6   310  309  308  307  306  305  304  303  302   301   300   299   298
#> Row5   241  242  243  244  245  246  247  248  249   250   251   252   253
#> Row4   240  239  238  237  236  235  234  233  232   231   230   229   228
#> Row3   171  172  173  174  175  176  177  178  179   180   181   182   183
#> Row2   170  169  168  167  166  165  164  163  162   161   160   159   158
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row20   787   786   785   784   783   782   781   780   779   778   777   776
#> Row19   744   745   746   747   748   749   750   751   752   753   754   755
#> Row18   717   716   715   714   713   712   711   710   709   708   707   706
#> Row17   674   675   676   677   678   679   680   681   682   683   684   685
#> Row16   647   646   645   644   643   642   641   640   639   638   637   636
#> Row15   604   605   606   607   608   609   610   611   612   613   614   615
#> Row14   577   576   575   574   573   572   571   570   569   568   567   566
#> Row13   534   535   536   537   538   539   540   541   542   543   544   545
#> Row12   507   506   505   504   503   502   501   500   499   498   497   496
#> Row11   464   465   466   467   468   469   470   471   472   473   474   475
#> Row10   437   436   435   434   433   432   431   430   429   428   427   426
#> Row9    394   395   396   397   398   399   400   401   402   403   404   405
#> Row8    367   366   365   364   363   362   361   360   359   358   357   356
#> Row7    324   325   326   327   328   329   330   331   332   333   334   335
#> Row6    297   296   295   294   293   292   291   290   289   288   287   286
#> Row5    254   255   256   257   258   259   260   261   262   263   264   265
#> Row4    227   226   225   224   223   222   221   220   219   218   217   216
#> Row3    184   185   186   187   188   189   190   191   192   193   194   195
#> Row2    157   156   155   154   153   152   151   150   149   148   147   146
#> Row1    114   115   116   117   118   119   120   121   122   123   124   125
#>       Col26 Col27 Col28 Col29 Col30 Col31 Col32 Col33 Col34 Col35
#> Row20   775   774   773   772   771   770   769   768   767   766
#> Row19   756   757   758   759   760   761   762   763   764   765
#> Row18   705   704   703   702   701   700   699   698   697   696
#> Row17   686   687   688   689   690   691   692   693   694   695
#> Row16   635   634   633   632   631   630   629   628   627   626
#> Row15   616   617   618   619   620   621   622   623   624   625
#> Row14   565   564   563   562   561   560   559   558   557   556
#> Row13   546   547   548   549   550   551   552   553   554   555
#> Row12   495   494   493   492   491   490   489   488   487   486
#> Row11   476   477   478   479   480   481   482   483   484   485
#> Row10   425   424   423   422   421   420   419   418   417   416
#> Row9    406   407   408   409   410   411   412   413   414   415
#> Row8    355   354   353   352   351   350   349   348   347   346
#> Row7    336   337   338   339   340   341   342   343   344   345
#> Row6    285   284   283   282   281   280   279   278   277   276
#> Row5    266   267   268   269   270   271   272   273   274   275
#> Row4    215   214   213   212   211   210   209   208   207   206
#> Row3    196   197   198   199   200   201   202   203   204   205
#> Row2    145   144   143   142   141   140   139   138   137   136
#> Row1    126   127   128   129   130   131   132   133   134   135

head(OptimAd2$fieldBook,12)

#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20YWA2    MINOT 2021  101   1      1      0   164      G164
#> 2   2 20YWA2    MINOT 2021  102   1      2      0   419      G419
#> 3   3 20YWA2    MINOT 2021  103   1      3      0   210      G210
#> 4   4 20YWA2    MINOT 2021  104   1      4      0   605      G605
#> 5   5 20YWA2    MINOT 2021  105   1      5      0    94       G94
#> 6   6 20YWA2    MINOT 2021  106   1      6      0   358      G358
#> 7   7 20YWA2    MINOT 2021  107   1      7      0   387      G387
#> 8   8 20YWA2    MINOT 2021  108   1      8      0   552      G552
#> 9   9 20YWA2    MINOT 2021  109   1      9      0   297      G297
#> 10 10 20YWA2    MINOT 2021  110   1     10      0   437      G437
#> 11 11 20YWA2    MINOT 2021  111   1     11      0   143      G143
#> 12 12 20YWA2    MINOT 2021  112   1     12      0   352      G352