Generates an Spatial Un-replicated Optimized Arrangement Design

Randomly generates a spatial un-replicated optimized arrangement design, where the distance between checks is maximized in such a way that each row and column have control plots. Note that design generation needs the dimension of the field (number of rows and columns).

Usage

optimized_arrangement(
  nrows = NULL,
  ncols = NULL,
  lines = NULL,
  amountChecks = NULL,
  checks = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  seed = NULL,
  exptName = NULL,
  locationNames = NULL,
  optim = TRUE,
  data = NULL
)

Arguments

nrows

Number of rows in the field.

ncols

Number of columns in the field.

lines

Number of genotypes, experimental lines or treatments.

amountChecks

Integer with the amount total of checks or a numeric vector with the replicates of each check label.

checks

Number of genotypes as checks.

planter

Option for serpentine or cartesian arrangement. By default planter = 'serpentine'.

l

Number of locations. By default l = 1.

plotNumber

Numeric vector with the starting plot number for each location. By default plotNumber = 101.

seed

(optional) Real number that specifies the starting seed to obtain reproducible designs.

exptName

(optional) Name of the experiment.

locationNames

(optional) Name for each location.

optim

By default optim = TRUE.

data

(optional) Data frame with 3 columns: ENTRY | NAME | REPS.

Value

A list with five elements.

infoDesign is a list with information on the design parameters.

layoutRandom is a matrix with the randomization layout.

plotNumber is a matrix with the layout plot number.

dataEntry is a data frame with the data input.

genEntries is a list with the entries for replicated and no replicated part.

fieldBook is a data frame with field book design. This includes the index (Row, Column).

References

Clarke, G. P. Y., & Stefanova, K. T. (2011). Optimal design for early-generation plant breeding trials with unreplicated or partially replicated test lines. Australian & New Zealand Journal of Statistics, 53(4), 461–480.

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a spatial unreplicated optimized arrangement design in one location
# with 108 genotypes + 12 check plots (4 checks) for a field with dimension 10 rows x 12 cols.
OptimAd1 <- optimized_arrangement(
  nrows = 10, 
  ncols = 12, 
  lines = 108, 
  amountChecks = 12, 
  checks = 1:4,
  planter = "cartesian", 
  plotNumber = 101,
  seed = 14,
  exptName = "20RW1",
  locationNames = "CASSELTON"
 )
OptimAd1$infoDesign
#> $rows
#> [1] 10
#> 
#> $columns
#> [1] 12
#> 
#> $treatments
#> [1] 108
#> 
#> $checks
#> [1] 4
#> 
#> $entry_checks
#> [1] 1 2 3 4
#> 
#> $rep_checks
#> [1] 3 3 3 3
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "cartesian"
#> 
#> $seed
#> [1] 14
#> 
#> $id_design
#> [1] 16
#> 
OptimAd1$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12
#> Row10   74    2   56   19   38  105   97   60   86    45   104    92
#> Row9   111    3   88   13   90   77   37    1   21    81    50   103
#> Row8    28   62  110   68   93   43   40   29   32    80     4   107
#> Row7    15   10   83   63  102   25   52   27   34     6    85    18
#> Row6    12   14   89   67   79    5  109   84   30   100   108    94
#> Row5    39   91   59   70   99   49    9   35   17    95    82    51
#> Row4     4   57   53   55   66  106   69   76   11    78    96    41
#> Row3    71   23    7   47   16   24   54   31   75   112     8     3
#> Row2    48    3   65   44  101   98   64   73   36     1    33    42
#> Row1     1    2   87   61   46   72   22   20    2     4    26    58
#> 
OptimAd1$plotNumber
#> [[1]]
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
#>  [1,]  209  210  211  212  213  214  215  216  217   218   219   220
#>  [2,]  197  198  199  200  201  202  203  204  205   206   207   208
#>  [3,]  185  186  187  188  189  190  191  192  193   194   195   196
#>  [4,]  173  174  175  176  177  178  179  180  181   182   183   184
#>  [5,]  161  162  163  164  165  166  167  168  169   170   171   172
#>  [6,]  149  150  151  152  153  154  155  156  157   158   159   160
#>  [7,]  137  138  139  140  141  142  143  144  145   146   147   148
#>  [8,]  125  126  127  128  129  130  131  132  133   134   135   136
#>  [9,]  113  114  115  116  117  118  119  120  121   122   123   124
#> [10,]  101  102  103  104  105  106  107  108  109   110   111   112
#> 
head(OptimAd1$fieldBook,12)
#>    ID  EXPT  LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20RW1 CASSELTON 2023  101   1      1      1     1       CH1
#> 2   2 20RW1 CASSELTON 2023  102   1      2      2     2       CH2
#> 3   3 20RW1 CASSELTON 2023  103   1      3      0    87       G87
#> 4   4 20RW1 CASSELTON 2023  104   1      4      0    61       G61
#> 5   5 20RW1 CASSELTON 2023  105   1      5      0    46       G46
#> 6   6 20RW1 CASSELTON 2023  106   1      6      0    72       G72
#> 7   7 20RW1 CASSELTON 2023  107   1      7      0    22       G22
#> 8   8 20RW1 CASSELTON 2023  108   1      8      0    20       G20
#> 9   9 20RW1 CASSELTON 2023  109   1      9      2     2       CH2
#> 10 10 20RW1 CASSELTON 2023  110   1     10      4     4       CH4
#> 11 11 20RW1 CASSELTON 2023  111   1     11      0    26       G26
#> 12 12 20RW1 CASSELTON 2023  112   1     12      0    58       G58
                  
# Example 2: Generates a spatial unreplicated optimized arrangement design in one location
# with 200 genotypes + 20 check plots (4 checks) for a field with dimension 10 rows x 22 cols.
# As example, we set up the data option with the entries list.
checks <- 4
list_checks <- paste("CH", 1:checks, sep = "")
treatments <- paste("G", 5:204, sep = "")
REPS <- c(5, 5, 5, 5, rep(1, 200))
treatment_list <- data.frame(list(ENTRY = 1:204, NAME = c(list_checks, treatments), REPS = REPS))
head(treatment_list, 12) 
#>    ENTRY NAME REPS
#> 1      1  CH1    5
#> 2      2  CH2    5
#> 3      3  CH3    5
#> 4      4  CH4    5
#> 5      5   G5    1
#> 6      6   G6    1
#> 7      7   G7    1
#> 8      8   G8    1
#> 9      9   G9    1
#> 10    10  G10    1
#> 11    11  G11    1
#> 12    12  G12    1
tail(treatment_list, 12)
#>     ENTRY NAME REPS
#> 193   193 G193    1
#> 194   194 G194    1
#> 195   195 G195    1
#> 196   196 G196    1
#> 197   197 G197    1
#> 198   198 G198    1
#> 199   199 G199    1
#> 200   200 G200    1
#> 201   201 G201    1
#> 202   202 G202    1
#> 203   203 G203    1
#> 204   204 G204    1
OptimAd2 <- optimized_arrangement(
  nrows = 10, 
  ncols = 22, 
  planter = "serpentine", 
  plotNumber = 101,
  seed = 120,
  exptName = "20YWA2",
  locationNames = "MINOT",
  data = treatment_list
)
#> [1] 220
OptimAd2$infoDesign
#> $rows
#> [1] 10
#> 
#> $columns
#> [1] 22
#> 
#> $treatments
#> [1] 200
#> 
#> $checks
#> [1] 4
#> 
#> $entry_checks
#> [1] 1 2 3 4
#> 
#> $rep_checks
#> [1] 5 5 5 5
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $seed
#> [1] 120
#> 
#> $id_design
#> [1] 16
#> 
OptimAd2$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row10  199   94   70  133   40   62   44   99   24    65    33   196   135
#> Row9    10  125   74  170  144   45  138  204    3   145     4   157    66
#> Row8    76  111  149   54  152  203   32   81  126    55   147   150   127
#> Row7   140   21   96   68   41  190  141   56  128    11   189   102    82
#> Row6     4  186  198  146  171  165   14  178   88     2    23    69     1
#> Row5   136  200  110  197  175  122    1   86   59   172   179   100     4
#> Row4   142   42  143   83   77  119  161   31   58    48   148   194     5
#> Row3     3    6   85  154  103   57  116   37  130   192   101   201     2
#> Row2   115   91   61   98    2   12  174   47  151     3    87   114     1
#> Row1   160    1  164  180    8   27  169  176  188    18    84    34   106
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22
#> Row10    38    15    79    52   155    16   121    90     1
#> Row9    118    80     3    26    35    73     4   162   181
#> Row8     95   153     4   173   132    51   166    53   156
#> Row7     22    97    89     2   113   123   168    63    29
#> Row6    139   177   159    39   108    20    78   129    75
#> Row5    105   158   185    28   137    13    19     7     3
#> Row4     25    43   124   193   112   183   120    50   182
#> Row3    134    30   167   187   202    49    17   131    71
#> Row2    104    72    36    60   191     9    92     2    64
#> Row1    184    46   107   195   117   109    67   163    93
#> 
OptimAd2$plotNumber
#> [[1]]
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  320  319  318  317  316  315  314  313  312   311   310   309   308
#>  [2,]  277  278  279  280  281  282  283  284  285   286   287   288   289
#>  [3,]  276  275  274  273  272  271  270  269  268   267   266   265   264
#>  [4,]  233  234  235  236  237  238  239  240  241   242   243   244   245
#>  [5,]  232  231  230  229  228  227  226  225  224   223   222   221   220
#>  [6,]  189  190  191  192  193  194  195  196  197   198   199   200   201
#>  [7,]  188  187  186  185  184  183  182  181  180   179   178   177   176
#>  [8,]  145  146  147  148  149  150  151  152  153   154   155   156   157
#>  [9,]  144  143  142  141  140  139  138  137  136   135   134   133   132
#> [10,]  101  102  103  104  105  106  107  108  109   110   111   112   113
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22]
#>  [1,]   307   306   305   304   303   302   301   300   299
#>  [2,]   290   291   292   293   294   295   296   297   298
#>  [3,]   263   262   261   260   259   258   257   256   255
#>  [4,]   246   247   248   249   250   251   252   253   254
#>  [5,]   219   218   217   216   215   214   213   212   211
#>  [6,]   202   203   204   205   206   207   208   209   210
#>  [7,]   175   174   173   172   171   170   169   168   167
#>  [8,]   158   159   160   161   162   163   164   165   166
#>  [9,]   131   130   129   128   127   126   125   124   123
#> [10,]   114   115   116   117   118   119   120   121   122
#> 
head(OptimAd2$fieldBook,12)
#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20YWA2    MINOT 2023  101   1      1      0   160      G160
#> 2   2 20YWA2    MINOT 2023  102   1      2      1     1       CH1
#> 3   3 20YWA2    MINOT 2023  103   1      3      0   164      G164
#> 4   4 20YWA2    MINOT 2023  104   1      4      0   180      G180
#> 5   5 20YWA2    MINOT 2023  105   1      5      0     8        G8
#> 6   6 20YWA2    MINOT 2023  106   1      6      0    27       G27
#> 7   7 20YWA2    MINOT 2023  107   1      7      0   169      G169
#> 8   8 20YWA2    MINOT 2023  108   1      8      0   176      G176
#> 9   9 20YWA2    MINOT 2023  109   1      9      0   188      G188
#> 10 10 20YWA2    MINOT 2023  110   1     10      0    18       G18
#> 11 11 20YWA2    MINOT 2023  111   1     11      0    84       G84
#> 12 12 20YWA2    MINOT 2023  112   1     12      0    34       G34