Generates an Spatial Un-replicated Optimized Arrangement Design

Randomly generates a spatial un-replicated optimized arrangement design, where the distance between checks is maximized in such a way that each row and column have control plots. Note that design generation needs the dimension of the field (number of rows and columns).

Usage

optimized_arrangement(
  nrows = NULL,
  ncols = NULL,
  lines = NULL,
  amountChecks = NULL,
  checks = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  seed = NULL,
  exptName = NULL,
  locationNames = NULL,
  optim = TRUE,
  data = NULL
)

Arguments

nrows

Number of rows in the field.

ncols

Number of columns in the field.

lines

Number of genotypes, experimental lines or treatments.

amountChecks

Integer with the amount total of checks or a numeric vector with the replicates of each check label.

checks

Number of genotypes as checks.

planter

Option for serpentine or cartesian arrangement. By default planter = 'serpentine'.

l

Number of locations. By default l = 1.

plotNumber

Numeric vector with the starting plot number for each location. By default plotNumber = 101.

seed

(optional) Real number that specifies the starting seed to obtain reproducible designs.

exptName

(optional) Name of the experiment.

locationNames

(optional) Name for each location.

optim

By default optim = TRUE.

data

(optional) Data frame with 3 columns: ENTRY | NAME | REPS.

Value

A list with five elements.

infoDesign is a list with information on the design parameters.

layoutRandom is a matrix with the randomization layout.

plotNumber is a matrix with the layout plot number.

dataEntry is a data frame with the data input.

genEntries is a list with the entries for replicated and no replicated part.

fieldBook is a data frame with field book design. This includes the index (Row, Column).

References

Clarke, G. P. Y., & Stefanova, K. T. (2011). Optimal design for early-generation plant breeding trials with unreplicated or partially replicated test lines. Australian & New Zealand Journal of Statistics, 53(4), 461–480.

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a spatial unreplicated optimized arrangement design in one location
# with 362 genotypes + 38 check plots (5 checks) for a field with dimension 20 rows x 20 cols.
OptimAd1 <- optimized_arrangement(nrows = 20, ncols = 20, lines = 362, 
                                  amountChecks = 38, 
                                  checks = 1:5,
                                  planter = "cartesian", 
                                  plotNumber = 101,
                                  seed = 14,
                                  exptName = "20RW1",
                                  locationNames = "CASSELTON")
OptimAd1$infoDesign
#> $rows
#> [1] 20
#> 
#> $columns
#> [1] 20
#> 
#> $treatments
#> [1] 362
#> 
#> $checks
#> [1] 5
#> 
#> $entry_checks
#> [1] 1 2 3 4 5
#> 
#> $rep_checks
#> [1]  7  7 10  7  7
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "cartesian"
#> 
#> $seed
#> [1] 14
#> 
#> $id_design
#> [1] 16
#> 
OptimAd1$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  140  173  287  306  161  327    5  352   93    12     2   106   206
#> Row19  143   82   31   98   65   49  349  136  367   103   209   260    60
#> Row18   57  340  309  229   58  110   35  202   15   134   310   346   207
#> Row17   72  360   30    2  114  213   91  252  180   342   357    14   291
#> Row16  326  296   90  178   83  344  293   73    4   116   215     4    97
#> Row15   32  285   87  167    1   67   18  321   21   338   183   245    52
#> Row14    4  141   70  308    2  162  271  250  364   228   219    25   100
#> Row13  118  353  241  227  189  190  214  297   66     2   254    99   262
#> Row12   37  355  157  240  146  154  105   84  216   232   263   235   121
#> Row11  191   94  301  335   23  284  366   55  270    68   238   320    63
#> Row10  348   95  345    2  132   47  135  204  171   292   126   239   187
#> Row9   299  303  257   48  127   54   69  220  217   312   122     5   268
#> Row8    61   71  365  275  156  267  195  307    2    59   133   148   281
#> Row7   282   88    3  170   26  218  223  125  164    86    79   272   120
#> Row6   153    5  212   44  230   11  304   80  341   274   288   273   249
#> Row5   198  316    3  295  192  276  169  298  111   185     1    33   336
#> Row4    46  151  351  331   96    4  211    4   27    29    41    39   201
#> Row3   203  359  179  142  165    4  333    7  337   231   124   112     3
#> Row2     5  317   81  222  318  115  280  221  158    36   305   358     3
#> Row1   123  236  313  139  152  108    1    2  199    75    64   155   109
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row20   163   226   205   144   363   246   160
#> Row19     1   328   194   294     3   264   248
#> Row18    40     3   138   196   278     1   117
#> Row17   283    77   330     5   324   300    92
#> Row16   168   107   332   347    45   325     8
#> Row15     4    38   130    56   184   266    53
#> Row14    20    42   210   237   224   233   119
#> Row13   159   314   319   277    43   137   193
#> Row12   259    50   172   197   361     3     3
#> Row11   269     3   290     5   302   289   175
#> Row10    34   188     1   251   247    85     6
#> Row9    101   129     1    13    16   354   253
#> Row8    315   131   265   208     3    22    78
#> Row7     19   102   128   243   147   244     5
#> Row6     51   166   350   311   322   279   329
#> Row5    145    76   255   181   150    89   234
#> Row4    113   174   261    74   104   176   286
#> Row3    149   186   339    24   182   256   362
#> Row2    258   225   334   323   177    10    62
#> Row1    242     9    28   200   356    17   343
#> 
OptimAd1$plotNumber
#> [[1]]
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  481  482  483  484  485  486  487  488  489   490   491   492   493
#>  [2,]  461  462  463  464  465  466  467  468  469   470   471   472   473
#>  [3,]  441  442  443  444  445  446  447  448  449   450   451   452   453
#>  [4,]  421  422  423  424  425  426  427  428  429   430   431   432   433
#>  [5,]  401  402  403  404  405  406  407  408  409   410   411   412   413
#>  [6,]  381  382  383  384  385  386  387  388  389   390   391   392   393
#>  [7,]  361  362  363  364  365  366  367  368  369   370   371   372   373
#>  [8,]  341  342  343  344  345  346  347  348  349   350   351   352   353
#>  [9,]  321  322  323  324  325  326  327  328  329   330   331   332   333
#> [10,]  301  302  303  304  305  306  307  308  309   310   311   312   313
#> [11,]  281  282  283  284  285  286  287  288  289   290   291   292   293
#> [12,]  261  262  263  264  265  266  267  268  269   270   271   272   273
#> [13,]  241  242  243  244  245  246  247  248  249   250   251   252   253
#> [14,]  221  222  223  224  225  226  227  228  229   230   231   232   233
#> [15,]  201  202  203  204  205  206  207  208  209   210   211   212   213
#> [16,]  181  182  183  184  185  186  187  188  189   190   191   192   193
#> [17,]  161  162  163  164  165  166  167  168  169   170   171   172   173
#> [18,]  141  142  143  144  145  146  147  148  149   150   151   152   153
#> [19,]  121  122  123  124  125  126  127  128  129   130   131   132   133
#> [20,]  101  102  103  104  105  106  107  108  109   110   111   112   113
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   494   495   496   497   498   499   500
#>  [2,]   474   475   476   477   478   479   480
#>  [3,]   454   455   456   457   458   459   460
#>  [4,]   434   435   436   437   438   439   440
#>  [5,]   414   415   416   417   418   419   420
#>  [6,]   394   395   396   397   398   399   400
#>  [7,]   374   375   376   377   378   379   380
#>  [8,]   354   355   356   357   358   359   360
#>  [9,]   334   335   336   337   338   339   340
#> [10,]   314   315   316   317   318   319   320
#> [11,]   294   295   296   297   298   299   300
#> [12,]   274   275   276   277   278   279   280
#> [13,]   254   255   256   257   258   259   260
#> [14,]   234   235   236   237   238   239   240
#> [15,]   214   215   216   217   218   219   220
#> [16,]   194   195   196   197   198   199   200
#> [17,]   174   175   176   177   178   179   180
#> [18,]   154   155   156   157   158   159   160
#> [19,]   134   135   136   137   138   139   140
#> [20,]   114   115   116   117   118   119   120
#> 
head(OptimAd1$fieldBook,12)
#>    ID  EXPT  LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20RW1 CASSELTON 2022  101   1      1      0   123      G123
#> 2   2 20RW1 CASSELTON 2022  102   1      2      0   236      G236
#> 3   3 20RW1 CASSELTON 2022  103   1      3      0   313      G313
#> 4   4 20RW1 CASSELTON 2022  104   1      4      0   139      G139
#> 5   5 20RW1 CASSELTON 2022  105   1      5      0   152      G152
#> 6   6 20RW1 CASSELTON 2022  106   1      6      0   108      G108
#> 7   7 20RW1 CASSELTON 2022  107   1      7      1     1       CH1
#> 8   8 20RW1 CASSELTON 2022  108   1      8      2     2       CH2
#> 9   9 20RW1 CASSELTON 2022  109   1      9      0   199      G199
#> 10 10 20RW1 CASSELTON 2022  110   1     10      0    75       G75
#> 11 11 20RW1 CASSELTON 2022  111   1     11      0    64       G64
#> 12 12 20RW1 CASSELTON 2022  112   1     12      0   155      G155
                  
# Example 2: Generates a spatial unreplicated optimized arrangement design in one location
# with 635 genotypes + 65 check plots (4 checks) for a field with dimension 20 rows x 35 cols.
# As example, we set up the data option with the entries list.
checks <- 4
list_checks <- paste("CH", 1:checks, sep = "")
treatments <- paste("G", 5:639, sep = "")
REPS <- c(17, 16, 16, 16, rep(1, 635))
treatment_list <- data.frame(list(ENTRY = 1:639, NAME = c(list_checks, treatments), REPS = REPS))
head(treatment_list, 12) 
#>    ENTRY NAME REPS
#> 1      1  CH1   17
#> 2      2  CH2   16
#> 3      3  CH3   16
#> 4      4  CH4   16
#> 5      5   G5    1
#> 6      6   G6    1
#> 7      7   G7    1
#> 8      8   G8    1
#> 9      9   G9    1
#> 10    10  G10    1
#> 11    11  G11    1
#> 12    12  G12    1
tail(treatment_list, 12)
#>     ENTRY NAME REPS
#> 628   628 G628    1
#> 629   629 G629    1
#> 630   630 G630    1
#> 631   631 G631    1
#> 632   632 G632    1
#> 633   633 G633    1
#> 634   634 G634    1
#> 635   635 G635    1
#> 636   636 G636    1
#> 637   637 G637    1
#> 638   638 G638    1
#> 639   639 G639    1
OptimAd2 <- optimized_arrangement(nrows = 20, ncols = 35, 
                                  planter = "serpentine", 
                                  plotNumber = 101,
                                  seed = 12,
                                  exptName = "20YWA2",
                                  locationNames = "MINOT",
                                  data = treatment_list)
OptimAd2$infoDesign
#> $rows
#> [1] 20
#> 
#> $columns
#> [1] 35
#> 
#> $treatments
#> [1] 635
#> 
#> $checks
#> [1] 4
#> 
#> $entry_checks
#> [1] 1 2 3 4
#> 
#> $rep_checks
#> [1] 17 16 16 16
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $seed
#> [1] 12
#> 
#> $id_design
#> [1] 16
#> 
OptimAd2$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row20  600  462  534    1  252  597  206   95  119   525    10    16   475
#> Row19  127   62  155  125  144  576  218    2    7   184   614   283   412
#> Row18  115  594    2  101  271   34  456   97   73   511   441   299   618
#> Row17  181  536  394  281    4  418  240  520  293   435   284   565   266
#> Row16   31  319  231   96  357  365  263   66  500   154     4   490   432
#> Row15  457  602  329  330  138  269  550   55  167   577   529   544    98
#> Row14  259   49   94  162   82    2  196    4  608     1   546   487   570
#> Row13    3  421  553  203  314  109  431  141  208   455   157     3    67
#> Row12  147  439  516  499  416  542  273   14  343   320   541     3   205
#> Row11   17  199  379    4  505  268  323  403  137   509   148   274   574
#> Row10  592  247  450  411  408   77    2  120   75   633   178   398     3
#> Row9   420  562  111  513   18  356  311  159   68     3    58   143     1
#> Row8     3   32   88  228  588  262  280  593  364   568   209   306   130
#> Row7   285  237    3  410  583  326   51  526  433   397    91   286   504
#> Row6   603    1  160   38  202   93  470  636  260   132     4   270   222
#> Row5    61  324   89  378  219    4  282  360    1   183   355   419   503
#> Row4   142    5  438  102  427  352  579  632  150    40   171   113   340
#> Row3   527  303  114  226   60  288    2  238  185   335   207   214   533
#> Row2   494  383  530  473  331  257  547   87  310   476    79   265   100
#> Row1   609  613  391  250   36  368  425  387  517   140   189    22   174
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20 Col21 Col22 Col23 Col24 Col25
#> Row20   152   172   195   307    42   333   290   201     2   187   463   396
#> Row19    90   611   639    44     3   117     3   405   386   213    92    46
#> Row18     4   116   512   629   446   135   465   429   413   223   472   124
#> Row17    11   555   545   193   424   393   558   596   342   619     3     1
#> Row16   566   318   604    99     4   177    59    72    15   506    29   501
#> Row15   277    63     2     1   467   423   123   198   543    43   392   339
#> Row14   186   110   315   447   166   300   585   573   353   474   449    69
#> Row13   328   514   358   191   371   606   385    19   377   468   480   362
#> Row12   105   133   220   508   399     8   591     1   366    84     1   466
#> Row11   363     1   498   477   151   149     3   163   188    57   235   179
#> Row10   128   451   336   369   146   471   507   537   175   469   107   254
#> Row9      1    70   616   417   560    47   272   389   532   540   492   638
#> Row8    121   249     4   607   232   161   103   561   129     2   212   539
#> Row7    489   559   194     1   348    20   190     2   502   510   192   458
#> Row6    244   245    81   126    23   522   346   344   309   156   521    37
#> Row5    317   267   234   233   347     2    26   563   582    27   586   287
#> Row4    305   197   158   428    56    39   297   538    41   345   444   239
#> Row3    380   453   571   229   327     2   599   302   334   414   304   349
#> Row2     25   460   164   617   631   434   590   182   204   401   215     2
#> Row1    276     3   404   301    28    21    78   275     4     3   294   134
#>       Col26 Col27 Col28 Col29 Col30 Col31 Col32 Col33 Col34 Col35
#> Row20    33    86   136    45   452   557    13   519     2   325
#> Row19   382   426   258   486   445   289    35   236   200   242
#> Row18   296   122   261   448   551   454     4   372   361   165
#> Row17   264   322   464   549   180   605   217   224   106   241
#> Row16     2   552   313   461   216     3   153    48   478   139
#> Row15   170    50   531   626   589     4   230   248   459   388
#> Row14   253   112   564     1     6   359    64   587   575   350
#> Row13   211   312     1   572    71   567   227   221   518   295
#> Row12   407   548    74    54   485   630   351   628   610   409
#> Row11   375   595    12   524     1    53   493    83   436   298
#> Row10   479   491   515   634   376   637   145     4   108   354
#> Row9    440   173   332   437   556   279   367   443    52   225
#> Row8    390   483   554   243   622   422   321   578   406     2
#> Row7     24    80    85   624   370   623   497   635   255   495
#> Row6    338   381   278   528     4   612   395   584   251   625
#> Row5    523   430   118   256   337   316   292   169   374   620
#> Row4      1   569   627     4   484     9   168   384   400     2
#> Row3    373     1   481   598   131   291   402    76     3   415
#> Row2     65     3   246   580   442   176     4   496   308   210
#> Row1    104   488    30   581   601   621   341   535   615   482
#> 
OptimAd2$plotNumber
#> [[1]]
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  800  799  798  797  796  795  794  793  792   791   790   789   788
#>  [2,]  731  732  733  734  735  736  737  738  739   740   741   742   743
#>  [3,]  730  729  728  727  726  725  724  723  722   721   720   719   718
#>  [4,]  661  662  663  664  665  666  667  668  669   670   671   672   673
#>  [5,]  660  659  658  657  656  655  654  653  652   651   650   649   648
#>  [6,]  591  592  593  594  595  596  597  598  599   600   601   602   603
#>  [7,]  590  589  588  587  586  585  584  583  582   581   580   579   578
#>  [8,]  521  522  523  524  525  526  527  528  529   530   531   532   533
#>  [9,]  520  519  518  517  516  515  514  513  512   511   510   509   508
#> [10,]  451  452  453  454  455  456  457  458  459   460   461   462   463
#> [11,]  450  449  448  447  446  445  444  443  442   441   440   439   438
#> [12,]  381  382  383  384  385  386  387  388  389   390   391   392   393
#> [13,]  380  379  378  377  376  375  374  373  372   371   370   369   368
#> [14,]  311  312  313  314  315  316  317  318  319   320   321   322   323
#> [15,]  310  309  308  307  306  305  304  303  302   301   300   299   298
#> [16,]  241  242  243  244  245  246  247  248  249   250   251   252   253
#> [17,]  240  239  238  237  236  235  234  233  232   231   230   229   228
#> [18,]  171  172  173  174  175  176  177  178  179   180   181   182   183
#> [19,]  170  169  168  167  166  165  164  163  162   161   160   159   158
#> [20,]  101  102  103  104  105  106  107  108  109   110   111   112   113
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>  [1,]   787   786   785   784   783   782   781   780   779   778   777   776
#>  [2,]   744   745   746   747   748   749   750   751   752   753   754   755
#>  [3,]   717   716   715   714   713   712   711   710   709   708   707   706
#>  [4,]   674   675   676   677   678   679   680   681   682   683   684   685
#>  [5,]   647   646   645   644   643   642   641   640   639   638   637   636
#>  [6,]   604   605   606   607   608   609   610   611   612   613   614   615
#>  [7,]   577   576   575   574   573   572   571   570   569   568   567   566
#>  [8,]   534   535   536   537   538   539   540   541   542   543   544   545
#>  [9,]   507   506   505   504   503   502   501   500   499   498   497   496
#> [10,]   464   465   466   467   468   469   470   471   472   473   474   475
#> [11,]   437   436   435   434   433   432   431   430   429   428   427   426
#> [12,]   394   395   396   397   398   399   400   401   402   403   404   405
#> [13,]   367   366   365   364   363   362   361   360   359   358   357   356
#> [14,]   324   325   326   327   328   329   330   331   332   333   334   335
#> [15,]   297   296   295   294   293   292   291   290   289   288   287   286
#> [16,]   254   255   256   257   258   259   260   261   262   263   264   265
#> [17,]   227   226   225   224   223   222   221   220   219   218   217   216
#> [18,]   184   185   186   187   188   189   190   191   192   193   194   195
#> [19,]   157   156   155   154   153   152   151   150   149   148   147   146
#> [20,]   114   115   116   117   118   119   120   121   122   123   124   125
#>       [,26] [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35]
#>  [1,]   775   774   773   772   771   770   769   768   767   766
#>  [2,]   756   757   758   759   760   761   762   763   764   765
#>  [3,]   705   704   703   702   701   700   699   698   697   696
#>  [4,]   686   687   688   689   690   691   692   693   694   695
#>  [5,]   635   634   633   632   631   630   629   628   627   626
#>  [6,]   616   617   618   619   620   621   622   623   624   625
#>  [7,]   565   564   563   562   561   560   559   558   557   556
#>  [8,]   546   547   548   549   550   551   552   553   554   555
#>  [9,]   495   494   493   492   491   490   489   488   487   486
#> [10,]   476   477   478   479   480   481   482   483   484   485
#> [11,]   425   424   423   422   421   420   419   418   417   416
#> [12,]   406   407   408   409   410   411   412   413   414   415
#> [13,]   355   354   353   352   351   350   349   348   347   346
#> [14,]   336   337   338   339   340   341   342   343   344   345
#> [15,]   285   284   283   282   281   280   279   278   277   276
#> [16,]   266   267   268   269   270   271   272   273   274   275
#> [17,]   215   214   213   212   211   210   209   208   207   206
#> [18,]   196   197   198   199   200   201   202   203   204   205
#> [19,]   145   144   143   142   141   140   139   138   137   136
#> [20,]   126   127   128   129   130   131   132   133   134   135
#> 
head(OptimAd2$fieldBook,12)
#>    ID   EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20YWA2    MINOT 2022  101   1      1      0   609      G609
#> 2   2 20YWA2    MINOT 2022  102   1      2      0   613      G613
#> 3   3 20YWA2    MINOT 2022  103   1      3      0   391      G391
#> 4   4 20YWA2    MINOT 2022  104   1      4      0   250      G250
#> 5   5 20YWA2    MINOT 2022  105   1      5      0    36       G36
#> 6   6 20YWA2    MINOT 2022  106   1      6      0   368      G368
#> 7   7 20YWA2    MINOT 2022  107   1      7      0   425      G425
#> 8   8 20YWA2    MINOT 2022  108   1      8      0   387      G387
#> 9   9 20YWA2    MINOT 2022  109   1      9      0   517      G517
#> 10 10 20YWA2    MINOT 2022  110   1     10      0   140      G140
#> 11 11 20YWA2    MINOT 2022  111   1     11      0   189      G189
#> 12 12 20YWA2    MINOT 2022  112   1     12      0    22       G22