Generates a Spatial Partially Replicated Arrangement Design

Randomly generates a spatial partially replicated design, where the distance between checks is maximized in such a way that each row and column have control plots. Note that design generation needs the dimension of the field (number of rows and columns).

partially_replicated(
  nrows = NULL,
  ncols = NULL,
  repGens = NULL,
  repUnits = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  seed = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows
Number of rows field.

ncols
Number of columns field.

repGens
Numeric vector with the amount genotypes to replicate.

repUnits
Numeric vector with the number of reps of each genotype.

planter
Option for serpentine or cartesian movement. By default planter = 'serpentine'.

l
Number of locations. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

exptName
(optional) Name of the experiment.

locationNames
(optional) Name for each location.

data
(optional) Dataframe with 3 columns: ENTRY | NAME | REPS.

Value

A list with information on the design parameters.

A matrix with the randomization layout.

A matrix with the layout plot number.

A data frame with fieldBook design. This includes the index (Row, Column).

References

Cullis, S., B. R., & Coombes, N. E. (2006). On the design of early generation variety trials with correlated data. Journal of Agricultural, Biological, and Environmental Statistics, 11, 381–393. https://doi.org/10.1198/108571106X154443

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a spatial optimized partially replicated arrangement design in one 
# location with 342 genotypes for a field with dimensions 25 rows x 18 cols. 
# Note that there are 280 genotypes unreplicated (only one time), 50 genotypes replicated 
# two times, and 10 genotypes replicated three times, and two checks 20 times each one.
SpatpREP1 <- partially_replicated(nrows = 25, 
                                  ncols = 18,  
                                  repGens = c(280,50,10,1,1),
                                  repUnits = c(1,2,3,20,20),
                                  planter = "cartesian", 
                                  plotNumber = 101,
                                  seed = 77)
SpatpREP1$infoDesign

#> $Greps
#> [1] 62
#> 
#> $RepGens
#>  [1] 20 20  3  3  3  3  3  3  3  3  3  3  2  2  2  2  2  2  2  2  2  2  2  2  2
#> [26]  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2  2
#> [51]  2  2  2  2  2  2  2  2  2  2  2  2
#> 
#> $EntryReps
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> [51] 51 52 53 54 55 56 57 58 59 60 61 62
#> 

SpatpREP1$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row25  155   15   11  249  274   56  339  289   34   168     3     9   176
#> Row24    9  271   29   28  251  177  317    1  282   253    23   338   178
#> Row23   17  134   43  214  318  323  305  333  329     2   185     2     2
#> Row22  132   56  272   59  108  309   83  216   68    59   105   233   261
#> Row21    1  215   52  334   45  161  101   10  264     2   294   208    30
#> Row20    7   30  260  191  210    2  321  268   31   225   133    81    33
#> Row19  231  239    3    5    9   22   38  109    2    86   332    77   129
#> Row18  303   57   76  241  206  218    8  328   19    41    67    80   336
#> Row17  320  154    2  166  252  267  137  118  242   213   142   146   135
#> Row16  299   36    5  255    7  254   36  122  326   110    70   209    42
#> Row15   48  262  295  188  322   69  220  102  217   229    50     1   307
#> Row14    1   14   31   26  331  245    2  232   39   284    20    74     2
#> Row13   78  169  223  125   25  117   53    2    2    53     1     2   152
#> Row12    1  314  256   54  167    1  165  270  281   162   153     1    17
#> Row11   90    1   89  330  115    1  159   27    4    84   179     5    85
#> Row10  189  199   99   73   46   60  144   71  292    55     1    88    57
#> Row9   130  187  139   58   51   91    1  186  126    21    72    62    50
#> Row8   175  311  247   13  300  269   66   60  131   306    13   337     1
#> Row7   259  182   12   48  298   61   16  288  200    43   160    54   228
#> Row6   193  141   95  192   62  204  194   92   40   230   123   138   243
#> Row5   236   20  341  148  257  234   11   15  136   297    26   149   273
#> Row4    24  266  151  219   12   39   47  174  120   202    23    94   280
#> Row3    58   41  181    1  150  114  173   47   45    16   157   258   226
#> Row2   296   61  340    6   25   38  107    1  286   119   127    21    87
#> Row1    79  308   49  147    2  211  112    2   28    10   196   212   302
#>       Col14 Col15 Col16 Col17 Col18
#> Row25   315   325   244     2   100
#> Row24    35     1    46     6    65
#> Row23   235     8    82    18   316
#> Row22    42     3   277   111   207
#> Row21   312   128    35   285    44
#> Row20   227   335    12    98   246
#> Row19   342     1   197    97   324
#> Row18   221     2   319     2    32
#> Row17   279   250    75   237     8
#> Row16   170    96    11     2   203
#> Row15   327   222    14   171    63
#> Row14     7   301   183   310   143
#> Row13   116    93   304   291   124
#> Row12    51   290   205   195    22
#> Row11    33    19   265   163   158
#> Row10    55   184   240     4    37
#> Row9     27   104   248    24   106
#> Row8    313     2   263   224   190
#> Row7    121   164    37    64   287
#> Row6     18    49    29     2   145
#> Row5     32   238    10   103     1
#> Row4    198    40   172   276    52
#> Row3    278   140    34     4   113
#> Row2    156   201   283     1     1
#> Row1      6   293   180   275    44

SpatpREP1$plotNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row25  533  534  535  536  537  538  539  540  541   542   543   544   545
#> Row24  515  516  517  518  519  520  521  522  523   524   525   526   527
#> Row23  497  498  499  500  501  502  503  504  505   506   507   508   509
#> Row22  479  480  481  482  483  484  485  486  487   488   489   490   491
#> Row21  461  462  463  464  465  466  467  468  469   470   471   472   473
#> Row20  443  444  445  446  447  448  449  450  451   452   453   454   455
#> Row19  425  426  427  428  429  430  431  432  433   434   435   436   437
#> Row18  407  408  409  410  411  412  413  414  415   416   417   418   419
#> Row17  389  390  391  392  393  394  395  396  397   398   399   400   401
#> Row16  371  372  373  374  375  376  377  378  379   380   381   382   383
#> Row15  353  354  355  356  357  358  359  360  361   362   363   364   365
#> Row14  335  336  337  338  339  340  341  342  343   344   345   346   347
#> Row13  317  318  319  320  321  322  323  324  325   326   327   328   329
#> Row12  299  300  301  302  303  304  305  306  307   308   309   310   311
#> Row11  281  282  283  284  285  286  287  288  289   290   291   292   293
#> Row10  263  264  265  266  267  268  269  270  271   272   273   274   275
#> Row9   245  246  247  248  249  250  251  252  253   254   255   256   257
#> Row8   227  228  229  230  231  232  233  234  235   236   237   238   239
#> Row7   209  210  211  212  213  214  215  216  217   218   219   220   221
#> Row6   191  192  193  194  195  196  197  198  199   200   201   202   203
#> Row5   173  174  175  176  177  178  179  180  181   182   183   184   185
#> Row4   155  156  157  158  159  160  161  162  163   164   165   166   167
#> Row3   137  138  139  140  141  142  143  144  145   146   147   148   149
#> Row2   119  120  121  122  123  124  125  126  127   128   129   130   131
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18
#> Row25   546   547   548   549   550
#> Row24   528   529   530   531   532
#> Row23   510   511   512   513   514
#> Row22   492   493   494   495   496
#> Row21   474   475   476   477   478
#> Row20   456   457   458   459   460
#> Row19   438   439   440   441   442
#> Row18   420   421   422   423   424
#> Row17   402   403   404   405   406
#> Row16   384   385   386   387   388
#> Row15   366   367   368   369   370
#> Row14   348   349   350   351   352
#> Row13   330   331   332   333   334
#> Row12   312   313   314   315   316
#> Row11   294   295   296   297   298
#> Row10   276   277   278   279   280
#> Row9    258   259   260   261   262
#> Row8    240   241   242   243   244
#> Row7    222   223   224   225   226
#> Row6    204   205   206   207   208
#> Row5    186   187   188   189   190
#> Row4    168   169   170   171   172
#> Row3    150   151   152   153   154
#> Row2    132   133   134   135   136
#> Row1    114   115   116   117   118

head(SpatpREP1$fieldBook,12)

#>    ID       EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20ExptpREP        1 2021  101   1      1      0    79       G79
#> 2   2 20ExptpREP        1 2021  102   1      2      0   308      G308
#> 3   3 20ExptpREP        1 2021  103   1      3      1    49       G49
#> 5   4 20ExptpREP        1 2021  104   1      4      0   147      G147
#> 6   5 20ExptpREP        1 2021  105   1      5      1     2        G2
#> 26  6 20ExptpREP        1 2021  106   1      6      0   211      G211
#> 27  7 20ExptpREP        1 2021  107   1      7      0   112      G112
#> 12  8 20ExptpREP        1 2021  108   1      8      1     2        G2
#> 28  9 20ExptpREP        1 2021  109   1      9      1    28       G28
#> 32 10 20ExptpREP        1 2021  110   1     10      1    10       G10
#> 33 11 20ExptpREP        1 2021  111   1     11      0   196      G196
#> 34 12 20ExptpREP        1 2021  112   1     12      0   212      G212

# Example 2: Generates a spatial optimized partially replicated arrangement design with 492 
# genotypes in a field with dimensions 30 rows x 20 cols. Note that there 384 genotypes 
# unreplicated (only one time), 108 genotypes replicated two times. 
# In this case we don't have check plots.
# As example, we set up the data option with the entries list.
NAME <- paste("G", 1:492, sep = "")
repGens = c(108, 384);repUnits = c(2,1)
REPS <- rep(repUnits, repGens)
treatment_list <- data.frame(list(ENTRY = 1:492, NAME = NAME, REPS = REPS))
head(treatment_list, 12) 

#>    ENTRY NAME REPS
#> 1      1   G1    2
#> 2      2   G2    2
#> 3      3   G3    2
#> 4      4   G4    2
#> 5      5   G5    2
#> 6      6   G6    2
#> 7      7   G7    2
#> 8      8   G8    2
#> 9      9   G9    2
#> 10    10  G10    2
#> 11    11  G11    2
#> 12    12  G12    2

tail(treatment_list, 12)

#>     ENTRY NAME REPS
#> 481   481 G481    1
#> 482   482 G482    1
#> 483   483 G483    1
#> 484   484 G484    1
#> 485   485 G485    1
#> 486   486 G486    1
#> 487   487 G487    1
#> 488   488 G488    1
#> 489   489 G489    1
#> 490   490 G490    1
#> 491   491 G491    1
#> 492   492 G492    1

SpatpREP2 <- partially_replicated(nrows = 30, 
                                  ncols = 20, 
                                  planter = "serpentine", 
                                  plotNumber = 101,
                                  seed = 41,
                                  data = treatment_list)
SpatpREP2$infoDesign

#> $Greps
#> [1] 108
#> 
#> $RepGens
#>   [1] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [38] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [75] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> 
#> $EntryReps
#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> 

SpatpREP2$layoutRandom

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30   80  439   34  362   92   77   80  347  119   149    58   170   478
#> Row29   38  472   41  106  434   60   26  159  302   420   234   341    68
#> Row28  390   38   37  286  426  153  306   88  109   373    36   101   174
#> Row27  139    2   54  116  210  345  113  104  381    47   375   284   232
#> Row26  450  367   28  324   70  383   40   69  329   407   224    82    24
#> Row25  140  230  274  147  283  239   68   12  484   143    46   332    23
#> Row24   99  355  281  238  319   60  173   84  465   145   326   317   350
#> Row23   64  120  339  291   12  295  424  205   95    86   244   267   199
#> Row22  344   56  333   14  380   57  372  107   35   327   437   135   421
#> Row21   18  282  277   58  360  307  187  157  400   171   310     5   130
#> Row20  197   93   82    3  300  262   54   75    9   491   388   299   315
#> Row19  191  246   16  270  102   59  289  293  364     7   220   207    43
#> Row18   49   91  405    6  351   10  104  214  154    32   288   325   374
#> Row17  313  382   76   51  430  108  468  445  100    72    97    39   328
#> Row16  391  179  178   37   39   15  368   20  155    71    51    33   318
#> Row15  126  417  201  404   49    8   91  463   79   309    17   453   395
#> Row14  414  186  455   53   42  162  330  117  208   252   212   438   422
#> Row13  320  394  237  389  150  399   21  203   45   235   305    67   481
#> Row12  256  471  264  393  285  480   72  346  296     4    31   129    22
#> Row11  308   19  416  287  398  200  349  363    4   443   204    74   132
#> Row10  271  100   66  243  273   86  451  385  371   312    44   172   431
#> Row9    25  470  433  454  474   98  321   87  492   359   419   101   213
#> Row8   357   25  248  215  144  137   29  469  401   418   142    95    30
#> Row7    98   81  194  444  297  163   52   61  138    73   188   183    77
#> Row6   125  112  202  475   48  118  458   65   42   105   272   222   335
#> Row5   102  165    7   28  397  265  410  348   48   216   379   254    29
#> Row4   169  476  340   70  427  477  136  376  226    64   461   361    27
#> Row3    62  253  314  247  449  378  278  429   16    46    35    96   106
#> Row2   206   41    9   66   57  148  353  146  403   290    78   436    63
#> Row1     5   71  446  298   14  488  334  266  485   459    55   107   121
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row30   268    27   479   369   460   134   377
#> Row29   111   412   193   196   358   384    24
#> Row28   490    62   181   218   425    26   128
#> Row27   294   124   242   245    89   108   462
#> Row26    53   392   217    15   467   158   365
#> Row25    11   110   103   304    78   189    94
#> Row24    36   241   336   483   164   337    74
#> Row23    33   448   423   160     8   233   331
#> Row22     1    22   343    61   249   161   176
#> Row21   263    83   223   168     1   441     6
#> Row20   486   432    85   311   406   231   225
#> Row19   261    92    75    13    31   457   409
#> Row18   396   323    11     3   250    90   133
#> Row17    50   356   370    73   131   366   408
#> Row16   428   198   303   190   122   251   338
#> Row15    67   276   354   127    63    88   152
#> Row14   413    79   103   442   487    65   387
#> Row13   166    96    52   227   415   322    44
#> Row12   435   229   301    81   195    47    83
#> Row11    13   257   221   151   259    56    69
#> Row10    50    45     2   182    94   452   192
#> Row9     21   466    87    10    43    30   489
#> Row8    447   411   185    89   342   440   209
#> Row7    386    85   177   211    55   482   456
#> Row6    240   115    19   228   279   123    18
#> Row5    175   292   473    97    76   316   402
#> Row4     40    17   184    90   280    32    34
#> Row3    167   236    20   269   464   141   219
#> Row2    255   275   258   180   105    99    59
#> Row1    352    23   114    84   156    93   260

SpatpREP2$plotNumber

#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  700  699  698  697  696  695  694  693  692   691   690   689   688
#> Row29  661  662  663  664  665  666  667  668  669   670   671   672   673
#> Row28  660  659  658  657  656  655  654  653  652   651   650   649   648
#> Row27  621  622  623  624  625  626  627  628  629   630   631   632   633
#> Row26  620  619  618  617  616  615  614  613  612   611   610   609   608
#> Row25  581  582  583  584  585  586  587  588  589   590   591   592   593
#> Row24  580  579  578  577  576  575  574  573  572   571   570   569   568
#> Row23  541  542  543  544  545  546  547  548  549   550   551   552   553
#> Row22  540  539  538  537  536  535  534  533  532   531   530   529   528
#> Row21  501  502  503  504  505  506  507  508  509   510   511   512   513
#> Row20  500  499  498  497  496  495  494  493  492   491   490   489   488
#> Row19  461  462  463  464  465  466  467  468  469   470   471   472   473
#> Row18  460  459  458  457  456  455  454  453  452   451   450   449   448
#> Row17  421  422  423  424  425  426  427  428  429   430   431   432   433
#> Row16  420  419  418  417  416  415  414  413  412   411   410   409   408
#> Row15  381  382  383  384  385  386  387  388  389   390   391   392   393
#> Row14  380  379  378  377  376  375  374  373  372   371   370   369   368
#> Row13  341  342  343  344  345  346  347  348  349   350   351   352   353
#> Row12  340  339  338  337  336  335  334  333  332   331   330   329   328
#> Row11  301  302  303  304  305  306  307  308  309   310   311   312   313
#> Row10  300  299  298  297  296  295  294  293  292   291   290   289   288
#> Row9   261  262  263  264  265  266  267  268  269   270   271   272   273
#> Row8   260  259  258  257  256  255  254  253  252   251   250   249   248
#> Row7   221  222  223  224  225  226  227  228  229   230   231   232   233
#> Row6   220  219  218  217  216  215  214  213  212   211   210   209   208
#> Row5   181  182  183  184  185  186  187  188  189   190   191   192   193
#> Row4   180  179  178  177  176  175  174  173  172   171   170   169   168
#> Row3   141  142  143  144  145  146  147  148  149   150   151   152   153
#> Row2   140  139  138  137  136  135  134  133  132   131   130   129   128
#> Row1   101  102  103  104  105  106  107  108  109   110   111   112   113
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row30   687   686   685   684   683   682   681
#> Row29   674   675   676   677   678   679   680
#> Row28   647   646   645   644   643   642   641
#> Row27   634   635   636   637   638   639   640
#> Row26   607   606   605   604   603   602   601
#> Row25   594   595   596   597   598   599   600
#> Row24   567   566   565   564   563   562   561
#> Row23   554   555   556   557   558   559   560
#> Row22   527   526   525   524   523   522   521
#> Row21   514   515   516   517   518   519   520
#> Row20   487   486   485   484   483   482   481
#> Row19   474   475   476   477   478   479   480
#> Row18   447   446   445   444   443   442   441
#> Row17   434   435   436   437   438   439   440
#> Row16   407   406   405   404   403   402   401
#> Row15   394   395   396   397   398   399   400
#> Row14   367   366   365   364   363   362   361
#> Row13   354   355   356   357   358   359   360
#> Row12   327   326   325   324   323   322   321
#> Row11   314   315   316   317   318   319   320
#> Row10   287   286   285   284   283   282   281
#> Row9    274   275   276   277   278   279   280
#> Row8    247   246   245   244   243   242   241
#> Row7    234   235   236   237   238   239   240
#> Row6    207   206   205   204   203   202   201
#> Row5    194   195   196   197   198   199   200
#> Row4    167   166   165   164   163   162   161
#> Row3    154   155   156   157   158   159   160
#> Row2    127   126   125   124   123   122   121
#> Row1    114   115   116   117   118   119   120

head(SpatpREP2$fieldBook,10)

#>    ID       EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 20ExptpREP        1 2021  101   1      1      1     5        G5
#> 3   2 20ExptpREP        1 2021  102   1      2      1    71       G71
#> 5   3 20ExptpREP        1 2021  103   1      3      0   446      G446
#> 6   4 20ExptpREP        1 2021  104   1      4      0   298      G298
#> 7   5 20ExptpREP        1 2021  105   1      5      1    14       G14
#> 9   6 20ExptpREP        1 2021  106   1      6      0   488      G488
#> 10  7 20ExptpREP        1 2021  107   1      7      0   334      G334
#> 11  8 20ExptpREP        1 2021  108   1      8      0   266      G266
#> 12  9 20ExptpREP        1 2021  109   1      9      0   485      G485
#> 13 10 20ExptpREP        1 2021  110   1     10      0   459      G459