Generates a Spatial Partially Replicated Arrangement Design

Randomly generates a spatial partially replicated design, where the distance between checks is maximized in such a way that each row and column have control plots. Note that design generation needs the dimension of the field (number of rows and columns).

Usage

partially_replicated(
  nrows = NULL,
  ncols = NULL,
  repGens = NULL,
  repUnits = NULL,
  planter = "serpentine",
  l = 1,
  plotNumber = 101,
  seed = NULL,
  exptName = NULL,
  locationNames = NULL,
  data = NULL
)

Arguments

nrows

Number of rows field.

ncols

Number of columns field.

repGens

Numeric vector with the amount genotypes to replicate.

repUnits

Numeric vector with the number of reps of each genotype.

planter

Option for serpentine or cartesian movement. By default planter = 'serpentine'.

l

Number of locations. By default l = 1.

plotNumber

Numeric vector with the starting plot number for each location. By default plotNumber = 101.

seed

(optional) Real number that specifies the starting seed to obtain reproducible designs.

exptName

(optional) Name of the experiment.

locationNames

(optional) Name for each location.

data

(optional) Dataframe with 3 columns: ENTRY | NAME | REPS.

Value

A list with five elements.

infoDesign is a list with information on the design parameters.

layoutRandom is a matrix with the randomization layout.

plotNumber is a matrix with the layout plot number.

dataEntry is a data frame with the data input.

genEntries is a list with the entries for replicated and no replicated part.

fieldBook is a data frame with field book design. This includes the index (Row, Column).

References

Cullis, S., B. R., & Coombes, N. E. (2006). On the design of early generation variety trials with correlated data. Journal of Agricultural, Biological, and Environmental Statistics, 11, 381–393. https://doi.org/10.1198/108571106X154443

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a spatial optimized partially replicated arrangement design in one 
# location with 342 genotypes for a field with dimensions 25 rows x 18 cols. 
# Note that there are 280 genotypes unreplicated (only one time), 50 genotypes replicated 
# two times, and 10 genotypes replicated three times, and two checks 20 times each one.
SpatpREP1 <- partially_replicated(nrows = 25, 
                                  ncols = 18,  
                                  repGens = c(280,50,10,1,1),
                                  repUnits = c(1,2,3,20,20),
                                  planter = "cartesian", 
                                  plotNumber = 101,
                                  seed = 77)
SpatpREP1$infoDesign
#> $rows
#> [1] 25
#> 
#> $columns
#> [1] 18
#> 
#> $treatments_with_reps
#> [1] 62
#> 
#> $treatments_with_no_reps
#> [1] 280
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "cartesian"
#> 
#> $seed
#> [1] 77
#> 
#> $id_design
#> [1] 13
#> 
SpatpREP1$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row25  239   12   35   41  158  169   88  292   48   324    46     4   154
#> Row24   38  281   53  180  338    2  185   50  287   107     2   159   249
#> Row23   36   70   11  204  247    2  245  280   66    27   182     2     1
#> Row22  155   18    2    8   10  303  309   89  106    29    57    76   290
#> Row21  173  127   19  298  110  135  261   29  294    69   227    34    25
#> Row20  285    1  216  160  156   60   74   98    9    47    91   231     9
#> Row19  221   68   53  141   58   28   24   78   39    84    96   257   196
#> Row18   49   31  125  214  262  342    1   26   32    55   310   260   268
#> Row17    8  137   48   15  237  276  233  105  277   283   163   109   240
#> Row16  272   38   22  147   50  117    2  321  332   224   282   279    61
#> Row15    2    6  242   86    1  152  275  172   90    35    40    21   208
#> Row14   27    1  111   54  270  198   13   39  197   130   114   222    51
#> Row13  183   94  184  251   43  191   30   31   47   134    33     1   217
#> Row12    2  145  146   32   25    6  306  211   17   322    73     6    20
#> Row11  244   41  341  263  144   42  126    2  334   181   317     4   228
#> Row10  121   64  210  337   52  284  241  293  194     4    24   335    20
#> Row9   179  318  278  143  323  325    2  103  133    59   299    17     7
#> Row8   316   95   63   14  122  250   71   59   45   218    46   253    33
#> Row7   170  161  330   23  291   11    1   85  305    54   189    44   328
#> Row6    11  295  331  301   37   80  178  139   12   201   320   215   243
#> Row5   212    3  157  203  312  112   51  118  192    93     2   193   238
#> Row4    62   72  319  232  138    2   10  286  252   176    56   119   230
#> Row3   209  113  115    8   87    2  151   61   16    18   162   339   246
#> Row2   207   52  258   19   21    2  124    1  131   213   190    43   264
#> Row1   168  297    2  195  308  327  248    2    1     7   226   254    82
#>       Col14 Col15 Col16 Col17 Col18
#> Row25    99   289   315     1   148
#> Row24     1   116    60    23   288
#> Row23    67    62   229     7   205
#> Row22     2    45   314   104   101
#> Row21   307    16     1   255    56
#> Row20   199   136    58   187   220
#> Row19   336     1   219   256   123
#> Row18   167    42   128    44    30
#> Row17    75   274   102   108     1
#> Row16   259   200    49     1   165
#> Row15   304    83     1   186   271
#> Row14     2   329   188   296   273
#> Row13   223   166   120   206   140
#> Row12    26   164   100    92    55
#> Row11     1     3   311   177   149
#> Row10   234   267   236     5     9
#> Row9     15   340    79    28   150
#> Row8    266    13   171    77   129
#> Row7    265     5     2   326   132
#> Row6     22     1    36    10    97
#> Row5     37   153    40   225    14
#> Row4    269    12   174   202     1
#> Row3    333   175     1    34   302
#> Row2    142   300   313     3     5
#> Row1     57   235    81    65     2
#> 
SpatpREP1$plotNumber
#> [[1]]
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  533  534  535  536  537  538  539  540  541   542   543   544   545
#>  [2,]  515  516  517  518  519  520  521  522  523   524   525   526   527
#>  [3,]  497  498  499  500  501  502  503  504  505   506   507   508   509
#>  [4,]  479  480  481  482  483  484  485  486  487   488   489   490   491
#>  [5,]  461  462  463  464  465  466  467  468  469   470   471   472   473
#>  [6,]  443  444  445  446  447  448  449  450  451   452   453   454   455
#>  [7,]  425  426  427  428  429  430  431  432  433   434   435   436   437
#>  [8,]  407  408  409  410  411  412  413  414  415   416   417   418   419
#>  [9,]  389  390  391  392  393  394  395  396  397   398   399   400   401
#> [10,]  371  372  373  374  375  376  377  378  379   380   381   382   383
#> [11,]  353  354  355  356  357  358  359  360  361   362   363   364   365
#> [12,]  335  336  337  338  339  340  341  342  343   344   345   346   347
#> [13,]  317  318  319  320  321  322  323  324  325   326   327   328   329
#> [14,]  299  300  301  302  303  304  305  306  307   308   309   310   311
#> [15,]  281  282  283  284  285  286  287  288  289   290   291   292   293
#> [16,]  263  264  265  266  267  268  269  270  271   272   273   274   275
#> [17,]  245  246  247  248  249  250  251  252  253   254   255   256   257
#> [18,]  227  228  229  230  231  232  233  234  235   236   237   238   239
#> [19,]  209  210  211  212  213  214  215  216  217   218   219   220   221
#> [20,]  191  192  193  194  195  196  197  198  199   200   201   202   203
#> [21,]  173  174  175  176  177  178  179  180  181   182   183   184   185
#> [22,]  155  156  157  158  159  160  161  162  163   164   165   166   167
#> [23,]  137  138  139  140  141  142  143  144  145   146   147   148   149
#> [24,]  119  120  121  122  123  124  125  126  127   128   129   130   131
#> [25,]  101  102  103  104  105  106  107  108  109   110   111   112   113
#>       [,14] [,15] [,16] [,17] [,18]
#>  [1,]   546   547   548   549   550
#>  [2,]   528   529   530   531   532
#>  [3,]   510   511   512   513   514
#>  [4,]   492   493   494   495   496
#>  [5,]   474   475   476   477   478
#>  [6,]   456   457   458   459   460
#>  [7,]   438   439   440   441   442
#>  [8,]   420   421   422   423   424
#>  [9,]   402   403   404   405   406
#> [10,]   384   385   386   387   388
#> [11,]   366   367   368   369   370
#> [12,]   348   349   350   351   352
#> [13,]   330   331   332   333   334
#> [14,]   312   313   314   315   316
#> [15,]   294   295   296   297   298
#> [16,]   276   277   278   279   280
#> [17,]   258   259   260   261   262
#> [18,]   240   241   242   243   244
#> [19,]   222   223   224   225   226
#> [20,]   204   205   206   207   208
#> [21,]   186   187   188   189   190
#> [22,]   168   169   170   171   172
#> [23,]   150   151   152   153   154
#> [24,]   132   133   134   135   136
#> [25,]   114   115   116   117   118
#> 
head(SpatpREP1$fieldBook,12)
#>    ID  EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 Expt1        1 2022  101   1      1      0   168      G168
#> 2   2 Expt1        1 2022  102   1      2      0   297      G297
#> 3   3 Expt1        1 2022  103   1      3      2     2        G2
#> 4   4 Expt1        1 2022  104   1      4      0   195      G195
#> 5   5 Expt1        1 2022  105   1      5      0   308      G308
#> 6   6 Expt1        1 2022  106   1      6      0   327      G327
#> 7   7 Expt1        1 2022  107   1      7      0   248      G248
#> 8   8 Expt1        1 2022  108   1      8      2     2        G2
#> 9   9 Expt1        1 2022  109   1      9      1     1        G1
#> 10 10 Expt1        1 2022  110   1     10      7     7        G7
#> 11 11 Expt1        1 2022  111   1     11      0   226      G226
#> 12 12 Expt1        1 2022  112   1     12      0   254      G254

# Example 2: Generates a spatial optimized partially replicated arrangement design with 492 
# genotypes in a field with dimensions 30 rows x 20 cols. Note that there 384 genotypes 
# unreplicated (only one time), 108 genotypes replicated two times. 
# In this case we don't have check plots.
# As example, we set up the data option with the entries list.
NAME <- paste("G", 1:492, sep = "")
repGens = c(108, 384);repUnits = c(2,1)
REPS <- rep(repUnits, repGens)
treatment_list <- data.frame(list(ENTRY = 1:492, NAME = NAME, REPS = REPS))
head(treatment_list, 12) 
#>    ENTRY NAME REPS
#> 1      1   G1    2
#> 2      2   G2    2
#> 3      3   G3    2
#> 4      4   G4    2
#> 5      5   G5    2
#> 6      6   G6    2
#> 7      7   G7    2
#> 8      8   G8    2
#> 9      9   G9    2
#> 10    10  G10    2
#> 11    11  G11    2
#> 12    12  G12    2
tail(treatment_list, 12)
#>     ENTRY NAME REPS
#> 481   481 G481    1
#> 482   482 G482    1
#> 483   483 G483    1
#> 484   484 G484    1
#> 485   485 G485    1
#> 486   486 G486    1
#> 487   487 G487    1
#> 488   488 G488    1
#> 489   489 G489    1
#> 490   490 G490    1
#> 491   491 G491    1
#> 492   492 G492    1
SpatpREP2 <- partially_replicated(nrows = 30, 
                                  ncols = 20, 
                                  planter = "serpentine", 
                                  plotNumber = 101,
                                  seed = 41,
                                  data = treatment_list)
SpatpREP2$infoDesign
#> $rows
#> [1] 30
#> 
#> $columns
#> [1] 20
#> 
#> $treatments_with_reps
#> [1] 108
#> 
#> $treatments_with_no_reps
#> [1] 384
#> 
#> $locations
#> [1] 1
#> 
#> $planter
#> [1] "serpentine"
#> 
#> $seed
#> [1] 41
#> 
#> $id_design
#> [1] 13
#> 
SpatpREP2$layoutRandom
#> [[1]]
#>       Col1 Col2 Col3 Col4 Col5 Col6 Col7 Col8 Col9 Col10 Col11 Col12 Col13
#> Row30  345  178    8  158  211   89   70  236   83   342   292   234    42
#> Row29  286  131   75  104  228   87   77  249  395   308    59   313    52
#> Row28  420  471    6  129  213  402  253   61  260   426    47   392   133
#> Row27  195  146   10  338  433  215  413   16  317    14   394    77    33
#> Row26  287  360  107  454   32  453    5  477  489   149   166    36   369
#> Row25  285  365  169  180  314  290    7   86  460   172    56   153   105
#> Row24   18  456   37  188  305   31  238   40  363   208   459   349   274
#> Row23   62   63   88  490   55  385  210  252  440    92   357   174   488
#> Row22  396   66  318   74  421  108  204   91   32   299    60   128   320
#> Row21   80  254  157   51  321  461  400  344  467   275   309    87   209
#> Row20  428   58   67  106  103  409   76   48   45   377   201   462   110
#> Row19  447   84  173  156   69   62  343  182  388    93   370   330    43
#> Row18   49   20  221   56  256   23   64  304  258    19   416   227   271
#> Row17  387  482   98   48  379   38  176  109  122    50   481   346   220
#> Row16  261  183  291   35   65   89  245  101  198    99   102    17   378
#> Row15  130  478  418  415   88   37   73  410   74   232   427   381   233
#> Row14  443  121  217   68   82  430  123  452  288   114   207    22    85
#> Row13  393  140  386  289  417  257   38  425   44     7   331    65   137
#> Row12  102  163  187  422  355  150   14  262   45    42     3   431   340
#> Row11  438   55  139  356  272  277  127  177    9   480   406    81    22
#> Row10  371    4  293  278  362   96  332  203  325   397    54    52   212
#> Row9    64  105  270  205  190   80  383  106   85   403   148    49   411
#> Row8   251   41  337    2  191  284   82  366  484   239    46    34   136
#> Row7     2   19  359  224  323  374   96   25  269    72   380   441    95
#> Row6   429  222  125  483   28  263  449   97   13    79   265   326   243
#> Row5   101  246  432   35  465  486  218  466    6   175   145   200    92
#> Row4   179  120  165   63  113  341  241  458  126     3   398   404    50
#> Row3    61  391   59  141  470  384  455  298   11   117     9    18    73
#> Row2   124   98   90  138   66  352  242  439  434   333    68   487   267
#> Row1    24  247  336  368   26  339  350  367  297   189    12   348   181
#>       Col14 Col15 Col16 Col17 Col18 Col19 Col20
#> Row30   335    28   193   196   310    24   472
#> Row29   382   435   283   451   444   319    71
#> Row28   401    10    76    29   282    46   328
#> Row27    25   436   408   111    16    51   457
#> Row26   375     1   358    58   108   300   273
#> Row25    53   334    79   206    86   312    13
#> Row24   303   160   414    78   115   155    72
#> Row23    83   450   159   295   281   464   474
#> Row22    21    26   266   135   324   170   255
#> Row21   229    41   307   353    39   216    33
#> Row20   361   202    30   259   399   276   162
#> Row19   448    90    75   103   240   405   264
#> Row18   144   407    36   469   147   100   372
#> Row17    69   327   134    40     4   491   468
#> Row16   389   419   118   112   424   479   437
#> Row15    71   412   364   316    39   104   164
#> Row14   184   219    94   446   154    44   279
#> Row13    60    11    94   423   197   390    47
#> Row12   167   306   315    70   311    67    31
#> Row11    81   442   347   116   376    78    57
#> Row10    27    57    84   248     8   230   142
#> Row9    280   226   294   107    27    95   268
#> Row8    475   492   354    12   192   151   152
#> Row7    171    53   237   301    23   231   225
#> Row6    329   223    91   132   186   250     5
#> Row5    161   463    15    99    20   445   476
#> Row4     54   100   185    17   119    15    30
#> Row3    473   302    43   244   143   199   373
#> Row2    485   296   322   351    34    29    93
#> Row1     21     1   235    97   214   194   168
#> 
SpatpREP2$plotNumber
#> [[1]]
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]  700  699  698  697  696  695  694  693  692   691   690   689   688
#>  [2,]  661  662  663  664  665  666  667  668  669   670   671   672   673
#>  [3,]  660  659  658  657  656  655  654  653  652   651   650   649   648
#>  [4,]  621  622  623  624  625  626  627  628  629   630   631   632   633
#>  [5,]  620  619  618  617  616  615  614  613  612   611   610   609   608
#>  [6,]  581  582  583  584  585  586  587  588  589   590   591   592   593
#>  [7,]  580  579  578  577  576  575  574  573  572   571   570   569   568
#>  [8,]  541  542  543  544  545  546  547  548  549   550   551   552   553
#>  [9,]  540  539  538  537  536  535  534  533  532   531   530   529   528
#> [10,]  501  502  503  504  505  506  507  508  509   510   511   512   513
#> [11,]  500  499  498  497  496  495  494  493  492   491   490   489   488
#> [12,]  461  462  463  464  465  466  467  468  469   470   471   472   473
#> [13,]  460  459  458  457  456  455  454  453  452   451   450   449   448
#> [14,]  421  422  423  424  425  426  427  428  429   430   431   432   433
#> [15,]  420  419  418  417  416  415  414  413  412   411   410   409   408
#> [16,]  381  382  383  384  385  386  387  388  389   390   391   392   393
#> [17,]  380  379  378  377  376  375  374  373  372   371   370   369   368
#> [18,]  341  342  343  344  345  346  347  348  349   350   351   352   353
#> [19,]  340  339  338  337  336  335  334  333  332   331   330   329   328
#> [20,]  301  302  303  304  305  306  307  308  309   310   311   312   313
#> [21,]  300  299  298  297  296  295  294  293  292   291   290   289   288
#> [22,]  261  262  263  264  265  266  267  268  269   270   271   272   273
#> [23,]  260  259  258  257  256  255  254  253  252   251   250   249   248
#> [24,]  221  222  223  224  225  226  227  228  229   230   231   232   233
#> [25,]  220  219  218  217  216  215  214  213  212   211   210   209   208
#> [26,]  181  182  183  184  185  186  187  188  189   190   191   192   193
#> [27,]  180  179  178  177  176  175  174  173  172   171   170   169   168
#> [28,]  141  142  143  144  145  146  147  148  149   150   151   152   153
#> [29,]  140  139  138  137  136  135  134  133  132   131   130   129   128
#> [30,]  101  102  103  104  105  106  107  108  109   110   111   112   113
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   687   686   685   684   683   682   681
#>  [2,]   674   675   676   677   678   679   680
#>  [3,]   647   646   645   644   643   642   641
#>  [4,]   634   635   636   637   638   639   640
#>  [5,]   607   606   605   604   603   602   601
#>  [6,]   594   595   596   597   598   599   600
#>  [7,]   567   566   565   564   563   562   561
#>  [8,]   554   555   556   557   558   559   560
#>  [9,]   527   526   525   524   523   522   521
#> [10,]   514   515   516   517   518   519   520
#> [11,]   487   486   485   484   483   482   481
#> [12,]   474   475   476   477   478   479   480
#> [13,]   447   446   445   444   443   442   441
#> [14,]   434   435   436   437   438   439   440
#> [15,]   407   406   405   404   403   402   401
#> [16,]   394   395   396   397   398   399   400
#> [17,]   367   366   365   364   363   362   361
#> [18,]   354   355   356   357   358   359   360
#> [19,]   327   326   325   324   323   322   321
#> [20,]   314   315   316   317   318   319   320
#> [21,]   287   286   285   284   283   282   281
#> [22,]   274   275   276   277   278   279   280
#> [23,]   247   246   245   244   243   242   241
#> [24,]   234   235   236   237   238   239   240
#> [25,]   207   206   205   204   203   202   201
#> [26,]   194   195   196   197   198   199   200
#> [27,]   167   166   165   164   163   162   161
#> [28,]   154   155   156   157   158   159   160
#> [29,]   127   126   125   124   123   122   121
#> [30,]   114   115   116   117   118   119   120
#> 
head(SpatpREP2$fieldBook,10)
#>    ID  EXPT LOCATION YEAR PLOT ROW COLUMN CHECKS ENTRY TREATMENT
#> 1   1 Expt1        1 2022  101   1      1     24    24       G24
#> 2   2 Expt1        1 2022  102   1      2      0   247      G247
#> 3   3 Expt1        1 2022  103   1      3      0   336      G336
#> 4   4 Expt1        1 2022  104   1      4      0   368      G368
#> 5   5 Expt1        1 2022  105   1      5     26    26       G26
#> 6   6 Expt1        1 2022  106   1      6      0   339      G339
#> 7   7 Expt1        1 2022  107   1      7      0   350      G350
#> 8   8 Expt1        1 2022  108   1      8      0   367      G367
#> 9   9 Expt1        1 2022  109   1      9      0   297      G297
#> 10 10 Expt1        1 2022  110   1     10      0   189      G189