Generates a Resolvable Row-Column Design (RowColD)

It randomly generates a resolvable row-column design (RowColD). The design is optimized in both rows and columns blocking factors. The randomization can be done across multiple locations.

Usage

row_column(
  t = NULL,
  nrows = NULL,
  r = NULL,
  l = 1,
  plotNumber = 101,
  locationNames = NULL,
  seed = NULL,
  iterations = 1000,
  data = NULL
)

Arguments

t

Number of treatments.

nrows

Number of rows of a full resolvable replicate.

r

Number of blocks (full resolvable replicates).

l

Number of locations. By default l = 1.

plotNumber

Numeric vector with the starting plot number for each location. By default plotNumber = 101.

locationNames

(optional) Names for each location.

seed

(optional) Real number that specifies the starting seed to obtain reproducible designs.

iterations

Number of iterations for design optimization. By default iterations = 1000.

data

(optional) Data frame with label list of treatments

Value

A list with four elements.

infoDesign is a list with information on the design parameters.

resolvableBlocks a list with the resolvable row columns blocks.

concurrence is the concurrence matrix.

fieldBook is a data frame with the row-column field book.

Details

The Row-Column design in FielDHub is built in two stages. The first step constructs the blocking factor Columns using Incomplete Block Units from an incomplete block design that sets the number of incomplete blocks as the number of Columns in the design, each of which has a dimension equal to the number of Rows. Once this design is generated, the Rows are used as the Row blocking factor that is optimized for A-Efficiency, but levels within the original Columns are fixed. To optimize the Rows while maintaining the current optimized Columns, we use a heuristic algorithm that swaps at random treatment positions within a given Column (Block) also selected at random. The algorithm begins by calculating the A-Efficiency on the initial design, performs a swap iteration, recalculates the A-Efficiency on the resulting design, and compares it with the previous one to decide whether to keep or discard the new design. This iterative process is repeated, by default, 1,000 times.

References

Edmondson., R. N. (2021). blocksdesign: Nested and crossed block designs for factorial and unstructured treatment sets. https://CRAN.R-project.org/package=blocksdesign

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a row-column design with 3 full blocks and 24 treatments
# and 6 rows. This for one location.
rowcold1 <- row_column(t = 24, nrows = 6, r = 3, l = 1, 
                       plotNumber= 101, 
                       locationNames = "Loc1",
                       iterations = 500,
                       seed = 21)
rowcold1$infoDesign
#> $rows
#> [1] 6
#> 
#> $columns
#> [1] 4
#> 
#> $reps
#> [1] 3
#> 
#> $treatments
#> [1] 24
#> 
#> $locations
#> [1] 1
#> 
#> $location_names
#> [1] "Loc1"
#> 
#> $seed
#> [1] 21
#> 
#> $id_design
#> [1] 9
#> 
rowcold1$resolvableBlocks
#> $Loc_Loc1
#> $Loc_Loc1$rep1
#>      [,1] [,2] [,3] [,4]
#> [1,]   NA   NA   NA   NA
#> [2,]   NA   NA   NA   NA
#> [3,]   NA   NA   NA   NA
#> [4,]   NA   NA   NA   NA
#> [5,]   NA   NA   NA   NA
#> [6,]   NA   NA   NA   NA
#> 
#> $Loc_Loc1$rep2
#>      [,1] [,2] [,3] [,4]
#> [1,]   NA   NA   NA   NA
#> [2,]   NA   NA   NA   NA
#> [3,]   NA   NA   NA   NA
#> [4,]   NA   NA   NA   NA
#> [5,]   NA   NA   NA   NA
#> [6,]   NA   NA   NA   NA
#> 
#> $Loc_Loc1$rep3
#>      [,1] [,2] [,3] [,4]
#> [1,]   NA   NA   NA   NA
#> [2,]   NA   NA   NA   NA
#> [3,]   NA   NA   NA   NA
#> [4,]   NA   NA   NA   NA
#> [5,]   NA   NA   NA   NA
#> [6,]   NA   NA   NA   NA
#> 
#> 
head(rowcold1$fieldBook,12)
#>    ID LOCATION PLOT REP ROW COLUMN ENTRY TREATMENT
#> 1   1     Loc1  101   1   1      1    24      G-24
#> 7   7     Loc1  102   1   1      2    22      G-22
#> 13 13     Loc1  103   1   1      3    16      G-16
#> 19 19     Loc1  104   1   1      4     4       G-4
#> 2   2     Loc1  105   1   2      1    15      G-15
#> 8   8     Loc1  106   1   2      2     1       G-1
#> 14 14     Loc1  107   1   2      3     8       G-8
#> 20 20     Loc1  108   1   2      4     3       G-3
#> 3   3     Loc1  109   1   3      1    21      G-21
#> 9   9     Loc1  110   1   3      2    11      G-11
#> 15 15     Loc1  111   1   3      3    10      G-10
#> 21 21     Loc1  112   1   3      4     7       G-7

# Example 2: Generates a row-column design with 3 full blocks and 30 treatments
# and 5 rows, for one location.
# In this case, we show how to use the option data.
treatments <- paste("ND-", 1:30, sep = "")
ENTRY <- 1:30
treatment_list <- data.frame(list(ENTRY = ENTRY, TREATMENT = treatments))
head(treatment_list)
#>   ENTRY TREATMENT
#> 1     1      ND-1
#> 2     2      ND-2
#> 3     3      ND-3
#> 4     4      ND-4
#> 5     5      ND-5
#> 6     6      ND-6
rowcold2 <- row_column(t = 30, nrows = 5, r = 3, l = 1, 
                       plotNumber= c(101,1001), 
                       locationNames = c("A", "B"),
                       seed = 15,
                       iterations = 500,
                       data = treatment_list)
#> Warning: Length of plot numbers is larger than number of locations.
rowcold2$infoDesign
#> $rows
#> [1] 5
#> 
#> $columns
#> [1] 6
#> 
#> $reps
#> [1] 3
#> 
#> $treatments
#> [1] 30
#> 
#> $locations
#> [1] 1
#> 
#> $location_names
#> [1] 1
#> 
#> $seed
#> [1] 15
#> 
#> $id_design
#> [1] 9
#> 
rowcold2$resolvableBlocks
#> $Loc_1
#> $Loc_1$rep1
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,] NA   NA   NA   NA   NA   NA  
#> [2,] NA   NA   NA   NA   NA   NA  
#> [3,] NA   NA   NA   NA   NA   NA  
#> [4,] NA   NA   NA   NA   NA   NA  
#> [5,] NA   NA   NA   NA   NA   NA  
#> 
#> $Loc_1$rep2
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,] NA   NA   NA   NA   NA   NA  
#> [2,] NA   NA   NA   NA   NA   NA  
#> [3,] NA   NA   NA   NA   NA   NA  
#> [4,] NA   NA   NA   NA   NA   NA  
#> [5,] NA   NA   NA   NA   NA   NA  
#> 
#> $Loc_1$rep3
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,] NA   NA   NA   NA   NA   NA  
#> [2,] NA   NA   NA   NA   NA   NA  
#> [3,] NA   NA   NA   NA   NA   NA  
#> [4,] NA   NA   NA   NA   NA   NA  
#> [5,] NA   NA   NA   NA   NA   NA  
#> 
#> 
head(rowcold2$fieldBook,12)
#>    ID LOCATION PLOT REP ROW COLUMN ENTRY TREATMENT
#> 1   1        1  101   1   1      1    26     ND-26
#> 6   6        1  102   1   1      2     8      ND-8
#> 11 11        1  103   1   1      3     9      ND-9
#> 16 16        1  104   1   1      4    18     ND-18
#> 21 21        1  105   1   1      5    19     ND-19
#> 26 26        1  106   1   1      6    21     ND-21
#> 2   2        1  107   1   2      1    11     ND-11
#> 7   7        1  108   1   2      2    15     ND-15
#> 12 12        1  109   1   2      3    22     ND-22
#> 17 17        1  110   1   2      4    24     ND-24
#> 22 22        1  111   1   2      5     7      ND-7
#> 27 27        1  112   1   2      6    28     ND-28