Generates a Resolvable Row-Column Design (RowColD)

It randomly generates a resolvable row-column designs (RowColD). Note that design optimization is only done at the level of rows and not columns; hence, design is suboptimal. The randomization can be done across locations.

row_column(
  t = NULL,
  nrows = NULL,
  r = NULL,
  l = 1,
  plotNumber = 101,
  locationNames = NULL,
  seed = NULL,
  data = NULL
)

Arguments

t
Number of treatments.

nrows
Number of rows of a full resolvable replicate.

r
Number of blocks (full resolvable replicates).

l
Number of locations. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

locationNames
(optional) Names for each location.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

data
(optional) Data frame with label list of treatments

Value

A list with four elements.

infoDesign is a list with information on the design parameters.

resolvableBlocks a list with the resolvable row columns blocks.

concurrence is the concurrence matrix.

fieldBook is a data frame with the row-column field book.

References

Edmondson., R. N. (2021). blocksdesign: Nested and crossed block designs for factorial and unstructured treatment sets. https://CRAN.R-project.org/package=blocksdesign

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a row-column design with 3 full blocks and 36 treatments
# and 6 rows. This for one location.
rowcold1 <- row_column(t = 36, nrows = 6, r = 3, l = 1, 
                       plotNumber= 101, 
                       locationNames = "Loc1",
                       seed = 21)
rowcold1$infoDesign

#> $nRows
#> [1] 6
#> 
#> $nCols
#> [1] 6
#> 
#> $Reps
#> [1] 3
#> 
#> $NumberTreatments
#> [1] 36
#> 
#> $NumberLocations
#> [1] 1
#> 
#> $Locations
#> [1] "Loc1"
#> 
#> $seed
#> [1] 21
#> 
#> $idDesign
#> [1] 9
#> 

rowcold1$resolvableBlocks

#> $Loc_Loc1
#> $Loc_Loc1$rep1
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]   11   16    9   36   14    1
#> [2,]   35    4    3   12    8   19
#> [3,]   17   28   15   24   20   25
#> [4,]    5   10   27   18   26    7
#> [5,]   29   34   21   30   32   31
#> [6,]   23   22   33    6    2   13
#> 
#> $Loc_Loc1$rep2
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]   17   20    1   27   10   34
#> [2,]   13   23    6   30   12   32
#> [3,]   16   24    2   25    7   33
#> [4,]   14   19    5   26    9   31
#> [5,]   15   22    4   28    8   36
#> [6,]   18   21    3   29   11   35
#> 
#> $Loc_Loc1$rep3
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]   33   13    4    6   32   10
#> [2,]    2    3    9   16   22   20
#> [3,]   23   29   14   21   12   36
#> [4,]   28    8   35   11   17    5
#> [5,]    7   34   19   26   27   25
#> [6,]   18   24   30   31    1   15
#> 
#> 

head(rowcold1$fieldBook,12)

#>    ID LOCATION PLOT REP ROW COLUMN ENTRY
#> 1   1     Loc1  101   1   1      1    11
#> 2   2     Loc1  102   1   1      2    16
#> 3   3     Loc1  103   1   1      3     9
#> 4   4     Loc1  104   1   1      4    36
#> 5   5     Loc1  105   1   1      5    14
#> 6   6     Loc1  106   1   1      6     1
#> 7   7     Loc1  107   1   2      1    35
#> 8   8     Loc1  108   1   2      2     4
#> 9   9     Loc1  109   1   2      3     3
#> 10 10     Loc1  110   1   2      4    12
#> 11 11     Loc1  111   1   2      5     8
#> 12 12     Loc1  112   1   2      6    19

# Example 2: Generates a row-column design with 3 full blocks and 30 treatments
# and 5 rows, for two locations.
# In this case, we show how to use the option data.
treatments <- paste("ND-", 1:30, sep = "")
ENTRY <- 1:30
treatment_list <- data.frame(list(ENTRY = ENTRY, TREATMENT = treatments))
head(treatment_list)

#>   ENTRY TREATMENT
#> 1     1      ND-1
#> 2     2      ND-2
#> 3     3      ND-3
#> 4     4      ND-4
#> 5     5      ND-5
#> 6     6      ND-6

rowcold2 <- row_column(t = 30, nrows = 5, r = 3, l = 2, 
                       plotNumber= c(101,1001), 
                       locationNames = c("A", "B"),
                       seed = 15,
                       data = treatment_list)
rowcold2$infoDesign

#> $nRows
#> [1] 5
#> 
#> $nCols
#> [1] 6
#> 
#> $Reps
#> [1] 3
#> 
#> $NumberTreatments
#> [1] 30
#> 
#> $NumberLocations
#> [1] 2
#> 
#> $Locations
#> [1] "A" "B"
#> 
#> $seed
#> [1] 15
#> 
#> $idDesign
#> [1] 9
#> 

rowcold2$resolvableBlocks

#> $Loc_A
#> $Loc_A$rep1
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-28" "ND-1"  "ND-21" "ND-2"  "ND-30" "ND-18"
#> [2,] "ND-16" "ND-11" "ND-24" "ND-29" "ND-19" "ND-22"
#> [3,] "ND-4"  "ND-25" "ND-26" "ND-17" "ND-14" "ND-8" 
#> [4,] "ND-5"  "ND-12" "ND-23" "ND-3"  "ND-13" "ND-10"
#> [5,] "ND-27" "ND-7"  "ND-15" "ND-9"  "ND-6"  "ND-20"
#> 
#> $Loc_A$rep2
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-12" "ND-22" "ND-16" "ND-19" "ND-25" "ND-1" 
#> [2,] "ND-14" "ND-4"  "ND-8"  "ND-9"  "ND-13" "ND-24"
#> [3,] "ND-20" "ND-11" "ND-30" "ND-28" "ND-21" "ND-10"
#> [4,] "ND-17" "ND-15" "ND-7"  "ND-18" "ND-5"  "ND-27"
#> [5,] "ND-26" "ND-6"  "ND-29" "ND-23" "ND-3"  "ND-2" 
#> 
#> $Loc_A$rep3
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-23" "ND-14" "ND-16" "ND-13" "ND-7"  "ND-21"
#> [2,] "ND-1"  "ND-28" "ND-6"  "ND-29" "ND-9"  "ND-8" 
#> [3,] "ND-30" "ND-25" "ND-26" "ND-18" "ND-15" "ND-27"
#> [4,] "ND-4"  "ND-22" "ND-10" "ND-24" "ND-5"  "ND-12"
#> [5,] "ND-17" "ND-2"  "ND-3"  "ND-11" "ND-20" "ND-19"
#> 
#> 
#> $Loc_B
#> $Loc_B$rep1
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-9"  "ND-20" "ND-7"  "ND-28" "ND-27" "ND-11"
#> [2,] "ND-29" "ND-16" "ND-4"  "ND-23" "ND-10" "ND-12"
#> [3,] "ND-22" "ND-24" "ND-26" "ND-13" "ND-14" "ND-19"
#> [4,] "ND-30" "ND-5"  "ND-1"  "ND-3"  "ND-21" "ND-8" 
#> [5,] "ND-25" "ND-18" "ND-2"  "ND-6"  "ND-15" "ND-17"
#> 
#> $Loc_B$rep2
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-9"  "ND-21" "ND-15" "ND-29" "ND-16" "ND-8" 
#> [2,] "ND-12" "ND-24" "ND-5"  "ND-7"  "ND-10" "ND-2" 
#> [3,] "ND-20" "ND-17" "ND-4"  "ND-11" "ND-30" "ND-6" 
#> [4,] "ND-3"  "ND-28" "ND-19" "ND-18" "ND-23" "ND-25"
#> [5,] "ND-27" "ND-26" "ND-22" "ND-13" "ND-1"  "ND-14"
#> 
#> $Loc_B$rep3
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-8"  "ND-7"  "ND-29" "ND-5"  "ND-4"  "ND-23"
#> [2,] "ND-15" "ND-10" "ND-16" "ND-3"  "ND-9"  "ND-25"
#> [3,] "ND-30" "ND-12" "ND-2"  "ND-11" "ND-18" "ND-20"
#> [4,] "ND-13" "ND-22" "ND-27" "ND-14" "ND-6"  "ND-26"
#> [5,] "ND-24" "ND-28" "ND-17" "ND-1"  "ND-21" "ND-19"
#> 
#> 

head(rowcold2$fieldBook,12)

#>    ID LOCATION PLOT REP ROW COLUMN ENTRY TREATMENT
#> 1   1        A  101   1   1      1    28     ND-28
#> 2   2        A  102   1   1      2     1      ND-1
#> 3   3        A  103   1   1      3    21     ND-21
#> 4   4        A  104   1   1      4     2      ND-2
#> 5   5        A  105   1   1      5    30     ND-30
#> 6   6        A  106   1   1      6    18     ND-18
#> 7   7        A  107   1   2      1    16     ND-16
#> 8   8        A  108   1   2      2    11     ND-11
#> 9   9        A  109   1   2      3    24     ND-24
#> 10 10        A  110   1   2      4    29     ND-29
#> 11 11        A  111   1   2      5    19     ND-19
#> 12 12        A  112   1   2      6    22     ND-22