Generates a Resolvable Row-Column Design (RowColD)

It randomly generates a resolvable row-column designs (RowColD). Note that design optimization is only done at the level of rows and not columns; hence, design is suboptimal. The randomization can be done across locations.

Usage

row_column(
  t = NULL,
  nrows = NULL,
  r = NULL,
  l = 1,
  plotNumber = 101,
  locationNames = NULL,
  seed = NULL,
  data = NULL
)

Arguments

t

Number of treatments.

nrows

Number of rows of a full resolvable replicate.

r

Number of blocks (full resolvable replicates).

l

Number of locations. By default l = 1.

plotNumber

Numeric vector with the starting plot number for each location. By default plotNumber = 101.

locationNames

(optional) Names for each location.

seed

(optional) Real number that specifies the starting seed to obtain reproducible designs.

data

(optional) Data frame with label list of treatments

Value

A list with four elements.

infoDesign is a list with information on the design parameters.

resolvableBlocks a list with the resolvable row columns blocks.

concurrence is the concurrence matrix.

fieldBook is a data frame with the row-column field book.

References

Edmondson., R. N. (2021). blocksdesign: Nested and crossed block designs for factorial and unstructured treatment sets. https://CRAN.R-project.org/package=blocksdesign

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Generates a row-column design with 3 full blocks and 36 treatments
# and 6 rows. This for one location.
rowcold1 <- row_column(t = 36, nrows = 6, r = 3, l = 1, 
                       plotNumber= 101, 
                       locationNames = "Loc1",
                       seed = 21)
rowcold1$infoDesign
#> $rows
#> [1] 6
#> 
#> $columns
#> [1] 6
#> 
#> $reps
#> [1] 3
#> 
#> $treatments
#> [1] 36
#> 
#> $locations
#> [1] 1
#> 
#> $location_names
#> [1] "Loc1"
#> 
#> $seed
#> [1] 21
#> 
#> $id_design
#> [1] 9
#> 
rowcold1$resolvableBlocks
#> $Loc_Loc1
#> $Loc_Loc1$rep1
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]   30    6   26   28   24   33
#> [2,]    4   15   36   12    9   25
#> [3,]   21   29   35   13    8   23
#> [4,]   11    3   34    5   18   16
#> [5,]   27    2   22   20   14   32
#> [6,]   10   17   31    1    7   19
#> 
#> $Loc_Loc1$rep2
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]   21    8   33   34    3    2
#> [2,]   19   10    1   20   12   14
#> [3,]    6   13    7   23   16   31
#> [4,]   24   25   11   18   26   32
#> [5,]   35   17   15   29    9   28
#> [6,]    5   22   36   27   30    4
#> 
#> $Loc_Loc1$rep3
#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]   31   19   15    1   14    3
#> [2,]    7   36   26    6   17    8
#> [3,]   10   27   24   22   12   28
#> [4,]   29    9    4   30   21   11
#> [5,]   16    2   25   18   34   23
#> [6,]    5   13   20   32   33   35
#> 
#> 
head(rowcold1$fieldBook,12)
#>    ID LOCATION PLOT REP ROW COLUMN ENTRY TREATMENT
#> 1   1     Loc1  101   1   1      1    30      G-30
#> 2   2     Loc1  102   1   1      2     6       G-6
#> 3   3     Loc1  103   1   1      3    26      G-26
#> 4   4     Loc1  104   1   1      4    28      G-28
#> 5   5     Loc1  105   1   1      5    24      G-24
#> 6   6     Loc1  106   1   1      6    33      G-33
#> 7   7     Loc1  107   1   2      1     4       G-4
#> 8   8     Loc1  108   1   2      2    15      G-15
#> 9   9     Loc1  109   1   2      3    36      G-36
#> 10 10     Loc1  110   1   2      4    12      G-12
#> 11 11     Loc1  111   1   2      5     9       G-9
#> 12 12     Loc1  112   1   2      6    25      G-25

# Example 2: Generates a row-column design with 3 full blocks and 30 treatments
# and 5 rows, for one location.
# In this case, we show how to use the option data.
treatments <- paste("ND-", 1:30, sep = "")
ENTRY <- 1:30
treatment_list <- data.frame(list(ENTRY = ENTRY, TREATMENT = treatments))
head(treatment_list)
#>   ENTRY TREATMENT
#> 1     1      ND-1
#> 2     2      ND-2
#> 3     3      ND-3
#> 4     4      ND-4
#> 5     5      ND-5
#> 6     6      ND-6
rowcold2 <- row_column(t = 30, nrows = 5, r = 3, l = 1, 
                       plotNumber= c(101,1001), 
                       locationNames = c("A", "B"),
                       seed = 15,
                       data = treatment_list)
#> Warning: Since plotNumber was missing, it was set up to default values.
#> Warning: Length of plot numbers is larger than number of locations.
rowcold2$infoDesign
#> $rows
#> [1] 5
#> 
#> $columns
#> [1] 6
#> 
#> $reps
#> [1] 3
#> 
#> $treatments
#> [1] 30
#> 
#> $locations
#> [1] 1
#> 
#> $location_names
#> [1] 1
#> 
#> $seed
#> [1] 15
#> 
#> $id_design
#> [1] 9
#> 
rowcold2$resolvableBlocks
#> $Loc_1
#> $Loc_1$rep1
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-28" "ND-2"  "ND-21" "ND-10" "ND-15" "ND-22"
#> [2,] "ND-13" "ND-17" "ND-14" "ND-1"  "ND-30" "ND-25"
#> [3,] "ND-19" "ND-9"  "ND-7"  "ND-23" "ND-20" "ND-8" 
#> [4,] "ND-18" "ND-6"  "ND-3"  "ND-26" "ND-29" "ND-27"
#> [5,] "ND-16" "ND-24" "ND-4"  "ND-5"  "ND-12" "ND-11"
#> 
#> $Loc_1$rep2
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-18" "ND-22" "ND-14" "ND-15" "ND-29" "ND-10"
#> [2,] "ND-24" "ND-17" "ND-5"  "ND-28" "ND-7"  "ND-19"
#> [3,] "ND-30" "ND-13" "ND-12" "ND-1"  "ND-6"  "ND-20"
#> [4,] "ND-25" "ND-4"  "ND-2"  "ND-8"  "ND-26" "ND-9" 
#> [5,] "ND-3"  "ND-23" "ND-27" "ND-21" "ND-16" "ND-11"
#> 
#> $Loc_1$rep3
#>      [,1]    [,2]    [,3]    [,4]    [,5]    [,6]   
#> [1,] "ND-26" "ND-23" "ND-4"  "ND-15" "ND-20" "ND-22"
#> [2,] "ND-14" "ND-8"  "ND-18" "ND-5"  "ND-25" "ND-12"
#> [3,] "ND-28" "ND-2"  "ND-10" "ND-9"  "ND-17" "ND-21"
#> [4,] "ND-24" "ND-16" "ND-27" "ND-3"  "ND-7"  "ND-19"
#> [5,] "ND-11" "ND-30" "ND-29" "ND-13" "ND-1"  "ND-6" 
#> 
#> 
head(rowcold2$fieldBook,12)
#>    ID LOCATION PLOT REP ROW COLUMN ENTRY TREATMENT
#> 1   1        1  101   1   1      1    28     ND-28
#> 2   2        1  102   1   1      2     2      ND-2
#> 3   3        1  103   1   1      3    21     ND-21
#> 4   4        1  104   1   1      4    10     ND-10
#> 5   5        1  105   1   1      5    15     ND-15
#> 6   6        1  106   1   1      6    22     ND-22
#> 7   7        1  107   1   2      1    13     ND-13
#> 8   8        1  108   1   2      2    17     ND-17
#> 9   9        1  109   1   2      3    14     ND-14
#> 10 10        1  110   1   2      4     1      ND-1
#> 11 11        1  111   1   2      5    30     ND-30
#> 12 12        1  112   1   2      6    25     ND-25