Unreplicated designs using the sparse allocation approach

Unreplicated designs using the sparse allocation approach

Usage

sparse_allocation(
  lines,
  nrows,
  ncols,
  l,
  planter = "serpentine",
  plotNumber,
  copies_per_entry,
  checks = NULL,
  exptName = NULL,
  locationNames,
  sparse_list,
  seed,
  data = NULL
)

Arguments

lines

Number of genotypes, experimental lines or treatments.

nrows

Number of rows in the field.

ncols

Number of columns in the field.

l

Number of locations or sites. By default l = 1.

planter

Option for serpentine or cartesian plot arrangement. By default planter = 'serpentine'.

plotNumber

Numeric vector with the starting plot number for each location. By default plotNumber = 101.

copies_per_entry

Number of copies per plant. When design is sparse then copies_per_entry < l

checks

Number of genotypes checks.

exptName

(optional) Name of the experiment.

locationNames

(optional) Names each location.

sparse_list

(optional) A class "Sparse" object generated by do_optim() function.

seed

(optional) Real number that specifies the starting seed to obtain reproducible designs.

data

(optional) Data frame with 2 columns: ENTRY | NAME . ENTRY must be numeric.

Value

A list with four elements.

designs is a list with each location unreplicated randomization.

list_locs is a list with each location list of entries.

allocation is a matrix with the allocation of treatments.

size_locations is a data frame with one column for each location and one row with the size of the location.

References

Edmondson, R.N. Multi-level Block Designs for Comparative Experiments. JABES 25, 500–522 (2020). https://doi.org/10.1007/s13253-020-00416-0

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb]

Examples

sparse <- sparse_allocation(
  lines = 120, 
  l = 4, 
  copies_per_entry = 3, 
  checks = 4, 
  locationNames = c("LOC1", "LOC2", "LOC3", "LOC4", "LOC5"), 
  seed = 1234
)