Split a population of genotypes randomly into several locations.

Split a population of genotypes randomly into several locations, with the aim of having approximatelly the same number of replicates of each genotype, line or treatment per location.

split_families(l = NULL, data = NULL)

Arguments

l
Number of locations.

data
Data frame with the entry (ENTRY) and the labels of each treatment (NAME) and number of individuals per family group (FAMILY).

Value

A list with two elements.

rowsEachlist is a table with a summary of cases.

data_locations is a data frame with the entries for each location

Author

Didier Murillo [aut], Salvador Gezan [aut], Ana Heilman [ctb], Thomas Walk [ctb], Johan Aparicio [ctb], Richard Horsley [ctb]

Examples

# Example 1: Split a population of 3000 and 200 families into 8 locations. 
# Original dataset is been simulated.
set.seed(77)
N <- 3000; families <- 200
ENTRY <- 1:N
NAME <- paste0("SB-", 1:N)
FAMILY <- vector(mode = "numeric", length = N)
x <- 1:N
for (i in x) { FAMILY[i] <- sample(1:families, size = 1, replace = TRUE) }
gen.list <- data.frame(list(ENTRY = ENTRY, NAME = NAME, FAMILY = FAMILY))
head(gen.list)

#>   ENTRY NAME FAMILY
#> 1     1 SB-1    146
#> 2     2 SB-2    173
#> 3     3 SB-3    197
#> 4     4 SB-4     57
#> 5     5 SB-5     37
#> 6     6 SB-6     29

# Now we are going to use the split_families() function.
split_population <- split_families(l = 8, data = gen.list)

#> Warning: Family 23 is not in all locations.

#> Warning: Family 41 is not in all locations.

print(split_population)

#> Split families: 
#> 
#> 
#>  Data frame with the summary of cases by location: 
#>     Location   n
#> 1 Location 1 380
#> 2 Location 2 367
#> 3 Location 3 384
#> 4 Location 4 394
#> 5 Location 5 365
#> 6 Location 6 371
#> 7 Location 7 367
#> 8 Location 8 372
#> 
#>  10 First observations of the data frame with the entries for each location: 
#>    ENTRY    NAME FAMILY   LOCATION
#> 1   1040 SB-1040      1 Location 1
#> 2   1671 SB-1671      2 Location 1
#> 3   1151 SB-1151      2 Location 1
#> 4   2231 SB-2231      3 Location 1
#> 5    972  SB-972      3 Location 1
#> 6   2499 SB-2499      4 Location 1
#> 7   2184 SB-2184      5 Location 1
#> 8   2144 SB-2144      5 Location 1
#> 9   2087 SB-2087      6 Location 1
#> 10  1020 SB-1020      6 Location 1

summary(split_population)

#> Split families: 
#> 
#> 1. Structure of the data frame with the summary of entries by location: 
#> 
#> 'data.frame':	8 obs. of  2 variables:
#>  $ Location: chr  "Location 1" "Location 2" "Location 3" "Location 4" ...
#>  $ n       : num  380 367 384 394 365 371 367 372
#> 2. Structure of the data frame with the entries for each location: 
#> 
#> 'data.frame':	3000 obs. of  4 variables:
#>  $ ENTRY   : int  1040 1671 1151 2231 972 2499 2184 2144 2087 1020 ...
#>  $ NAME    : chr  "SB-1040" "SB-1671" "SB-1151" "SB-2231" ...
#>  $ FAMILY  : num  1 2 2 3 3 4 5 5 6 6 ...
#>  $ LOCATION: chr  "Location 1" "Location 1" "Location 1" "Location 1" ...

head(split_population$data_locations,12)

#>    ENTRY    NAME FAMILY   LOCATION
#> 1   1040 SB-1040      1 Location 1
#> 2   1671 SB-1671      2 Location 1
#> 3   1151 SB-1151      2 Location 1
#> 4   2231 SB-2231      3 Location 1
#> 5    972  SB-972      3 Location 1
#> 6   2499 SB-2499      4 Location 1
#> 7   2184 SB-2184      5 Location 1
#> 8   2144 SB-2144      5 Location 1
#> 9   2087 SB-2087      6 Location 1
#> 10  1020 SB-1020      6 Location 1
#> 11  1706 SB-1706      7 Location 1
#> 12   939  SB-939      8 Location 1