Generates a Square Lattice Design.

It randomly generates a square lattice design across locations.

square_lattice(
  t = NULL,
  k = NULL,
  r = NULL,
  l = 1,
  plotNumber = 101,
  locationNames = NULL,
  seed = NULL,
  data = NULL
)

Arguments

t
Number of treatments.

k
Size of incomplete blocks (number of unites per incomplete block).

r
Number of blocks (full resolvable replicates).

l
Number of locations. By default l = 1.

plotNumber
Numeric vector with the starting plot number for each location. By default plotNumber = 101.

locationNames
(optional) Names for each location.

seed
(optional) Real number that specifies the starting seed to obtain reproducible designs.

data
(optional) Data frame with label list of treatments.

Value

A list with information on the design parameters.

Data frame with the square lattice design field book.

References

Edmondson, R.N. 2020. Package blocksdesign: Nested and Crossed Block Designs for Factorial and Unstructured Treatment Sets.

Examples

#Example 1: Generates a square lattice design with 5 full blocks, 8 units per iBlock,
# 8 iBlocks for a square number of treatmens of 64 in two locations.
squareLattice1 <- square_lattice(t = 64, k = 8, r = 5, l = 2, 
                                 plotNumber = c(1001, 2001),
                                 locationNames = c("FARGO", "MINOT"), 
                                 seed = 1986)
squareLattice1$infoDesign

#> $Reps
#> [1] 5
#> 
#> $iBlocks
#> [1] 8
#> 
#> $NumberTreatments
#> [1] 64
#> 
#> $NumberLocations
#> [1] 2
#> 
#> $Locations
#> [1] "FARGO" "MINOT"
#> 
#> $seed
#> [1] 1986
#> 
#> $lambda
#> [1] 0.5555556
#> 

head(squareLattice1$fieldBook,12)

#>    ID LOCATION PLOT REP IBLOCK UNIT ENTRY
#> 1   1    FARGO 1001   1      1    1     3
#> 2   2    FARGO 1002   1      1    2     4
#> 3   3    FARGO 1003   1      1    3     5
#> 4   4    FARGO 1004   1      1    4     6
#> 5   5    FARGO 1005   1      1    5     2
#> 6   6    FARGO 1006   1      1    6     8
#> 7   7    FARGO 1007   1      1    7     7
#> 8   8    FARGO 1008   1      1    8     1
#> 9   9    FARGO 1009   1      2    1    51
#> 10 10    FARGO 1010   1      2    2    55
#> 11 11    FARGO 1011   1      2    3    54
#> 12 12    FARGO 1012   1      2    4    53

#Example 2: Generates a square lattice design with 3 full blocks, 12 units per iBlock,
# 12 iBlocks for a square number of treatmens of 144 in one location.
squareLattice2 <- square_lattice(t = 100, k = 10, r = 3, l = 1, 
                                 plotNumber = 1001,
                                 locationNames = "CASSELTON", 
                                 seed = 1986)
squareLattice2$infoDesign

#> $Reps
#> [1] 3
#> 
#> $iBlocks
#> [1] 10
#> 
#> $NumberTreatments
#> [1] 100
#> 
#> $NumberLocations
#> [1] 1
#> 
#> $Locations
#> [1] "CASSELTON"
#> 
#> $seed
#> [1] 1986
#> 
#> $lambda
#> [1] 0.2727273
#> 

head(squareLattice2$fieldBook,12)

#>    ID  LOCATION PLOT REP IBLOCK UNIT ENTRY
#> 1   1 CASSELTON 1001   1      1    1    83
#> 2   2 CASSELTON 1002   1      1    2    53
#> 3   3 CASSELTON 1003   1      1    3    73
#> 4   4 CASSELTON 1004   1      1    4    33
#> 5   5 CASSELTON 1005   1      1    5    13
#> 6   6 CASSELTON 1006   1      1    6    63
#> 7   7 CASSELTON 1007   1      1    7     3
#> 8   8 CASSELTON 1008   1      1    8    23
#> 9   9 CASSELTON 1009   1      1    9    43
#> 10 10 CASSELTON 1010   1      1   10    93
#> 11 11 CASSELTON 1011   1      2    1    65
#> 12 12 CASSELTON 1012   1      2    2    95